### 2026-05-04 #10527 @editorialbot
[PRE REVIEW]: Joule Profiler: A phase-based energy measurement tool
**Submitting author:** @jwoirhaye (<a href="http://orcid.org/0009-0005-0777-9090">Jérémy Woirhaye</a>)
**Repository:** https://github.com/joule-profiler/joule-profiler/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v2.1.1
