### 2026-05-04 #10527 @danielskatz
👋 @jwoirhaye - note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html) and note that you can click on the error above to find out more about it. Please feel free to make changes to your .md file, then use the command `@editorialbot generate pdf` to try to make a new PDF. An editorialbot command needs to be the first entry in a new comment.
