### 2026-05-04 #10527 @danielskatz
Also, your paper is too long - JOSS papers are recommended to be 750 - 1750 words. You can use the command '@editorialbot check repository` to check this after you make changes.
