### 2026-05-04 4e99c4f
docs: Update alumet citation with doi
