+++ joss/paper/paper.bib
@@ -1,15 +1,14 @@
-@unpublished{alumet,
-  TITLE = {{Dissecting the software-based measurement of CPU energy consumption: a comparative analysis}},
-  AUTHOR = {Raffin, Guillaume and Trystram, Denis},
-  URL = {https://hal.science/hal-04420527},
-  NOTE = {working paper or preprint},
-  YEAR = {2024},
-  MONTH = Jul,
-  KEYWORDS = {energy consumption ; energy efficiency ; performance analysis ; software measurement ; RAPL library (Running Average Power Limit)},
-  PDF = {https://hal.science/hal-04420527v2/file/all_together.pdf},
-  HAL_ID = {hal-04420527},
-  HAL_VERSION = {v2},
-}
+@ARTICLE{10746340,
+  author={Raffin, Guillaume and Trystram, Denis},
+  journal={IEEE Transactions on Parallel and Distributed Systems}, 
+  title={Dissecting the Software-Based Measurement of CPU Energy Consumption: A Comparative Analysis}, 
+  year={2025},
+  volume={36},
+  number={1},
+  pages={96-107},
+  keywords={Energy consumption;Energy efficiency;Performance analysis;Software measurement;Global warming;Weather forecasting;Central Processing Unit;Closed box;Climate change;Environmental monitoring;Energy consumption;energy efficiency;performance analysis;software measurement;RAPL library (Running Average Power Limit)},
+  doi={10.1109/TPDS.2024.3492336}}
+
 
 @misc{scaphandre,
   author       = {{Hubblo contributors}},
