+++ joss/paper/paper.md
@@ -31,100 +31,90 @@ repository: https://github.com/joule-profiler/joule-profiler
 
 # Summary
 
-Joule Profiler is a lightweight command-line tool for Linux measuring a running program's energy consumption. Designed to be as non-intrusive as possible in terms of instrumentation, it allows developers and researchers to better understand their software’s energy consumption without significant overhead. Joule Profiler enables users to decompose program execution into distinct phases, such as data loading, computation, and result generation, and to attribute energy consumption to each phase. It monitors the target program to detect user-defined phase triggers and automatically queries various sources, such as Intel RAPL for CPUs or NVML for GPUs, to report system-wide energy consumption.
+Joule Profiler is a lightweight Linux command-line tool for measuring a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and to attribute energy use to each. The tool detects phase triggers from program output and automatically queries sources like Intel RAPL (CPUs) or NVML (GPUs) to report energy consumption.
 
-# Statement of need  
+# Statement of need
 
-The energy consumption of computing infrastructures has become a major concern for both research and industry. Programs deployed in clouds, data centres, and edge environments contribute significantly to global energy use. Improving their efficiency has become a key goal, and understanding how software uses energy requires tools that can measure consumption during execution. Modern CPUs and GPUs provide hardware counters via interfaces such as Intel RAPL, enabling software-based energy measurement without external hardware.  
+Energy use in computing is a growing concern across research and industry. Software running in clouds, data centres, and edge devices contributes significantly to global energy consumption. Improving efficiency requires tools that measure energy during execution. Hardware counters provided by modern CPUs and GPUs (e.g., Intel RAPL) make software-based energy measurement possible without external devices.
 
-Researchers and developers aiming to optimise energy consumption may need a simple tool to measure the energy use of code segments without deploying a full monitoring infrastructure. Joule Profiler covers this gap by implementing a phase-based energy profiling tool that integrates easily into experimental workflows.  
+Researchers and developers need simple tools to measure the energy use of code segments without complex infrastructure. Joule Profiler addresses this with phase-based profiling that integrates easily into workflows.
 
-# State of the field  
+# State of the field
 
-Numerous tools use these counters to monitor energy. PowerAPI [@powerapi] offers infrastructure for monitoring and attributing energy consumption in distributed systems. Alumet [@alumet] provides a framework for measuring energy and studying the reliability of software-based measurement. Scaphandre [@scaphandre] exports energy metrics for hosts and containers for integration with observability platforms like Prometheus. EnergiBridge [@sallou_energibridge_2024] provides cross-platform energy measurement via a command-line wrapper. JouleIt [@jouleit], the direct predecessor that inspired this work, demonstrated the viability of a lightweight wrapper approach but lacks phase decomposition, GPU support, and a modular source architecture.  
+Existing tools like PowerAPI [@powerapi], Alumet [@alumet], Scaphandre [@alumet], and EnergiBridge [@sallou_energibridge_2024] monitor energy using these counters, often focusing on distributed and system-level observability. JouleIt [@jouleit], which inspired this work, demonstrated a light wrapper approach but lacked phase decomposition, GPU support, and modularity.
 
[…]
