### 2026-05-04 a9f92cc
feat: Update figure with last results
