+++ joss/paper/paper.bib
@@ -1,4 +1,4 @@
-@ARTICLE{10746340,
+@ARTICLE{alumet,
   author={Raffin, Guillaume and Trystram, Denis},
   journal={IEEE Transactions on Parallel and Distributed Systems}, 
   title={Dissecting the Software-Based Measurement of CPU Energy Consumption: A Comparative Analysis}, 
