+++ joss/paper/paper.md
@@ -31,7 +31,7 @@ repository: https://github.com/joule-profiler/joule-profiler
 
 # Summary
 
-Joule Profiler is a lightweight Linux command-line tool for measuring a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and to attribute energy use to each. The tool detects phase triggers from program output and automatically queries sources like Intel RAPL (CPUs) or NVML (GPUs) to report energy consumption.
+Joule Profiler is a lightweight Linux command-line tool for measuring a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and to attribute energy use to each. The tool detects phase triggers from program output and automatically queries sources like Intel RAPL (CPUs) or NVML (GPUs) to report system-wide energy consumption.
 
 # Statement of need
 
@@ -41,7 +41,7 @@ Researchers and developers need simple tools to measure the energy use of code s
 
 # State of the field
 
-Existing tools like PowerAPI [@powerapi], Alumet [@alumet], Scaphandre [@alumet], and EnergiBridge [@sallou_energibridge_2024] monitor energy using these counters, often focusing on distributed and system-level observability. JouleIt [@jouleit], which inspired this work, demonstrated a light wrapper approach but lacked phase decomposition, GPU support, and modularity.
+Existing tools like PowerAPI [@powerapi], Alumet [@alumet], Scaphandre [@scaphandre], and EnergiBridge [@sallou_energibridge_2024] monitor energy using these counters, often focusing on distributed and system-level observability. JouleIt [@jouleit], which inspired this work, demonstrated a light wrapper approach but lacked phase decomposition, GPU support, and modularity.
 
 These solutions are suited for system-level monitoring, not fine-grained analysis of program phases. Joule Profiler, in contrast, is designed for lightweight, single-invocation use, enabling energy attribution to specific phases within programs.
 
@@ -68,53 +68,50 @@ The tool uses a layered structure: the core detects phases and aggregates metric
 
 # Validity of the energy measurement
 
-To validate its measurements, Joule Profiler was compared with reference tools perf [@perfwiki] and Alumet, both using RAPL counters but different strategies. This checks whether Joule Profiler introduces measurement bias.
+To validate its measurements, Joule Profiler was compared with reference tools perf [@perfwiki] and Alumet [@alumet], both using RAPL counters but different strategies. This checks whether Joule Profiler introduces measurement bias.
 
 Three scenarios were tested: (1) parallel runs of Joule Profiler and perf (with CPU load) or Alumet (with GPU load) alongside a sleep command, ensuring identical hardware activity and measurement noise ; (2) Sequential execution of Joule Profiler, perf, and Alumet with workload pinned to a single CPU core, to compare overhead and variability; and (3) A custom workload with periodic output tokens tested phase detection precision.
 
-Experiments used Grid’5000 nodes: Chirop (Intel Xeon, RAPL, 512 GiB RAM) and Chifflot (Nvidia Tesla V100, NVML, 192 GiB RAM). Energy was measured from RAPL (PACKAGE, DRAM) and NVML (GPU). perf_event was used for access. Hyper-threading was disabled, and CPU frequency was fixed to reduce variability.
[…]
