### 2026-05-04 771b702
docs: Update figure parallel execution and typo
