### 2026-05-04 699b731
Improve text clarity in paper.md
Clarified text for accuracy and consistency.
