### 2026-05-04 d4ac520
Improve text clarity in paper.md
Clarified text for accuracy and consistency.
