### 2026-05-12 85ac41c
feat: procfs source implementation, memory done and io operations unstable
