### 2026-05-12 2b92b5a
feat: procfs source implementation, memory done and io operations unstable
