### 2026-05-12 86a2592
feat: procfs source implementation, memory done and io operations unstable
