+++ cli/src/bin/main.rs
@@ -7,6 +7,7 @@ use joule_profiler_core::JouleProfiler;
 use joule_profiler_core::config::{Command, Config};
 use joule_profiler_source_nvml::Nvml;
 use joule_profiler_source_perf_event::PerfEvent;
+use joule_profiler_source_procfs::Procfs;
 use joule_profiler_source_rapl::{perf, powercap};
 use log::{trace, warn};
 
@@ -64,6 +65,12 @@ async fn main() -> Result<()> {
         trace!("Initializing perf_event source");
         let perf_event = PerfEvent::new()?;
         profiler.add_source(perf_event);
+    }
+
+    if cli.sources.contains(&Source::Procfs) {
+        trace!("Initializing procfs source");
+        let procfs = Procfs::default();
+        profiler.add_source(procfs);
     }
 
     let config = Config::try_from(cli)?;
