+++ sources/procfs/src/lib.rs
@@ -1,0 +1,302 @@
+use futures::StreamExt;
+use joule_profiler_core::{
+    sensor::{Sensor, Sensors},
+    source::MetricReader,
+    types::{Metric, MetricValue, Metrics},
+    unit::{MetricUnit, Unit, UnitPrefix},
+};
+use procfs::process::Process;
+use std::sync::Arc;
+use std::time::Duration;
+use thiserror::Error;
+use tokio::sync::Mutex;
+use tokio::task::JoinHandle;
+use tokio_timerfd::Interval;
+
+use crate::counters::{Counters, IoCounters, MemoryCounters};
+
+pub mod counters;
+
+const IO_COUNTERS_METRIC_UNIT: MetricUnit = MetricUnit {prefix: UnitPrefix::None, unit: Unit::Byte};
+
+#[derive(Error, Debug)]
+pub enum ProcfsError {
+    #[error("Procfs not initialized, call init first")]
+    NotInitialized,
+    #[error(transparent)]
+    Procfs(#[from] procfs::ProcError),
+    #[error(transparent)]
+    Io(#[from] std::io::Error),
+}
+
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
+pub enum MemoryUnit {
+    Bytes,
+    Kilo,
+    #[default]
+    Mega,
+    Giga,
+}
+
+impl From<MemoryUnit> for MetricUnit {
+    fn from(unit: MemoryUnit) -> Self {
+        match unit {
+            MemoryUnit::Bytes => MetricUnit {
+                prefix: UnitPrefix::None,
+                unit: Unit::Byte,
+            },
+            MemoryUnit::Kilo => MetricUnit {
+                prefix: UnitPrefix::Kilo,
+                unit: Unit::Byte,
+            },
+            MemoryUnit::Mega => MetricUnit {
+                prefix: UnitPrefix::Mega,
+                unit: Unit::Byte,
+            },
+            MemoryUnit::Giga => MetricUnit {
+                prefix: UnitPrefix::Giga,
+                unit: Unit::Byte,
+            },
+        }
+    }
+}
+
+type Result<T> = std::result::Result<T, ProcfsError>;
+
+#[derive(Debug, Default)]
+pub struct Procfs {
+    process: Option<Process>,
+    poll_interval: Option<Duration>,
+    memory_counters: Arc<Mutex<MemoryCounters>>,
+    io_counters: IoCounters,
+    handle: Option<JoinHandle<Result<()>>>,
+    memory_unit: MemoryUnit,
+}
+
+fn read_memory(process: &Process) -> Result<(u64, u64, u64, u64, u64)> {
+    let vm_size = process.stat()?.vsize;
+    let smaps = process.smaps_rollup()?;
+
+    let (rss, pss, anon, shared) = smaps
+        .memory_map_rollup
+        .iter()
+        .next()
+        .map(|entry| {
+            let m = &entry.extension.map;
+            let rss = m.get("Rss").copied().unwrap_or(0);
+            let pss = m.get("Pss").copied().unwrap_or(0);
+            let anon = m.get("Anonymous").copied().unwrap_or(0);
+            let shared = m.get("Shared_Clean").copied().unwrap_or(0)
+                + m.get("Shared_Dirty").copied().unwrap_or(0);
+            (rss, pss, anon, shared)
+        })
+        .unwrap_or_default();
+
+    Ok((vm_size, rss, pss, shared, anon))
+}
+
+fn read_io(process: &Process) -> Result<(u64, u64)> {
+    let io = process.io()?;
+    Ok((io.rchar, io.wchar))
+}
+
+impl Procfs {
+    pub fn new(poll_interval: Option<Duration>) -> Self {
+        Self {
[…]
