### 2026-05-13 8a312ab
fix: force stdin to /dev/tty (current terminal) to fix entry when piped to bash
