### 2026-05-13 d89b5db
docs: update quickstart prints after installation
