+++ install.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 #
 # joule-profiler installer
-# https://github.com/jwoirhaye/joule-profiler
+# https://github.com/joule-profiler/joule-profiler
 #
 set -e
 
@@ -567,22 +567,19 @@ print_usage() {
     echo "" >&2
     echo -e "${COLOR_GREEN}Installation complete!${COLOR_RESET}" >&2
     echo "" >&2
-    echo "To get started with joule-profiler:" >&2
+    echo "To get started with Joule Profiler:" >&2
     echo "" >&2
-    echo "  # List available RAPL domains" >&2
-    echo -e "  ${COLOR_BLUE}sudo joule-profiler list-domains${COLOR_RESET}" >&2
+    echo "  # List available sensors" >&2
+    echo -e "  ${COLOR_BLUE}joule-profiler list-sensors${COLOR_RESET}" >&2
     echo "" >&2
-    echo "  # Measure energy consumption (simple mode)" >&2
-    echo -e "  ${COLOR_BLUE}sudo joule-profiler simple -- <your-command>${COLOR_RESET}" >&2
-    echo "" >&2
-    echo "  # Measure with phase detection" >&2
-    echo -e "  ${COLOR_BLUE}sudo joule-profiler phases -- <your-command>${COLOR_RESET}" >&2
-    echo "" >&2
-    echo "  # Export to JSON" >&2
-    echo -e "  ${COLOR_BLUE}sudo joule-profiler simple --json -- <your-command>${COLOR_RESET}" >&2
+    echo "  # Phase-based profiling" >&2
+    echo -e "  ${COLOR_BLUE}sudo joule-profiler profile -- <your-command>${COLOR_RESET}" >&2
     echo "" >&2
     echo "  # Show help" >&2
     echo -e "  ${COLOR_BLUE}joule-profiler --help${COLOR_RESET}" >&2
+    echo "" >&2
+    echo "If you want to launch Joule Profiler without root privileges (root), configure the perf_event_paranoid level:" >&2
+    echo -e "  ${COLOR_BLUE}sudo sysctl kernel.perf_event_paranoid=0${COLOR_RESET}" >&2
     echo "" >&2
     echo "For more information, visit:" >&2
     echo "  https://github.com/$REPO" >&2
@@ -594,10 +591,10 @@ main() {
     parse_args "$@"
 
     echo "" >&2
-    echo "╔══════════════════════════════════════════╗" >&2
-    echo "║   joule-profiler installer v0.1.0        ║" >&2
-    echo "║   github.com/jwoirhaye/joule-profiler    ║" >&2
-    echo "╚══════════════════════════════════════════╝" >&2
+    echo "╔════════════════════════════════════════════╗" >&2
+    echo "║          Joule Profiler installer          ║" >&2
+    echo "║  github.com/joule-profiler/joule-profiler  ║" >&2
+    echo "╚════════════════════════════════════════════╝" >&2
     echo "" >&2
 
     log_debug "Installation directory: $INSTALL_DIR"
