### 2026-05-13 #10575 @editorialbot
[REVIEW]: Joule Profiler: A phase-based energy measurement tool
**Submitting author:** @jwoirhaye (<a href="http://orcid.org/0009-0005-0777-9090">Jérémy Woirhaye</a>)
**Repository:** https://github.com/joule-profiler/joule-profiler/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** cli-v3.0.1
