### 2026-05-18 1eb3c6f
feat(procfs): removed delta and first measure metrics and added per-phase minimum
