### 2026-05-18 916c5a3
feat(procfs): removed delta and first measure metrics and added per-phase minimum
