### 2026-05-18 fea6588
feat(procfs): map child processes recursively and measure all
