### 2026-05-18 e449cab
feat(procfs): map child processes recursively and measure all
