### 2026-05-18 e057da5
chore: added logging for procfs source
