### 2026-05-18 bb7c91f
docs(installer): improve post-install usage message
