+++ cli/Cargo.toml
@@ -18,6 +18,7 @@ joule-profiler-source-rapl = { path = "../sources/rapl", version = "1.0.2" }
 joule-profiler-source-nvml = { path = "../sources/nvml", version = "1.0.2" }
 joule-profiler-source-perf_event = { path = "../sources/perf_event", version = "1.0.2" }
 joule-profiler-source-procfs = { path = "../sources/procfs", version = "0.1.0" }
+joule-profiler-source-cgroup = { path = "../sources/cgroup", version = "0.1.0" }
 
 joule-profiler-core.workspace = true
 log.workspace = true
