+++ sources/cgroup/README.md
@@ -1,0 +1,6 @@
+# joule-profiler-sources-cgroup
+
+CGroup v2 metric source for [joule-profiler](https://github.com/joule-profiler/joule-profiler).
+
+Places the profiled process (and **all its children**) into a dedicated cgroup and
+samples memory, CPU and I/O.
