### 2026-05-18 3e902e1
feat(sources): add cgroup v2 source base implementation
Add initial `joule-profiler-source-cgroup` crate implementing `MetricReader` via Linux cgroup v2.

- Create sub-cgroup and attach child PID in `init(pid)`
- Sample memory.current, memory.peak, cpu.stat and io.stat in `measure()`
- Add `--cgroup` flag to the CLI

WIP: memory.stat breakdown (anon/file) not yet reliable
