+++ cli/src/lib.rs
@@ -65,6 +65,11 @@ pub struct CliArgs {
     #[arg(long)]
     pub perf: bool,
 
+    /// CGroup v2 metrics
+    #[arg(long)]
+    pub cgroup: bool,
+
+
     /// Choose RAPL backend between powercap or perf
     #[arg(long = "rapl-backend", value_enum, default_value_t = RaplBackend::Perf)]
     pub rapl_backend: RaplBackend,
