+++ sources/cgroup/src/lib.rs
@@ -1,0 +1,277 @@
+use std::fs;
+use std::path::PathBuf;
+use log::{debug, trace, warn};
+use joule_profiler_core::sensor::{Sensor, Sensors};
+use joule_profiler_core::source::MetricReader;
+use joule_profiler_core::types::{Metric, Metrics};
+use joule_profiler_core::unit::{MetricUnit, Unit, UnitPrefix};
+use crate::error::CgroupError;
+use crate::snapshot::Snapshot;
+use crate::util::{read_flat_keyed, read_io_stat, read_u64_opt};
+
+mod error;
+mod snapshot;
+mod util;
+
+const SOURCE_NAME: &str = "CGroup";
+
+const BYTE_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Byte,
+};
+
+const MICRO_SECOND_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::Micro,
+    unit: Unit::Second,
+};
+
+const COUNT_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Count,
+};
+
+
+#[derive(Debug, Clone)]
+pub struct CgroupConfig {
+    pub cgroup_root: PathBuf,
+    pub cgroup_name: String,
+}
+
+impl Default for CgroupConfig {
+    fn default() -> Self {
+        Self {
+            cgroup_root: PathBuf::from("/sys/fs/cgroup"),
+            cgroup_name: format!("joule-profiler-{}", std::process::id()),
+        }
+    }
+}
+
+pub struct CgroupSource {
+    config: CgroupConfig,
+    cgroup_path: PathBuf,
+    last_snapshot: Snapshot,
+}
+
+impl CgroupSource {
+    pub fn new(config: CgroupConfig) -> Result<Self, CgroupError> {
+        let controllers = config.cgroup_root.join("cgroup.controllers");
+        if !controllers.exists() {
+            return Err(CgroupError::NotAvailable(
+                config.cgroup_root.display().to_string(),
+            ));
+        }
+
+        let cgroup_path = config.cgroup_root.join(&config.cgroup_name);
+
+        Ok(Self {config, cgroup_path,last_snapshot: Snapshot::default()})
+    }
+
+    pub fn try_default() -> Result<Self, CgroupError> {
+        Self::new(CgroupConfig::default())
+    }
+
+    fn create_and_enable_controllers(&self) -> Result<(), CgroupError> {
+        if !self.cgroup_path.exists() {
+            fs::create_dir_all(&self.cgroup_path).map_err(|e| CgroupError::Io {
+                path: self.cgroup_path.display().to_string(),
+                source: e,
+            })?;
+            debug!("Created cgroup at {}", self.cgroup_path.display());
+        }
+
+        let subtree_control = self.config.cgroup_root.join("cgroup.subtree_control");
+        for controller in ["memory", "cpu", "io"] {
+            let token = format!("+{controller}");
+            let controller_static: &'static str = match controller {
+                "memory" => "memory",
+                "cpu" => "cpu",
+                "io" => "io",
+                _ => unreachable!(),
+            };
+            fs::write(&subtree_control, &token).map_err(|e| CgroupError::EnableController {
+                controller: controller_static,
+                path: subtree_control.display().to_string(),
+                source: e,
+            })?;
+            trace!(
+                "Enabled controller `{controller}` in {}",
+                subtree_control.display()
+            );
+        }
+
+        Ok(())
+    }
+
+
+    fn attach_pid(&self, pid: i32) -> Result<(), CgroupError> {
[…]
