### 2026-05-19 c3663b1
chore(procfs): fmt, fix clippy warnings and clear sentinel values if any in to_metrics
