### 2026-05-19 4ba3035
chore(procfs): fmt, fix clippy warnings and clear sentinel values if any in to_metrics
