### 2026-05-19 013d269
fix: crash when inheriting program's stderr with non-blocking I/O fix
