+++ sources/cgroup/src/lib.rs
@@ -1,13 +1,19 @@
-use std::fs;
-use std::path::PathBuf;
-use log::{debug, trace, warn};
+use crate::error::CgroupError;
+use crate::snapshot::Snapshot;
+use crate::util::{read_flat_keyed, read_io_stat, read_u64_opt};
+use futures::StreamExt;
 use joule_profiler_core::sensor::{Sensor, Sensors};
 use joule_profiler_core::source::MetricReader;
 use joule_profiler_core::types::{Metric, Metrics};
 use joule_profiler_core::unit::{MetricUnit, Unit, UnitPrefix};
-use crate::error::CgroupError;
-use crate::snapshot::Snapshot;
-use crate::util::{read_flat_keyed, read_io_stat, read_u64_opt};
+use log::{debug, trace, warn};
+use std::fs;
+use std::path::{Path, PathBuf};
+use std::sync::Arc;
+use std::time::Duration;
+use tokio::sync::Mutex;
+use tokio::task::JoinHandle;
+use tokio_timerfd::Interval;
 
 mod error;
 mod snapshot;
@@ -30,11 +36,11 @@ const COUNT_UNIT: MetricUnit = MetricUnit {
     unit: Unit::Count,
 };
 
-
 #[derive(Debug, Clone)]
 pub struct CgroupConfig {
     pub cgroup_root: PathBuf,
     pub cgroup_name: String,
+    pub poll_interval: Option<Duration>,
 }
 
 impl Default for CgroupConfig {
@@ -42,6 +48,7 @@ impl Default for CgroupConfig {
         Self {
             cgroup_root: PathBuf::from("/sys/fs/cgroup"),
             cgroup_name: format!("joule-profiler-{}", std::process::id()),
+            poll_interval: None,
         }
     }
 }
@@ -49,7 +56,9 @@ impl Default for CgroupConfig {
 pub struct CgroupSource {
     config: CgroupConfig,
     cgroup_path: PathBuf,
-    last_snapshot: Snapshot,
+    shared_snapshot: Arc<Mutex<Snapshot>>,
+    handle: Option<JoinHandle<Result<(), CgroupError>>>,
+    last_ponctual: Snapshot,
 }
 
 impl CgroupSource {
@@ -63,7 +72,13 @@ impl CgroupSource {
 
         let cgroup_path = config.cgroup_root.join(&config.cgroup_name);
 
-        Ok(Self {config, cgroup_path,last_snapshot: Snapshot::default()})
+        Ok(Self {
+            config,
+            cgroup_path,
+            shared_snapshot: Arc::new(Mutex::new(Snapshot::default())),
+            handle: None,
+            last_ponctual: Snapshot::default(),
+        })
     }
 
     pub fn try_default() -> Result<Self, CgroupError> {
@@ -72,7 +87,7 @@ impl CgroupSource {
 
     fn create_and_enable_controllers(&self) -> Result<(), CgroupError> {
         if !self.cgroup_path.exists() {
-            fs::create_dir_all(&self.cgroup_path).map_err(|e| CgroupError::Io {
+            fs::create_dir_all(&self.cgroup_path).map_err(|e| CgroupError::IoPath {
                 path: self.cgroup_path.display().to_string(),
                 source: e,
             })?;
@@ -102,10 +117,10 @@ impl CgroupSource {
         Ok(())
     }
 
-
     fn attach_pid(&self, pid: i32) -> Result<(), CgroupError> {
         let procs_path = self.cgroup_path.join("cgroup.procs");
-        fs::write(&procs_path, pid.to_string()).map_err(|e| CgroupError::Attach { pid, source: e })?;
+        fs::write(&procs_path, pid.to_string())
[…]
