### 2026-05-26 3653e00
feat(cli): removed per-source cli option and replaced it with comma-separated list
