+++ cli/src/lib.rs
@@ -41,7 +41,7 @@ pub struct CliArgs {
     #[arg(long = "rapl-path")]
     pub rapl_path: Option<String>,
 
-    /// Sockets to measure (e.g. 0 or 0,1)
+    /// Sockets to measure (e.g., 0 or 0,1)
     #[arg(short = 's', long = "sockets")]
     pub sockets: Option<String>,
 
@@ -57,17 +57,13 @@ pub struct CliArgs {
     #[arg(short = 'o', long = "output-file")]
     pub output_file: Option<String>,
 
-    /// GPU support
-    #[arg(long)]
-    pub gpu: bool,
-
-    /// `perf_event` counters support
-    #[arg(long)]
-    pub perf: bool,
-
     /// Choose RAPL backend between powercap or perf
     #[arg(long = "rapl-backend", value_enum, default_value_t = RaplBackend::Perf)]
     pub rapl_backend: RaplBackend,
+
+    /// Sources activation list. All sources must be separated with a comma (e.g., "perf,nvml").
+    #[arg(long, value_delimiter = ',', default_value = "rapl")]
+    pub sources: Vec<Source>,
 
     /// The command to execute
     #[command(subcommand)]
@@ -98,6 +94,14 @@ impl From<CliArgs> for Config {
             rapl_path: cli_args.rapl_path,
         }
     }
+}
+
+#[derive(Clone, Debug, PartialEq, Eq, Hash, ValueEnum)]
+pub enum Source {
+    Rapl,
+    Nvml,
+    #[value(alias = "perf_event")]
+    Perf,
 }
 
 #[derive(Clone, Debug, ValueEnum)]
