### 2026-05-27 5250794
feat: procfs global memory counters (mem usage, cache, anon, swap)
