+++ sources/procfs/src/counters.rs
@@ -1,48 +1,60 @@
-#[derive(Debug, Default, Clone)]
+#[derive(Debug, Default, Clone, Copy)]
 pub struct MemoryCounters {
-    pub min_vm_size: u64,
-    pub min_rss: u64,
-    pub min_pss: u64,
-    pub min_shared: u64,
-    pub min_anon: u64,
+    pub proc: ProcMemoryCounters,
+    pub global: GlobalMemoryCounters,
+}
 
+#[derive(Debug, Default, Clone, Copy)]
+pub struct ProcMemoryCounters {
+    pub min_vm_size: u64,
     pub max_vm_size: u64,
+
+    pub min_rss: u64,
     pub max_rss: u64,
+
+    pub min_pss: u64,
     pub max_pss: u64,
+
+    pub min_shared: u64,
     pub max_shared: u64,
+
+    pub min_anon: u64,
     pub max_anon: u64,
 }
 
-impl MemoryCounters {
+impl ProcMemoryCounters {
     pub fn update(&mut self, vm_size: u64, rss: u64, pss: u64, shared: u64, anon: u64) {
         self.min_vm_size = self.min_vm_size.min(vm_size);
         self.max_vm_size = self.max_vm_size.max(vm_size);
+
         self.min_rss = self.min_rss.min(rss);
         self.max_rss = self.max_rss.max(rss);
+
         self.min_pss = self.min_pss.min(pss);
         self.max_pss = self.max_pss.max(pss);
+
         self.min_shared = self.min_shared.min(shared);
         self.max_shared = self.max_shared.max(shared);
+
         self.min_anon = self.min_anon.min(anon);
         self.max_anon = self.max_anon.max(anon);
     }
 
     pub fn reset_phase(&mut self) {
-        self.max_vm_size = 0;
-        self.max_rss = 0;
-        self.max_pss = 0;
-        self.max_shared = 0;
-        self.max_anon = 0;
-
         self.min_vm_size = u64::MAX;
-        self.min_rss = u64::MAX;
-        self.min_pss = u64::MAX;
-        self.min_shared = u64::MAX;
-        self.min_anon = u64::MAX;
-    }
+        self.max_vm_size = 0;
 
-    pub fn delta(current: u64, start: u64) -> i64 {
-        current.cast_signed() - start.cast_signed()
+        self.min_rss = u64::MAX;
+        self.max_rss = 0;
+
+        self.min_pss = u64::MAX;
+        self.max_pss = 0;
+
+        self.min_shared = u64::MAX;
+        self.max_shared = 0;
+
+        self.min_anon = u64::MAX;
+        self.max_anon = 0;
     }
 
     pub fn remove_sentinel_values(&mut self) {
@@ -62,6 +74,141 @@ impl MemoryCounters {
 }
 
 #[derive(Debug, Default, Clone, Copy)]
+pub struct GlobalMemoryCounters {
+    pub min_mem_available: Option<u64>,
+    pub max_mem_available: Option<u64>,
+
+    pub min_mem_free: u64,
+    pub max_mem_free: u64,
+
+    pub min_cached: u64,
+    pub max_cached: u64,
+
+    pub min_anon: Option<u64>,
+    pub max_anon: Option<u64>,
+
+    pub min_swap_free: u64,
+    pub max_swap_free: u64,
+}
+
+impl GlobalMemoryCounters {
+    pub fn update(
+        &mut self,
+        mem_available: Option<u64>,
+        mem_free: u64,
+        cached: u64,
+        anon: Option<u64>,
+        swap_free: u64,
+    ) {
+        if let Some(mem_available) = mem_available {
+            self.min_mem_available =
+                Some(if let Some(min_mem_available) = self.min_mem_available {
[…]
