### 2026-05-27 0b99794
feat: procfs global memory counters (mem usage, cache, anon, swap)
