+++ sources/procfs/src/lib.rs
@@ -5,8 +5,8 @@ use joule_profiler_core::{
     types::{Metric, MetricValue, Metrics},
     unit::{MetricUnit, Unit, UnitPrefix},
 };
-use log::{debug, trace};
-use procfs::process::Process;
+use log::{debug, trace, warn};
+use procfs::{Current, FromRead, process::Process};
 use std::sync::Arc;
 use std::time::Duration;
 use thiserror::Error;
@@ -14,10 +14,7 @@ use tokio::sync::Mutex;
 use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 
-use crate::{
-    counters::{Counters, IoCounters, MemoryCounters},
-    utils::collect_all_children,
-};
+use crate::{counters::Counters, utils::collect_all_children};
 
 pub mod counters;
 mod utils;
@@ -71,17 +68,31 @@ impl From<MemoryUnit> for MetricUnit {
 
 type Result<T> = std::result::Result<T, ProcfsError>;
 
+pub struct MemorySnapshot {
+    proc_vm_size: u64,
+    proc_rss: u64,
+    proc_pss: u64,
+    proc_shared: u64,
+    proc_anon: u64,
+
+    global_mem_available: Option<u64>,
+    global_mem_free: u64,
+    global_cached: u64,
+    global_anon: Option<u64>,
+    global_swap_free: u64,
+}
+
 #[derive(Debug, Default)]
 pub struct Procfs {
     pid: i32,
     poll_interval: Option<Duration>,
-    memory_counters: Arc<Mutex<MemoryCounters>>,
-    io_counters: IoCounters,
+    counters: Arc<Mutex<Counters>>,
     handle: Option<JoinHandle<Result<()>>>,
-    memory_unit: MemoryUnit,
+    proc_memory_unit: MemoryUnit,
+    global_memory_unit: MemoryUnit,
 }
 
-fn read_memory(process: &Process) -> Result<(u64, u64, u64, u64, u64)> {
+fn read_proc_memory(process: &Process) -> Result<(u64, u64, u64, u64, u64)> {
     let vm_size = process.stat()?.vsize;
     trace!("Querying process {} smaps_rollup.", process.pid);
     let smaps = process.smaps_rollup()?;
@@ -104,6 +115,19 @@ fn read_memory(process: &Process) -> Result<(u64, u64, u64, u64, u64)> {
     Ok((vm_size, rss, pss, shared, anon))
 }
 
+fn read_global_memory() -> Result<(Option<u64>, u64, u64, Option<u64>, u64)> {
+    let meminfo = procfs::Meminfo::from_file(procfs::Meminfo::PATH)?;
+    trace!("Querying global meminfo from {}", procfs::Meminfo::PATH);
+
+    Ok((
+        meminfo.mem_available,
+        meminfo.mem_free,
+        meminfo.cached,
+        meminfo.anon_pages,
+        meminfo.swap_free,
+    ))
+}
+
 fn read_io(process: &Process) -> Result<(u64, u64)> {
     trace!("Querying process {} io.", process.pid);
     let io = process.io()?;
@@ -114,9 +138,42 @@ impl Procfs {
     pub fn new(poll_interval: Option<Duration>) -> Self {
         Self {
             poll_interval,
+            global_memory_unit: MemoryUnit::Giga,
             ..Default::default()
         }
     }
+}
+
+fn read_memory_counters(pids: &[i32]) -> Result<MemorySnapshot> {
+    let (proc_vm_size, proc_rss, proc_pss, proc_shared, proc_anon) = pids
+        .iter()
+        .filter_map(|p| Process::new(*p).ok())
[…]
