### 2026-05-27 5006b03
refactor(procfs): improved procfs source code
