+++ sources/procfs/src/counters.rs
@@ -1,223 +1,154 @@
-#[derive(Debug, Default, Clone, Copy)]
-pub struct MemoryCounters {
-    pub proc: ProcMemoryCounters,
-    pub global: GlobalMemoryCounters,
-}
+use crate::snapshot::{GlobalSnapshot, ProcSnapshot};
 
-#[derive(Debug, Default, Clone, Copy)]
-pub struct ProcMemoryCounters {
-    pub min_vm_size: u64,
-    pub max_vm_size: u64,
+#[derive(Debug, Clone, Copy)]
+pub struct MinMax(pub u64, pub u64);
 
-    pub min_rss: u64,
-    pub max_rss: u64,
-
-    pub min_pss: u64,
-    pub max_pss: u64,
-
-    pub min_shared: u64,
-    pub max_shared: u64,
-
-    pub min_anon: u64,
-    pub max_anon: u64,
-}
-
-impl ProcMemoryCounters {
-    pub fn update(&mut self, vm_size: u64, rss: u64, pss: u64, shared: u64, anon: u64) {
-        self.min_vm_size = self.min_vm_size.min(vm_size);
-        self.max_vm_size = self.max_vm_size.max(vm_size);
-
-        self.min_rss = self.min_rss.min(rss);
-        self.max_rss = self.max_rss.max(rss);
-
-        self.min_pss = self.min_pss.min(pss);
-        self.max_pss = self.max_pss.max(pss);
-
-        self.min_shared = self.min_shared.min(shared);
-        self.max_shared = self.max_shared.max(shared);
-
-        self.min_anon = self.min_anon.min(anon);
-        self.max_anon = self.max_anon.max(anon);
+impl MinMax {
+    pub fn update(&mut self, value: u64) {
+        self.0 = self.0.min(value);
+        self.1 = self.1.max(value);
     }
 
-    pub fn reset_phase(&mut self) {
-        self.min_vm_size = u64::MAX;
-        self.max_vm_size = 0;
-
-        self.min_rss = u64::MAX;
-        self.max_rss = 0;
-
-        self.min_pss = u64::MAX;
-        self.max_pss = 0;
-
-        self.min_shared = u64::MAX;
-        self.max_shared = 0;
-
-        self.min_anon = u64::MAX;
-        self.max_anon = 0;
+    pub fn reset(&mut self) {
+        self.0 = u64::MAX;
+        self.1 = 0;
     }
 
-    pub fn remove_sentinel_values(&mut self) {
-        let values = [
-            &mut self.min_vm_size,
-            &mut self.min_rss,
-            &mut self.min_pss,
-            &mut self.min_shared,
-            &mut self.min_anon,
-        ];
-        for v in values {
-            if *v == u64::MAX {
-                *v = 0;
-            }
+    pub fn remove_sentinel(&mut self) {
+        if self.0 == u64::MAX {
+            self.0 = 0;
         }
+    }
+
+    pub fn min(&self) -> u64 {
+        self.0
+    }
+
+    pub fn max(&self) -> u64 {
+        self.1
+    }
+}
+
+impl Default for MinMax {
+    fn default() -> Self {
+        Self(u64::MAX, 0)
     }
 }
 
 #[derive(Debug, Default, Clone, Copy)]
-pub struct GlobalMemoryCounters {
-    pub min_mem_available: Option<u64>,
-    pub max_mem_available: Option<u64>,
+pub struct ProcCounters {
+    pub vm_size: MinMax,
+    pub rss: MinMax,
+    pub pss: MinMax,
+    pub shared: MinMax,
+    pub anon: MinMax,
 
-    pub min_mem_free: u64,
-    pub max_mem_free: u64,
-
[…]
