### 2026-05-27 b4453b2
refactor(procfs): improved procfs source code
