+++ cli/src/bin/main.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use anyhow::Result;
 use joule_profiler_cli::{
     CliArgs, ProfilerCommand, RaplBackend, Source, init_logging, output_format_to_displayer,
@@ -69,7 +71,7 @@ async fn main() -> Result<()> {
 
     if cli.sources.contains(&Source::Procfs) {
         trace!("Initializing procfs source");
-        let procfs = Procfs::default();
+        let procfs = Procfs::new(Some(Duration::from_millis(1)))?;
         profiler.add_source(procfs);
     }
 
