### 2026-05-27 69c7f8b
refactor(procfs): improved procfs source code
