### 2026-05-27 49e9b41
refactor(procfs): improved procfs source code
