+++ sources/procfs/src/lib.rs
@@ -5,18 +5,24 @@ use joule_profiler_core::{
     types::{Metric, MetricValue, Metrics},
     unit::{MetricUnit, Unit, UnitPrefix},
 };
-use log::{debug, trace, warn};
-use procfs::{Current, FromRead, process::Process};
+use log::{debug, trace};
+use procfs::{Current, FromRead, Meminfo};
 use std::sync::Arc;
 use std::time::Duration;
-use thiserror::Error;
 use tokio::sync::Mutex;
 use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 
-use crate::{counters::Counters, utils::collect_all_children};
+use crate::{
+    counters::{Counters, MinMax},
+    error::ProcfsError,
+    snapshot::{GlobalSnapshot, ProcSnapshot, read_global, read_proc},
+    utils::{MemoryUnit, collect_all_children},
+};
 
 pub mod counters;
+pub mod error;
+mod snapshot;
 mod utils;
 
 const IO_COUNTERS_METRIC_UNIT: MetricUnit = MetricUnit {
@@ -24,156 +30,33 @@ const IO_COUNTERS_METRIC_UNIT: MetricUnit = MetricUnit {
     unit: Unit::Byte,
 };
 
-#[derive(Error, Debug)]
-pub enum ProcfsError {
-    #[error("Procfs not initialized, call init first")]
-    NotInitialized,
-    #[error(transparent)]
-    Procfs(#[from] procfs::ProcError),
-    #[error(transparent)]
-    Io(#[from] std::io::Error),
-}
-
-#[derive(Debug, Default, Clone, Copy, PartialEq, Eq)]
-pub enum MemoryUnit {
-    Bytes,
-    Kilo,
-    #[default]
-    Mega,
-    Giga,
-}
-
-impl From<MemoryUnit> for MetricUnit {
-    fn from(unit: MemoryUnit) -> Self {
-        match unit {
-            MemoryUnit::Bytes => MetricUnit {
-                prefix: UnitPrefix::None,
-                unit: Unit::Byte,
-            },
-            MemoryUnit::Kilo => MetricUnit {
-                prefix: UnitPrefix::Kilo,
-                unit: Unit::Byte,
-            },
-            MemoryUnit::Mega => MetricUnit {
-                prefix: UnitPrefix::Mega,
-                unit: Unit::Byte,
-            },
-            MemoryUnit::Giga => MetricUnit {
-                prefix: UnitPrefix::Giga,
-                unit: Unit::Byte,
-            },
-        }
-    }
-}
-
 type Result<T> = std::result::Result<T, ProcfsError>;
-
-pub struct MemorySnapshot {
-    proc_vm_size: u64,
-    proc_rss: u64,
-    proc_pss: u64,
-    proc_shared: u64,
-    proc_anon: u64,
-
-    global_mem_available: Option<u64>,
-    global_mem_free: u64,
-    global_cached: u64,
-    global_anon: Option<u64>,
-    global_swap_free: u64,
-}
 
 #[derive(Debug, Default)]
 pub struct Procfs {
     pid: i32,
+
     poll_interval: Option<Duration>,
     counters: Arc<Mutex<Counters>>,
     handle: Option<JoinHandle<Result<()>>>,
+
     proc_memory_unit: MemoryUnit,
     global_memory_unit: MemoryUnit,
-}
 
-fn read_proc_memory(process: &Process) -> Result<(u64, u64, u64, u64, u64)> {
-    let vm_size = process.stat()?.vsize;
-    trace!("Querying process {} smaps_rollup.", process.pid);
-    let smaps = process.smaps_rollup()?;
-
-    let (rss, pss, anon, shared) = smaps
-        .memory_map_rollup
[…]
