### 2026-05-27 9bf4607
fix: io counters fix (ignore permission denied error after process termination)
