### 2026-05-27 8c64842
refactor(procfs): improved to_metrics clarity
