### 2026-05-27 8062a22
refactor(procfs): improved to_metrics clarity
