### 2026-05-27 facf9e3
docs(procfs): documented procfs source code
