+++ sources/procfs/src/counters.rs
@@ -1,19 +1,27 @@
 use crate::snapshot::{GlobalSnapshot, ProcSnapshot};
 
+/// A min/max accumulator for a single metric.
+///
+/// Initialized with `min = u64::MAX` and `max = 0` as sentinels,
+/// so the first [`update`](MinMax::update) call sets both bounds correctly.
+/// Call [`remove_sentinel`](MinMax::remove_sentinel) before reading min.
 #[derive(Debug, Clone, Copy)]
 pub struct MinMax(pub u64, pub u64);
 
 impl MinMax {
+    /// Updates the min/max bounds with the provided value.
     pub fn update(&mut self, value: u64) {
         self.0 = self.0.min(value);
         self.1 = self.1.max(value);
     }
 
+    /// Resets to sentinel state (`min = u64::MAX`, `max = 0`).
     pub fn reset(&mut self) {
         self.0 = u64::MAX;
         self.1 = 0;
     }
 
+    /// Replaces the `u64::MAX` sentinel with `0` if no update was received.
     pub fn remove_sentinel(&mut self) {
         if self.0 == u64::MAX {
             self.0 = 0;
@@ -35,21 +43,46 @@ impl Default for MinMax {
     }
 }
 
+/// Accumulated memory and I/O counters for a process hierarchy over a phase.
+///
+/// Memory fields track min/max over all snapshots in the phase.
+/// I/O fields track cumulative byte counts: begin is set at phase start.
+/// All metrics are in bytes.
 #[derive(Debug, Default, Clone, Copy)]
 pub struct ProcCounters {
+    /// Virtual memory size.
     pub vm_size: MinMax,
+
+    /// Resident set size.
     pub rss: MinMax,
+
+    /// Proportional set size.
     pub pss: MinMax,
+
+    /// Shared memory, clean + dirty.
     pub shared: MinMax,
+
+    /// Anonymous memory.
     pub anon: MinMax,
 
+    /// Cumulative bytes read at the beginning of a phase.
     pub begin_read_bytes: u64,
+
+    /// Cumulative bytes written at the beginning of a phase.
     pub begin_write_bytes: u64,
+
+    /// Highest cumulative bytes read observed during this phase.
     pub end_read_bytes: u64,
+
+    /// Highest cumulative bytes written observed during this phase.
     pub end_write_bytes: u64,
 }
 
 impl ProcCounters {
+    /// Merges a [`ProcSnapshot`] into the counters.
+    ///
+    /// I/O fields take the max of the current end and the snapshot value,
+    /// since `/proc/{pid}/io` counters are monotonically increasing.
     pub fn update(&mut self, snapshot: &ProcSnapshot) {
         self.vm_size.update(snapshot.vm_size);
         self.rss.update(snapshot.rss);
@@ -61,6 +94,7 @@ impl ProcCounters {
         self.end_write_bytes = self.end_write_bytes.max(snapshot.write_bytes);
     }
 
+    /// Resets memory min/max for the next phase and carries I/O end values forward as the new begin.
     pub fn reset(&mut self) {
         self.vm_size.reset();
         self.rss.reset();
@@ -72,6 +106,7 @@ impl ProcCounters {
         self.begin_write_bytes = self.end_write_bytes;
     }
 
+    /// Replaces `u64::MAX` sentinels with `0` on all memory fields before reading.
     pub fn remove_sentinels(&mut self) {
         self.vm_size.remove_sentinel();
         self.rss.remove_sentinel();
@@ -81,16 +116,31 @@ impl ProcCounters {
     }
 }
 
+/// Accumulated system-wide memory counters over a phase.
+///
+/// `mem_available` and `anon` are `Option` because they may not be present
[…]
