### 2026-05-27 0ac7caf
docs(procfs): documented procfs source code
