+++ sources/procfs/src/snapshot.rs
@@ -1,68 +1,107 @@
 use log::trace;
 use procfs::{Current, FromRead, Meminfo, process::Process};
 
-use crate::Result;
+use crate::{Result, utils::collect_all_children};
 
+/// Snapshot of memory and I/O measurements for a process hierarchy.
+///
+/// All fields are summed across the process and all its descendants.
+/// Memory values are in bytes.
 #[derive(Default)]
 pub struct ProcSnapshot {
+    /// Virtual memory size, from `/proc/{pid}/stat`.
     pub vm_size: u64,
+
+    /// Resident set size.
     pub rss: u64,
+
+    /// Proportional set size (shared pages divided by the count of process using them).
     pub pss: u64,
+
+    /// Shared memory (clean + dirty).
     pub shared: u64,
+
+    /// Anonymous memory (not backed by a file).
     pub anon: u64,
 
+    /// Cumulative bytes read since process start.
     pub read_bytes: u64,
+
+    /// Cumulative bytes written since process start.
     pub write_bytes: u64,
 }
 
-pub fn read_proc(pids: &[i32]) -> Result<ProcSnapshot> {
-    let mut snapshot = ProcSnapshot::default();
+/// Reads memory and I/O stats for a single pid and adds them into `snapshot`.
+///
+/// Silently ignores `PermissionDenied` on I/O reads, which can happen
+/// if the process exits between the smaps and io reads.
+fn read_proc_pid(pid: i32, snapshot: &mut ProcSnapshot) -> Result<()> {
+    let process = Process::new(pid)?;
 
-    for pid in pids {
-        let process = Process::new(*pid)?;
-        trace!("Querying process {} stat.", process.pid);
-        let vm_size = process.stat()?.vsize;
-        snapshot.vm_size += vm_size;
+    trace!("Querying process {} stat.", process.pid);
+    snapshot.vm_size += process.stat()?.vsize;
 
-        trace!("Querying process {} smaps_rollup.", process.pid);
-        let smaps = process.smaps_rollup()?;
-        let map = if let Some(entry) = smaps.memory_map_rollup.0.first() {
-            &entry.extension.map
-        } else {
-            continue;
-        };
-
+    trace!("Querying process {} smaps_rollup.", process.pid);
+    let smaps = process.smaps_rollup()?;
+    if let Some(entry) = smaps.memory_map_rollup.0.first() {
+        let map = &entry.extension.map;
         snapshot.rss += map.get("Rss").copied().unwrap_or(0);
         snapshot.pss += map.get("Pss").copied().unwrap_or(0);
         snapshot.anon += map.get("Anonymous").copied().unwrap_or(0);
         snapshot.shared += map.get("Shared_Clean").copied().unwrap_or(0)
             + map.get("Shared_Dirty").copied().unwrap_or(0);
-
-        trace!("Querying process {} io.", process.pid);
-
-        // Match because procfs throws permission denied if process terminated.
-        match process.io() {
-            Ok(io) => {
-                snapshot.read_bytes += io.rchar;
-                snapshot.write_bytes += io.wchar;
-            }
-            Err(procfs::ProcError::PermissionDenied(_)) => {}
-            Err(err) => return Err(err.into()),
-        }
     }
 
+    trace!("Querying process {} io.", process.pid);
+    match process.io() {
+        Ok(io) => {
+            snapshot.read_bytes += io.rchar;
+            snapshot.write_bytes += io.wchar;
+        }
+        Err(procfs::ProcError::PermissionDenied(_)) => {}
+        Err(err) => return Err(err.into()),
+    }
+
+    Ok(())
+}
+
[…]
