### 2026-05-27 fcd1d42
docs(procfs): documented procfs source code
