+++ sources/procfs/src/lib.rs
@@ -1,3 +1,8 @@
+//! Procfs metric source for Joule Profiler.
+//!
+//! Provides memory and I/O metrics for a process and it's children and system-wide,
+//! by reading Linux's `/proc` filesystem via the `procfs` crate.
+
 use futures::StreamExt;
 use joule_profiler_core::{
     sensor::{Sensor, Sensors},
@@ -14,12 +19,14 @@ use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 
 use crate::{
+    config::ProcfsConfig,
     counters::{Counters, MinMax},
     error::ProcfsError,
-    snapshot::{GlobalSnapshot, ProcSnapshot, read_global, read_proc},
-    utils::{MemoryUnit, collect_all_children, make_conversion},
+    snapshot::{ProcSnapshot, measure_global, measure_proc},
+    utils::{MemoryUnit, make_conversion},
 };
 
+pub mod config;
 pub mod counters;
 pub mod error;
 mod snapshot;
@@ -32,28 +39,61 @@ const IO_COUNTERS_METRIC_UNIT: MetricUnit = MetricUnit {
 
 type Result<T> = std::result::Result<T, ProcfsError>;
 
+/// Procfs-based metric source.
+///
+/// Reads memory and I/O metrics from `/proc` for a target process and its
+/// entire child hierarchy, as well as system-wide memory statistics.
+///
+/// Metrics are accumulated as min/max over each measurement phase.
+/// A phase starts on [`MetricReader::measure`] and ends on [`MetricReader::retrieve`],
+/// which returns the counters and resets them for the next phase.
+///
+/// If a `poll_interval` is configured, a background task polls `/proc` continuously
+/// at that interval in addition to explicit [`MetricReader::measure`] calls.
+///
+/// # Example
+///
+/// ```no_run
+/// use joule_profiler_source_procfs::{Procfs, config::ProcfsConfig};
+///
+/// let source = Procfs::new(&ProcfsConfig::default()).unwrap();
+/// ```
 #[derive(Debug, Default)]
 pub struct Procfs {
+    /// pid of the profiled process, initialized at -1.
     pid: i32,
 
+    /// The polling interval of the source if set, else no polling.
     poll_interval: Option<Duration>,
+
+    /// Current metrics counters.
     counters: Arc<Mutex<Counters>>,
+
+    /// The handle to the polling task.
     handle: Option<JoinHandle<Result<()>>>,
 
+    /// The unit in which to convert proc memory metrics.
     proc_memory_unit: MemoryUnit,
+
+    /// The unit in which to convert global memory metrics.
     global_memory_unit: MemoryUnit,
 
+    /// Total physical memory in bytes, read once at construction from `/proc/meminfo`.
     mem_total: u64,
 }
 
 impl Procfs {
-    pub fn new(poll_interval: Option<Duration>) -> Result<Self> {
+    /// Creates a new `Procfs` source from the given configuration.
+    ///
+    /// Reads `MemTotal` from `/proc/meminfo` at construction time.
+    pub fn new(config: &ProcfsConfig) -> Result<Self> {
         let mem_total = Meminfo::from_file(Meminfo::PATH)?.mem_total;
         Ok(Self {
             pid: -1,
             mem_total,
-            poll_interval,
-            global_memory_unit: MemoryUnit::Giga,
+            poll_interval: config.poll_interval,
+            global_memory_unit: config.global_memory_unit,
+            proc_memory_unit: config.proc_memory_unit,
             ..Default::default()
         })
     }
@@ -63,6 +103,7 @@ impl MetricReader for Procfs {
     type Type = Counters;
     type Error = ProcfsError;
 
[…]
