### 2026-05-27 ef7b821
feat(procfs): added config struct
