### 2026-05-27 894bc3c
feat(procfs): added config struct
