### 2026-05-27 18fab0a
chore: forbid source duplication in cli --sources
