### 2026-05-27 492bc64
chore: forbid source duplication in cli --sources
