### 2026-05-27 604ff21
chore: move configuration directly into procfs struct
