### 2026-05-27 769c1c5
chore: move configuration directly into procfs struct
