### 2026-05-28 ba98b6e
refactor(cgroup): source refactoring, separation into modules and improved logic
