### 2026-05-28 3e6c44f
refactor(cgroup): source refactoring, separation into modules and improved logic
