### 2026-05-28 257fd91
refactor(cgroup): source refactoring, separation into modules and improved logic
