+++ sources/cgroup/src/counters.rs
@@ -1,0 +1,192 @@
+use crate::snapshot::{CpuSnapshot, IoSnapshot, MemorySnapshot};
+
+/// A min/max accumulator for a single metric.
+#[derive(Debug, Clone, Copy, Default)]
+pub struct MinMax(Option<u64>, Option<u64>);
+
+impl MinMax {
+    /// Updates the min/max bounds with the provided value.
+    pub fn update(&mut self, value: u64) {
+        self.0 = Some(self.0.map_or(value, |m| m.min(value)));
+        self.1 = Some(self.1.map_or(value, |m| m.max(value)));
+    }
+
+    pub fn reset(&mut self) {
+        self.0 = None;
+        self.1 = None;
+    }
+
+    pub fn min(&self) -> Option<u64> {
+        self.0
+    }
+
+    pub fn max(&self) -> Option<u64> {
+        self.1
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy)]
+pub struct BeginEnd(u64, u64);
+
+impl BeginEnd {
+    pub fn update(&mut self, value: u64) {
+        self.1 = value;
+    }
+
+    pub fn diff(&self) -> u64 {
+        self.1.saturating_sub(self.0)
+    }
+
+    pub fn new_phase(&mut self) {
+        self.0 = self.1;
+    }
+}
+
+trait NewPhase {
+    fn new_phase(&mut self);
+}
+
+impl NewPhase for BeginEnd {
+    fn new_phase(&mut self) {
+        self.0 = self.1;
+    }
+}
+
+impl NewPhase for Option<BeginEnd> {
+    fn new_phase(&mut self) {
+        if let Some(v) = self {
+            v.new_phase();
+        }
+    }
+}
+
+trait UpdateOpt {
+    fn update(&mut self, value: Option<u64>);
+}
+
+impl UpdateOpt for Option<BeginEnd> {
+    fn update(&mut self, value: Option<u64>) {
+        if let Some(v) = value {
+            self.get_or_insert_default().update(v);
+        }
+    }
+}
+
+impl UpdateOpt for Option<MinMax> {
+    fn update(&mut self, value: Option<u64>) {
+        if let Some(v) = value {
+            self.get_or_insert_default().update(v);
+        }
+    }
+}
+
+#[derive(Debug, Default, Clone, Copy)]
+pub struct MemoryCounters {
+    pub current: MinMax,
+    pub swap_current: Option<MinMax>,
+    pub anon: Option<MinMax>,
+    pub file: Option<MinMax>,
+    pub peak: Option<MinMax>,
+    pub shmem: Option<MinMax>,
+    pub kernel: Option<MinMax>,
+    pub kernel_stack: Option<MinMax>,
+    pub slab: Option<MinMax>,
+}
+
+impl MemoryCounters {
+    pub fn update(&mut self, snapshot: &MemorySnapshot) {
+        self.current.update(snapshot.current);
+        self.swap_current.update(snapshot.swap_current);
+        self.anon.update(snapshot.anon);
+        self.file.update(snapshot.file);
+        self.peak.update(snapshot.peak);
+        self.shmem.update(snapshot.shmem);
+        self.kernel.update(snapshot.kernel);
+        self.kernel_stack.update(snapshot.kernel_stack);
+        self.slab.update(snapshot.slab);
+    }
+
+    pub fn reset(&mut self) {
+        self.current.reset();
+        self.swap_current = None;
+        self.anon = None;
+        self.file = None;
+        self.peak = None;
+        self.shmem = None;
+        self.kernel = None;
+        self.kernel_stack = None;
+        self.slab = None;
+    }
+}
+
[…]
