### 2026-05-28 1382c2c
refactor: removed sentinel values and use option instead in a more idiomatic way
