### 2026-05-28 96e52c5
refactor: removed sentinel values and use option instead in a more idiomatic way
