### 2026-05-28 680da9c
refactor(procfs): improved procfs polling with cancellation token to gracefully stop the tokio task
