+++ sources/procfs/Cargo.toml
@@ -17,6 +17,7 @@ futures.workspace = true
 
 procfs = "0.18.0"
 tokio-timerfd = "0.2.0"
+tokio-util = "0.7.18"
 
 [dev-dependencies]
 mockall.workspace = true
