### 2026-05-28 df6def0
refactor(procfs): improved procfs polling with cancellation token to gracefully stop the tokio task
