### 2026-05-29 7a15039
refactor: split root cgroup and child cgroups in two separate structs
