+++ sources/cgroup/src/lib.rs
@@ -13,7 +13,7 @@ use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
-use crate::cgroup::{Cgroup, Controller};
+use crate::cgroup::{Cgroup, Controller, RootCgroup, StatsReader};
 use crate::counters::{Counters, CpuCounters, IoCounters, MemoryCounters};
 use crate::error::CgroupError;
 
@@ -67,32 +67,40 @@ impl Default for CgroupConfig {
 pub struct CgroupSource {
     config: CgroupConfig,
     handle: Option<WorkerHandle>,
-    cgroup: Arc<Cgroup>,
-    shared_memory_counters: Arc<Mutex<MemoryCounters>>,
-    cpu_counters: CpuCounters,
-    io_counters: IoCounters,
+    proc_cgroup: Arc<Cgroup>,
+    root_cgroup: Arc<RootCgroup>,
+    proc_memory_counters: Arc<Mutex<MemoryCounters>>,
+    global_memory_counters: Arc<Mutex<MemoryCounters>>,
+    proc_cpu_counters: CpuCounters,
+    global_cpu_counters: CpuCounters,
+    proc_io_counters: IoCounters,
+    global_io_counters: IoCounters,
 }
 
 impl CgroupSource {
     pub fn new(config: CgroupConfig) -> Result<Self> {
-        let cgroup = Arc::new(Cgroup::new(
-            config.cgroup_root.clone(),
-            &config.cgroup_name,
-        )?);
+        let root_cgroup = Arc::new(RootCgroup::new(config.cgroup_root.clone()));
+        let proc_cgroup = Arc::new(root_cgroup.child(&config.cgroup_name));
 
         Ok(Self {
             config,
             handle: None,
-            cgroup,
-            shared_memory_counters: Arc::default(),
-            cpu_counters: CpuCounters::default(),
-            io_counters: IoCounters::default(),
+            root_cgroup,
+            proc_cgroup,
+            proc_memory_counters: Arc::default(),
+            global_memory_counters: Arc::default(),
+            proc_cpu_counters: CpuCounters::default(),
+            global_cpu_counters: CpuCounters::default(),
+            proc_io_counters: IoCounters::default(),
+            global_io_counters: IoCounters::default(),
         })
     }
 
     pub fn create_worker(
-        cgroup: Arc<Cgroup>,
-        shared_memory_counters: Arc<Mutex<MemoryCounters>>,
+        root_cgroup: Arc<RootCgroup>,
+        proc_cgroup: Arc<Cgroup>,
+        proc_memory_counters: Arc<Mutex<MemoryCounters>>,
+        global_memory_counters: Arc<Mutex<MemoryCounters>>,
         poll_interval: Duration,
     ) -> Result<WorkerHandle> {
         let mut ticker = Interval::new_interval(poll_interval)?;
@@ -107,9 +115,16 @@ impl CgroupSource {
                 tokio::select! {
                     _ = ticker.next() => {
                         trace!("Polled cgroup source.");
-                        let memory_snapshot = cgroup.read_memory()?;
-                        let mut lock = shared_memory_counters.lock().await;
-                        lock.update(&memory_snapshot);
+                        {
+                            let snapshot = proc_cgroup.stats().memory()?;
+                            let mut lock = proc_memory_counters.lock().await;
+                            lock.update(&snapshot);
+                        }
+                        {
+                            let snapshot = root_cgroup.stats().memory()?;
+                            let mut lock = global_memory_counters.lock().await;
+                            lock.update(&snapshot);
+                        }
                     }
 
[…]
