### 2026-05-29 d49a707
refactor: split root cgroup and child cgroups in two separate structs
