+++ sources/cgroup/src/cgroup.rs
@@ -1,12 +1,16 @@
-use std::{fmt::Display, fs, path::PathBuf};
+use std::{
+    fmt::Display,
+    fs,
+    path::{Path, PathBuf},
+};
 
 use log::{debug, warn};
 
 use crate::{
     Result,
-    error::CgroupError::{self},
+    error::CgroupError,
     snapshot::{CpuSnapshot, IoSnapshot, MemorySnapshot},
-    util::{read_flat_keyed_file, read_io_stat, read_u64, read_u64_opt},
+    util::{read_flat_keyed_file, read_io_stat, read_u64_opt},
 };
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
@@ -26,32 +30,31 @@ impl Display for Controller {
     }
 }
 
-pub struct Cgroup {
-    cgroup_root: PathBuf,
-    cgroup_path: PathBuf,
+pub trait StatsReader {
+    fn memory(&self) -> Result<MemorySnapshot>;
+    fn cpu(&self) -> Result<CpuSnapshot>;
+    fn io(&self) -> Result<IoSnapshot>;
 }
 
-impl Cgroup {
-    pub fn new(cgroup_root: PathBuf, cgroup_name: &str) -> Result<Self> {
-        let cgroup_path = cgroup_root.join(cgroup_name);
-        if cgroup_path.exists() {
-            return Err(CgroupError::AlreadyCreated(cgroup_name.to_owned()));
+pub struct RootCgroup {
+    path: PathBuf,
+}
+
+impl Default for RootCgroup {
+    fn default() -> Self {
+        Self {
+            path: PathBuf::from("/sys/fs/cgroup"),
         }
-
-        Ok(Self {
-            cgroup_root,
-            cgroup_path,
-        })
     }
+}
 
-    pub fn initialize_cgroup(&self) -> Result<()> {
-        debug!("Creating cgroup at \"{}\"", self.cgroup_path.display());
-        fs::create_dir_all(&self.cgroup_path)?;
-        Ok(())
+impl RootCgroup {
+    pub fn new(path: PathBuf) -> Self {
+        Self { path }
     }
 
     pub fn activate_controller(&self, controller: Controller) -> Result<()> {
-        let subtree_control_path = self.cgroup_root.join("cgroup.subtree_control");
+        let subtree_control_path = self.path.join("cgroup.subtree_control");
         fs::write(&subtree_control_path, format!("+{controller}")).map_err(|err| {
             CgroupError::FailedToCreateController {
                 controller,
@@ -62,53 +65,114 @@ impl Cgroup {
         Ok(())
     }
 
+    pub fn child(&self, name: &str) -> Cgroup {
+        Cgroup {
+            path: self.path.join(name),
+            root: self.path.clone(),
+        }
+    }
+
+    pub fn stats(&self) -> CgroupStat<'_> {
+        CgroupStat { path: &self.path }
+    }
+}
+
+pub struct Cgroup {
+    path: PathBuf,
+    root: PathBuf,
+}
+
+impl Cgroup {
+    pub fn initialize(&self) -> Result<()> {
+        debug!("Creating cgroup at \"{}\"", self.path.display());
+        fs::create_dir_all(&self.path)?;
+        Ok(())
+    }
+
     pub fn attach_pid(&self, pid: i32) -> Result<()> {
-        let procs_path = self.cgroup_path.join("cgroup.procs");
+        let procs_path = self.path.join("cgroup.procs");
         fs::write(&procs_path, pid.to_string()).map_err(|err| CgroupError::FailedToAttachPid {
             pid,
             path: procs_path,
             source: err,
         })?;
-        debug!(
-            "Attached PID {pid} to cgroup {}",
[…]
