### 2026-05-29 5b64c7b
refactor: split root cgroup and child cgroups in two separate structs
