### 2026-05-29 5929035
docs(cgroup): documented cgroup source
