+++ sources/cgroup/src/counters.rs
@@ -25,24 +25,30 @@ impl MinMax {
     }
 }
 
+/// Tracks the beginning and latest values of a cumulative counter for a phase.
 #[derive(Debug, Default, Clone, Copy)]
 pub struct BeginEnd(u64, u64);
 
 impl BeginEnd {
+    /// Updates the latest counter value.
     pub fn update(&mut self, value: u64) {
         self.1 = value;
     }
 
+    /// Returns the difference between end and begin values.
     pub fn diff(&self) -> u64 {
         self.1.saturating_sub(self.0)
     }
 
+    /// Starts a new measurement phase by setting begin to end.
     pub fn new_phase(&mut self) {
         self.0 = self.1;
     }
 }
 
+/// Internal trait for phase-based counters.
 trait NewPhase {
+    /// Starts a new measurement phase.
     fn new_phase(&mut self);
 }
 
@@ -60,7 +66,9 @@ impl NewPhase for Option<BeginEnd> {
     }
 }
 
+/// Internal trait for updating optional counters.
 trait UpdateOpt {
+    /// Updates the counter if a value is present.
     fn update(&mut self, value: Option<u64>);
 }
 
@@ -82,18 +90,36 @@ impl UpdateOpt for Option<MinMax> {
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct MemoryCounters {
+    /// Current memory usage.
     pub current: Option<MinMax>,
+
+    /// Current swap usage.
     pub swap_current: Option<MinMax>,
+
+    /// Anonymous memory usage (stack, heap, anon memory mappings).
     pub anon: Option<MinMax>,
+
+    /// File-backed memory usage.
     pub file: Option<MinMax>,
+
+    /// Peak memory usage.
     pub peak: Option<MinMax>,
+
+    /// Shared memory usage.
     pub shmem: Option<MinMax>,
+
+    /// Kernel memory usage.
     pub kernel: Option<MinMax>,
+
+    /// Kernel stack usage.
     pub kernel_stack: Option<MinMax>,
+
+    /// Slab allocator usage.
     pub slab: Option<MinMax>,
 }
 
 impl MemoryCounters {
+    /// Updates counters from a memory snapshot.
     pub fn update(&mut self, snapshot: &MemorySnapshot) {
         self.current.update(snapshot.current);
         self.swap_current.update(snapshot.swap_current);
@@ -106,6 +132,7 @@ impl MemoryCounters {
         self.slab.update(snapshot.slab);
     }
 
+    /// Resets all tracked memory metrics.
     pub fn reset(&mut self) {
         self.current = None;
         self.swap_current = None;
@@ -121,18 +148,33 @@ impl MemoryCounters {
 
 #[derive(Debug, Default, Clone, Copy)]
 pub struct CpuCounters {
+    /// Total CPU usage time.
     pub usage_usec: BeginEnd,
+
+    /// User-space CPU usage time.
     pub user_usec: BeginEnd,
+
+    /// Kernel-space CPU usage time.
     pub system_usec: BeginEnd,
 
+    /// Number of scheduler periods.
     pub nr_periods: Option<BeginEnd>,
+
+    /// Number of throttled periods.
     pub nr_throttled: Option<BeginEnd>,
+
+    /// Total throttled time.
     pub throttled_usec: Option<BeginEnd>,
+
+    /// Number of CPU bursts.
     pub nr_bursts: Option<BeginEnd>,
+
+    /// Total burst time.
     pub burst_usec: Option<BeginEnd>,
 }
 
 impl CpuCounters {
+    /// Updates counters from a CPU snapshot.
     pub fn update(&mut self, snapshot: &CpuSnapshot) {
[…]
