+++ sources/cgroup/src/cgroup.rs
@@ -1,4 +1,13 @@
+//! cgroup v2 management utilities.
+//!
+//! This module provides:
+//! - controller activation (`cpu`, `memory`, `io`);
+//! - child cgroup creation and cleanup;
+//! - process attachment to cgroups;
+//! - CPU, memory, and I/O statistics collection.
+
 use std::{
+    collections::HashSet,
     fmt::Display,
     fs,
     path::{Path, PathBuf},
@@ -13,10 +22,16 @@ use crate::{
     util::{read_flat_keyed_file, read_io_stat, read_u64_opt},
 };
 
+/// Available cgroup v2 controllers.
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
 pub enum Controller {
+    /// I/O controller.
     Io,
-    Mem,
+
+    /// Memory controller.
+    Memory,
+
+    /// CPU controller.
     Cpu,
 }
 
@@ -24,47 +39,71 @@ impl Display for Controller {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_str(match self {
             Controller::Io => "io",
-            Controller::Mem => "memory",
+            Controller::Memory => "memory",
             Controller::Cpu => "cpu",
         })
     }
 }
 
+/// Interface for reading cgroup statistics.
 pub trait StatsReader {
+    /// Returns memory statistics.
     fn memory(&self) -> Result<MemorySnapshot>;
+
+    /// Returns CPU statistics.
     fn cpu(&self) -> Result<CpuSnapshot>;
+
+    /// Returns I/O statistics.
     fn io(&self) -> Result<IoSnapshot>;
 }
 
+/// Root cgroup manager.
+///
+/// Used to enable controllers and create child cgroups.
 pub struct RootCgroup {
+    /// The path of the root cgroup, default is `/sys/fs/cgroup` (override for testing).
     path: PathBuf,
+
+    /// List of controllers activated for the root cgroup.
+    controllers: HashSet<Controller>,
 }
 
 impl Default for RootCgroup {
     fn default() -> Self {
         Self {
             path: PathBuf::from("/sys/fs/cgroup"),
+            controllers: HashSet::new(),
         }
     }
 }
 
 impl RootCgroup {
+    /// Creates a new root cgroup manager from a custom path.
     pub fn new(path: PathBuf) -> Self {
-        Self { path }
+        Self {
+            path,
+            ..Default::default()
+        }
     }
 
+    /// Enables a controller in `cgroup.subtree_control`.
+    ///
+    /// If the provided controller is already activated, then nothing happens.
     pub fn activate_controller(&self, controller: Controller) -> Result<()> {
-        let subtree_control_path = self.path.join("cgroup.subtree_control");
-        fs::write(&subtree_control_path, format!("+{controller}")).map_err(|err| {
-            CgroupError::FailedToCreateController {
-                controller,
-                path: subtree_control_path,
-                source: err,
-            }
-        })?;
+        if !self.controllers.contains(&controller) {
+            let subtree_control_path = self.path.join("cgroup.subtree_control");
+            fs::write(&subtree_control_path, format!("+{controller}")).map_err(|err| {
+                CgroupError::FailedToCreateController {
+                    controller,
+                    path: subtree_control_path,
+                    source: err,
+                }
+            })?;
+        }
         Ok(())
     }
 
+    /// Creates a child cgroup handle.
     pub fn child(&self, name: &str) -> Cgroup {
[…]
