### 2026-05-29 1a1510c
docs(cgroup): documented cgroup source
