### 2026-05-29 05cd057
docs(cgroup): documented cgroup source
