### 2026-05-29 22bc067
docs(example): update subcommand name in usage example from `phases` to `profile`
