+++ sources/cgroup/src/cgroup.rs
@@ -62,7 +62,7 @@ pub trait StatsReader {
 /// Used to enable controllers and create child cgroups.
 pub struct RootCgroup {
     /// The path of the root cgroup, default is `/sys/fs/cgroup` (override for testing).
-    path: PathBuf,
+    pub path: PathBuf,
 
     /// List of controllers activated for the root cgroup.
     controllers: HashSet<Controller>,
@@ -120,10 +120,10 @@ impl RootCgroup {
 /// Represents a child cgroup.
 pub struct Cgroup {
     /// The path of the cgroup, normally it is `/sys/fs/cgroup/{cgroup_name}` but it can be override for testing.
-    path: PathBuf,
+    pub path: PathBuf,
 
     /// The path of the root cgroup, default is `/sys/fs/cgroup` (override for testing).
-    root: PathBuf,
+    pub root: PathBuf,
 }
 
 impl Cgroup {
@@ -248,5 +248,164 @@ impl StatsReader for CgroupStat<'_> {
             nr_bursts: cpu_stat.remove("nr_bursts"),
             burst_usec: cpu_stat.remove("burst_usec"),
         })
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::{fs, io::Write, path::PathBuf};
+
+    fn temp_cgroup_dir(name: &str) -> PathBuf {
+        let mut path = std::env::temp_dir();
+        path.push(format!("cgroup_test_{name}"));
+        let _ = fs::create_dir_all(&path);
+        path
+    }
+
+    fn write_file(path: &Path, content: &str) {
+        let mut f = fs::File::create(path).unwrap();
+        writeln!(f, "{content}").unwrap();
+    }
+
+    fn setup_root(name: &str) -> RootCgroup {
+        let root = temp_cgroup_dir(name);
+        RootCgroup::new(root)
+    }
+
+    #[test]
+    fn test_initialize_and_attach_pid() {
+        let root = setup_root("init");
+        let cg = root.child("test_group");
+
+        cg.initialize().unwrap();
+
+        let pid_file = cg.path.join("cgroup.procs");
+        write_file(&pid_file, "1234");
+
+        cg.attach_pid(42).unwrap();
+
+        let content = fs::read_to_string(&pid_file).unwrap();
+        assert!(content.contains("42"));
+
+        let _ = cg.cleanup();
+    }
+
+    #[test]
+    fn test_pids_parsing() {
+        let root = setup_root("pids");
+        let cg = root.child("pids_group");
+
+        cg.initialize().unwrap();
+
+        write_file(&cg.path.join("cgroup.procs"), "1\n2\n3");
+
+        let pids = cg.pids().unwrap();
+        assert_eq!(pids, vec![1, 2, 3]);
+
+        let _ = cg.cleanup();
+    }
+
+    #[test]
+    fn test_cleanup_moves_pids() {
+        let root = setup_root("cleanup");
+        let cg = root.child("cleanup_group");
+
+        cg.initialize().unwrap();
+
+        let root_procs = root.path.join("cgroup.procs");
+        write_file(&cg.path.join("cgroup.procs"), "99");
+
+        cg.cleanup().unwrap();
+
+        let moved = fs::read_to_string(root_procs).unwrap();
+        assert!(moved.contains("99"));
+    }
+
+    #[test]
+    fn test_memory_stats() {
+        let root = setup_root("mem");
+        let cg = root.child("mem_group");
+
+        cg.initialize().unwrap();
+
+        write_file(&cg.path.join("memory.stat"), "anon 100\nfile 200\nslab 300");
+        write_file(&cg.path.join("memory.current"), "123");
[…]
