### 2026-05-29 06c48d7
docs: update sources readme
