+++ sources/nvml/README.md
@@ -9,8 +9,6 @@ This crate implements the `MetricSource` trait from `joule-profiler-core` by que
 NVML is the C-based API used internally by 'nvidia-smi'. It provides direct access to the NVIDIA GPU driver and exposes hardware energy counters, power draw, utilisation, clock speeds, memory usage, and more.
 `joule-profiler` uses NVML exclusively for energy measurement: it reads the cumulative energy counter (nvmlDeviceGetTotalEnergyConsumption) at each phase boundary and reports the delta in millijoules (mJ)
 
----
-
 ## Metrics collected
 
 | Metric | Unit | Description |
@@ -26,8 +24,3 @@ NVML is the C-based API used internally by 'nvidia-smi'. It provides direct acce
 | GPU | NVIDIA GPU |
 | Driver | NVIDIA driver with `libnvidia-ml.so` (included in standard driver packages) |
 | Permissions | Typically no extra permissions beyond driver access |
-
-
-## See also
-
-> Main project: [joule-profiler](https://github.com/joule-profiler/joule-profiler)
