### 2026-05-29 d5dda32
docs: update sources readme
