### 2026-05-29 9a1853e
docs(cgroup): rename IO metrics in README to match the exported one
