### 2026-06-01 5eef460
doc: Add source link to nbody.c comments
Added source link for nbody program

closes #29
