### 2026-06-01 80dafc9
test: added test for cgroup, mocked cgroup backend to abstract from sysfs
