+++ sources/cgroup/src/lib.rs
@@ -19,7 +19,7 @@ use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
-use crate::cgroup::{Cgroup, RootCgroup, StatsReader};
+use crate::cgroup::{Cgroup, CgroupBackend, SysFsBackend};
 use crate::counters::{Counters, CpuCounters, IoCounters, MemoryCounters};
 use crate::error::CgroupError;
 
@@ -56,11 +56,14 @@ pub(crate) type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
 ///
 /// Owns the process cgroup and root cgroup handles, and maintains
 /// internal counters for both process-level and system-wide metrics.
-pub struct CgroupSource {
+pub struct CgroupSource<B = SysFsBackend>
+where
+    B: CgroupBackend,
+{
     config: CgroupConfig,
     handle: Option<WorkerHandle>,
-    pub proc_cgroup: Arc<Cgroup>,
-    pub root_cgroup: Arc<RootCgroup>,
+    pub proc_cgroup: Arc<Cgroup<B>>,
+    pub root_cgroup: Arc<Cgroup<B>>,
     proc_memory_counters: Arc<Mutex<MemoryCounters>>,
     global_memory_counters: Arc<Mutex<MemoryCounters>>,
     proc_cpu_counters: CpuCounters,
@@ -76,14 +79,18 @@ impl CgroupSource {
     ///
     /// Initializes internal cgroup handles but does initialize and attach any PID yet.
     pub fn new(config: CgroupConfig) -> Result<Self> {
-        let root_cgroup = Arc::new(RootCgroup::new(config.cgroup_root.clone()));
-        let proc_cgroup = Arc::new(root_cgroup.child(&config.cgroup_name));
+        let root_cgroup = if let Some(cgroup_root) = &config.cgroup_root {
+            Cgroup::at(cgroup_root.clone())
+        } else {
+            Cgroup::root()
+        };
+        let proc_cgroup = root_cgroup.child(&config.cgroup_name);
 
         Ok(Self {
             config,
             handle: None,
-            root_cgroup,
-            proc_cgroup,
+            root_cgroup: Arc::new(root_cgroup),
+            proc_cgroup: Arc::new(proc_cgroup),
             proc_memory_counters: Arc::default(),
             global_memory_counters: Arc::default(),
             proc_cpu_counters: CpuCounters::default(),
@@ -94,14 +101,16 @@ impl CgroupSource {
             end_timestamp: 0,
         })
     }
+}
 
+impl<B: CgroupBackend> CgroupSource<B> {
     /// Creates a background worker that periodically samples memory usage.
     ///
     /// This worker updates process and global memory counters at a fixed interval.
     /// It can be cancelled using the returned `CancellationToken`.
     pub fn create_worker(
-        root_cgroup: Arc<RootCgroup>,
-        proc_cgroup: Arc<Cgroup>,
+        root_cgroup: Arc<Cgroup<B>>,
+        proc_cgroup: Arc<Cgroup<B>>,
         proc_memory_counters: Arc<Mutex<MemoryCounters>>,
         global_memory_counters: Arc<Mutex<MemoryCounters>>,
         poll_interval: Duration,
@@ -144,7 +153,7 @@ impl CgroupSource {
     }
 }
 
-impl MetricReader for CgroupSource {
+impl<B: CgroupBackend> MetricReader for CgroupSource<B> {
     type Type = Counters;
     type Error = CgroupError;
 
@@ -155,12 +164,7 @@ impl MetricReader for CgroupSource {
     /// - attaches the target PID
     /// - optionally starts background polling worker
[…]
