### 2026-06-01 a20a24d
test: added test for cgroup, mocked cgroup backend to abstract from sysfs
