+++ sources/cgroup/src/cgroup.rs
@@ -6,12 +6,7 @@
 //! - process attachment to cgroups;
 //! - CPU, memory, and I/O statistics collection.
 
-use std::{
-    collections::HashSet,
-    fmt::Display,
-    fs,
-    path::{Path, PathBuf},
-};
+use std::{collections::HashSet, fmt::Display, fs, path::PathBuf};
 
 use log::{debug, warn};
 
@@ -46,7 +41,10 @@ impl Display for Controller {
 }
 
 /// Interface for reading cgroup statistics.
-pub trait StatsReader {
+pub trait CgroupBackend: Send + Sync + 'static {
+    /// Initializes the backend.
+    fn initialize(&self, pid: i32, controllers: &HashSet<Controller>) -> Result<()>;
+
     /// Returns memory statistics.
     fn memory(&self) -> Result<MemorySnapshot>;
 
@@ -55,82 +53,114 @@ pub trait StatsReader {
 
     /// Returns I/O statistics.
     fn io(&self) -> Result<IoSnapshot>;
+
+    /// Cleanup the backend.
+    fn cleanup(&self) -> Result<()>;
 }
 
-/// Root cgroup manager.
-///
-/// Used to enable controllers and create child cgroups.
-pub struct RootCgroup {
-    /// The path of the root cgroup, default is `/sys/fs/cgroup` (override for testing).
+/// A cgroup node. If `parent` is `None`, this is the root cgroup.
+pub struct Cgroup<B: CgroupBackend = SysFsBackend> {
     pub path: PathBuf,
-
-    /// List of controllers activated for the root cgroup.
-    controllers: HashSet<Controller>,
+    parent: Option<PathBuf>,
+    backend: B,
 }
 
-impl Default for RootCgroup {
-    fn default() -> Self {
-        Self {
-            path: PathBuf::from("/sys/fs/cgroup"),
-            controllers: HashSet::new(),
-        }
+impl Cgroup {
+    /// Gets the root cgroup in sys fs `/sys/fs/cgroup`.
+    pub fn root() -> Self {
+        let path = PathBuf::from("/sys/fs/cgroup");
+        Self::at(path)
     }
-}
 
-impl RootCgroup {
-    /// Creates a new root cgroup manager from a custom path.
-    pub fn new(path: PathBuf) -> Self {
+    /// Builds a cgroup handle based on the provided directory.
+    pub fn at(path: PathBuf) -> Self {
+        let backend = SysFsBackend {
+            path: path.clone(),
+            root: path.clone(),
+        };
         Self {
             path,
-            ..Default::default()
+            parent: None,
+            backend,
         }
     }
 
+    /// Gets a child handle of the current cgroup based on its name.
+    pub fn child(&self, name: &str) -> Cgroup {
+        let child_path = self.path.join(name);
+        Cgroup {
+            parent: Some(self.path.clone()),
+            backend: SysFsBackend {
+                root: self.path.clone(),
+                path: child_path.clone(),
+            },
+            path: child_path,
+        }
+    }
+}
+
+impl<B: CgroupBackend> Cgroup<B> {
+    pub fn new(path: PathBuf, parent: Option<PathBuf>, backend: B) -> Self {
+        Self {
+            path,
+            parent,
+            backend,
+        }
+    }
+
+    /// True if the cgroup is the root cgroup, else false.
+    pub fn is_root(&self) -> bool {
+        self.parent.is_none()
+    }
+
+    /// Return the backend for stats querying.
+    pub fn stats(&self) -> &B {
+        &self.backend
+    }
+
+    /// Initializes the cgroup backend.
+    pub fn initialize(&self, pid: i32, controllers: &HashSet<Controller>) -> Result<()> {
[…]
