### 2026-06-01 da94c34
chore: add CUDA C mandelbrot workload and removed OpenGL python one
