+++ examples/programs/mandelbrot_nvidia.cu
@@ -1,0 +1,175 @@
+/*
+  Source: https://github.com/canonizer/mandelbrot-dyn
+
+  Compile: nvcc -O3 -Xcompiler -fopenmp mandelbrot.cu -o mandelbrot
+  Usage: joule-profiler --gpu profile -- ./mandelbrot {n}
+*/
+
+/*
+  The MIT License (MIT)
+
+  Copyright (c) 2014 Andrew V. Adinetz
+
+  Permission is hereby granted, free of charge, to any person obtaining a copy
+  of this software and associated documentation files (the "Software"), to deal
+  in the Software without restriction, including without limitation the rights
+  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  copies of the Software, and to permit persons to whom the Software is
+  furnished to do so, subject to the following conditions:
+
+  The above copyright notice and this permission notice shall be included in all
+  copies or substantial portions of the Software.
+
+  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+  SOFTWARE.
+*/
+
+/** @file histo-global.cu histogram with global memory atomics */
+
+#include <omp.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
+
+/** CUDA check macro */
+#define cucheck(call)                                                          \
+  {                                                                            \
+    cudaError_t res = (call);                                                  \
+    if (res != cudaSuccess) {                                                  \
+      const char *err_str = cudaGetErrorString(res);                           \
+      fprintf(stderr, "%s (%d): %s in %s", __FILE__, __LINE__, err_str,        \
+              #call);                                                          \
+      exit(-1);                                                                \
+    }                                                                          \
+  }
+
+/** time spent in device */
+double gpu_time = 0;
+
+/** a useful function to compute the number of threads */
+int divup(int x, int y) { return x / y + (x % y ? 1 : 0); }
+
+/** a simple complex type */
+struct complex {
+  __host__ __device__ complex(float re, float im = 0) {
+    this->re = re;
+    this->im = im;
+  }
+  /** real and imaginary part */
+  float re, im;
+}; // struct complex
+
+// operator overloads for complex numbers
+inline __host__ __device__ complex operator+(const complex &a,
+                                             const complex &b) {
+  return complex(a.re + b.re, a.im + b.im);
+}
+inline __host__ __device__ complex operator-(const complex &a) {
+  return complex(-a.re, -a.im);
+}
+inline __host__ __device__ complex operator-(const complex &a,
+                                             const complex &b) {
+  return complex(a.re - b.re, a.im - b.im);
+}
+inline __host__ __device__ complex operator*(const complex &a,
[…]
