### 2026-06-01 fc2add8
chore: update mandelbrot workload name in example usage
