### 2026-06-01 667f47b
refactor: constrained root and child cgroup types for safety
