+++ sources/cgroup/src/lib.rs
@@ -19,7 +19,7 @@ use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
-use crate::cgroup::{Cgroup, CgroupBackend, SysFsBackend};
+use crate::cgroup::{CgroupBackend, ChildCgroup, RootCgroup, SysFsBackend};
 use crate::counters::{Counters, CpuCounters, IoCounters, MemoryCounters};
 use crate::error::CgroupError;
 
@@ -62,8 +62,8 @@ where
 {
     config: CgroupConfig,
     handle: Option<WorkerHandle>,
-    pub proc_cgroup: Arc<Cgroup<B>>,
-    pub root_cgroup: Arc<Cgroup<B>>,
+    pub proc_cgroup: Arc<ChildCgroup<B>>,
+    pub root_cgroup: Arc<RootCgroup<B>>,
     proc_memory_counters: Arc<Mutex<MemoryCounters>>,
     global_memory_counters: Arc<Mutex<MemoryCounters>>,
     proc_cpu_counters: CpuCounters,
@@ -80,9 +80,9 @@ impl CgroupSource {
     /// Initializes internal cgroup handles but does initialize and attach any PID yet.
     pub fn new(config: CgroupConfig) -> Result<Self> {
         let root_cgroup = if let Some(cgroup_root) = &config.cgroup_root {
-            Cgroup::at(cgroup_root.clone())
+            RootCgroup::at(cgroup_root.clone())
         } else {
-            Cgroup::root()
+            RootCgroup::default()
         };
         let proc_cgroup = root_cgroup.child(&config.cgroup_name);
 
@@ -109,8 +109,8 @@ impl<B: CgroupBackend> CgroupSource<B> {
     /// This worker updates process and global memory counters at a fixed interval.
     /// It can be cancelled using the returned `CancellationToken`.
     pub fn create_worker(
-        root_cgroup: Arc<Cgroup<B>>,
-        proc_cgroup: Arc<Cgroup<B>>,
+        proc_cgroup: Arc<ChildCgroup<B>>,
+        root_cgroup: Arc<RootCgroup<B>>,
         proc_memory_counters: Arc<Mutex<MemoryCounters>>,
         global_memory_counters: Arc<Mutex<MemoryCounters>>,
         poll_interval: Duration,
@@ -128,12 +128,12 @@ impl<B: CgroupBackend> CgroupSource<B> {
                     _ = ticker.next() => {
                         trace!("Polled cgroup source.");
                         {
-                            let snapshot = proc_cgroup.stats().memory()?;
+                            let snapshot = proc_cgroup.memory()?;
                             let mut lock = proc_memory_counters.lock().await;
                             lock.update(&snapshot);
                         }
                         {
-                            let snapshot = root_cgroup.stats().memory()?;
+                            let snapshot = root_cgroup.memory()?;
                             let mut lock = global_memory_counters.lock().await;
                             lock.update(&snapshot);
                         }
@@ -168,8 +168,8 @@ impl<B: CgroupBackend> MetricReader for CgroupSource<B> {
 
         if let Some(poll_interval) = self.config.poll_interval {
             self.handle = Some(Self::create_worker(
-                self.root_cgroup.clone(),
                 self.proc_cgroup.clone(),
+                self.root_cgroup.clone(),
                 self.proc_memory_counters.clone(),
                 self.global_memory_counters.clone(),
                 poll_interval,
@@ -190,24 +190,20 @@ impl<B: CgroupBackend> MetricReader for CgroupSource<B> {
 
         self.end_timestamp = get_timestamp_micros();
         {
[…]
