### 2026-06-01 b604066
refactor: constrained root and child cgroup types for safety
