+++ sources/cgroup/src/cgroup.rs
@@ -6,7 +6,12 @@
 //! - process attachment to cgroups;
 //! - CPU, memory, and I/O statistics collection.
 
-use std::{collections::HashSet, fmt::Display, fs, path::PathBuf};
+use std::{
+    collections::HashSet,
+    fmt::Display,
+    fs,
+    path::{Path, PathBuf},
+};
 
 use log::{debug, warn};
 
@@ -43,117 +48,40 @@ impl Display for Controller {
 /// Interface for reading cgroup statistics.
 pub trait CgroupBackend: Send + Sync + 'static {
     /// Initializes the backend.
-    fn initialize(&self, pid: i32, controllers: &HashSet<Controller>) -> Result<()>;
-
-    /// Returns memory statistics.
-    fn memory(&self) -> Result<MemorySnapshot>;
-
-    /// Returns CPU statistics.
-    fn cpu(&self) -> Result<CpuSnapshot>;
-
-    /// Returns I/O statistics.
-    fn io(&self) -> Result<IoSnapshot>;
+    fn initialize(
+        &self,
+        path: &Path,
+        root: &Path,
+        pid: i32,
+        controllers: &HashSet<Controller>,
+    ) -> Result<()>;
 
     /// Cleanup the backend.
-    fn cleanup(&self) -> Result<()>;
+    fn cleanup(&self, path: &Path, root: &Path) -> Result<()>;
+
+    /// Returns memory statistics.
+    fn memory(&self, path: &Path) -> Result<MemorySnapshot>;
+
+    /// Returns CPU statistics.
+    fn cpu(&self, path: &Path) -> Result<CpuSnapshot>;
+
+    /// Returns I/O statistics.
+    fn io(&self, path: &Path) -> Result<IoSnapshot>;
 }
 
-/// A cgroup node. If `parent` is `None`, this is the root cgroup.
-pub struct Cgroup<B: CgroupBackend = SysFsBackend> {
-    pub path: PathBuf,
-    parent: Option<PathBuf>,
-    backend: B,
-}
-
-impl Cgroup {
-    /// Gets the root cgroup in sys fs `/sys/fs/cgroup`.
-    pub fn root() -> Self {
-        let path = PathBuf::from("/sys/fs/cgroup");
-        Self::at(path)
-    }
-
-    /// Builds a cgroup handle based on the provided directory.
-    pub fn at(path: PathBuf) -> Self {
-        let backend = SysFsBackend {
-            path: path.clone(),
-            root: path.clone(),
-        };
-        Self {
-            path,
-            parent: None,
-            backend,
-        }
-    }
-
-    /// Gets a child handle of the current cgroup based on its name.
-    pub fn child(&self, name: &str) -> Cgroup {
-        let child_path = self.path.join(name);
-        Cgroup {
-            parent: Some(self.path.clone()),
-            backend: SysFsBackend {
-                root: self.path.clone(),
-                path: child_path.clone(),
-            },
-            path: child_path,
-        }
-    }
-}
-
-impl<B: CgroupBackend> Cgroup<B> {
-    pub fn new(path: PathBuf, parent: Option<PathBuf>, backend: B) -> Self {
-        Self {
-            path,
-            parent,
-            backend,
-        }
-    }
-
-    /// True if the cgroup is the root cgroup, else false.
-    pub fn is_root(&self) -> bool {
-        self.parent.is_none()
-    }
-
-    /// Return the backend for stats querying.
-    pub fn stats(&self) -> &B {
-        &self.backend
-    }
-
-    /// Initializes the cgroup backend.
-    pub fn initialize(&self, pid: i32, controllers: &HashSet<Controller>) -> Result<()> {
[…]
