### 2026-06-02 72d5734
feat: amdsmi source implementation
