### 2026-06-02 0800e9c
feat: amdsmi source implementation
