+++ sources/amdsmi/src/lib.rs
@@ -1,0 +1,299 @@
+use std::{collections::HashMap, sync::Arc, time::Duration};
+
+use bitflags::bitflags;
+use futures::StreamExt;
+use joule_profiler_core::{
+    sensor::Sensors,
+    source::MetricReader,
+    types::{Metric, Metrics},
+    unit::{MetricUnit, Unit, UnitPrefix},
+};
+use log::{debug, trace};
+use tokio::{sync::Mutex, task::JoinHandle};
+use tokio_timerfd::Interval;
+use tokio_util::sync::CancellationToken;
+
+use crate::{
+    config::AmdSmiConfig,
+    counters::{Counter, EnergyCounter, PowerCounter, VramCounter},
+    error::AmdSmiError::{self},
+    hardware::{AmdSmi, Hardware},
+};
+
+pub mod config;
+pub mod counters;
+pub mod error;
+mod hardware;
+
+pub type UUID = String;
+
+type Result<T> = std::result::Result<T, AmdSmiError>;
+pub(crate) type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
+
+bitflags! {
+    #[derive(Debug, Clone, Copy)]
+    struct ProcessorSupport: u8 {
+        const Energy = 1;
+        const Power = 1 << 1;
+        const Vram = 1 << 2;
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct Processor {
+    uuid: UUID,
+    support: ProcessorSupport,
+}
+
+pub struct AmdSmiSource<H: Hardware> {
+    config: AmdSmiConfig,
+    hardware: Arc<H>,
+    handle: Option<WorkerHandle>,
+    processors: Arc<HashMap<usize, Processor>>,
+    energy_counters: HashMap<usize, EnergyCounter>,
+    vram_counters: Arc<Mutex<HashMap<usize, VramCounter>>>,
+    power_counters: Arc<Mutex<HashMap<usize, PowerCounter>>>,
+}
+
+impl AmdSmiSource<AmdSmi> {
+    pub fn new(config: AmdSmiConfig) -> Result<Self> {
+        let amdsmi = AmdSmi::new()?;
+        let processors = amdsmi.get_processors()?.into_iter().enumerate().collect();
+
+        Ok(Self {
+            config,
+            hardware: Arc::new(amdsmi),
+            handle: None,
+            processors: Arc::new(processors),
+            energy_counters: HashMap::new(),
+            vram_counters: Arc::default(),
+            power_counters: Arc::default(),
+        })
+    }
+}
+
+impl<H: Hardware> AmdSmiSource<H> {
+    pub fn create_worker(
+        hardware: Arc<H>,
+        processors: Arc<HashMap<usize, Processor>>,
+        power_counters: Arc<Mutex<HashMap<usize, PowerCounter>>>,
+        vram_counters: Arc<Mutex<HashMap<usize, VramCounter>>>,
+        poll_interval: Duration,
+    ) -> Result<WorkerHandle> {
+        let mut ticker = Interval::new_interval(poll_interval)?;
+
+        let cancellation_token = CancellationToken::new();
+        let cancellation_token_clone = cancellation_token.clone();
+
+        let handle = tokio::spawn(async move {
+            debug!("Starting AMD SMI source polling.");
+
+            loop {
+                tokio::select! {
+                    _ = ticker.next() => {
+                        trace!("Polled AMD SMI source.");
+                        Self::read_polled_counters(&hardware, &processors, &power_counters, &vram_counters).await?;
+                    }
+
+                    () = cancellation_token.cancelled() => {
+                        debug!("AMD SMI worker stopped.");
+                        break;
+                    }
+                }
+            }
+
[…]
