+++ cli/Cargo.toml
@@ -19,6 +19,7 @@ joule-profiler-source-nvml = { path = "../sources/nvml", version = "1.0.2" }
 joule-profiler-source-perf_event = { path = "../sources/perf_event", version = "1.0.2" }
 joule-profiler-source-procfs = { path = "../sources/procfs", version = "0.1.0" }
 joule-profiler-source-cgroup = { path = "../sources/cgroup", version = "0.1.0" }
+joule-profiler-source-amdsmi = { path = "../sources/amdsmi", version = "0.1.0" }
 
 joule-profiler-core.workspace = true
 log.workspace = true
