### 2026-06-02 a49f324
feat: amdsmi source implementation
