+++ cli/src/bin/main.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use anyhow::Result;
 use joule_profiler_cli::{
     CliArgs, ProfilerCommand, RaplBackend, Source, init_logging, output_format_to_displayer,
@@ -6,6 +8,8 @@ use joule_profiler_cli::{
 use joule_profiler_core::JouleProfiler;
 use joule_profiler_core::config::{Command, Config};
 use joule_profiler_source_cgroup::{CgroupConfig, CgroupSource};
+use joule_profiler_source_amdsmi::AmdSmiSource;
+use joule_profiler_source_amdsmi::config::AmdSmiConfig;
 use joule_profiler_source_nvml::Nvml;
 use joule_profiler_source_nvml::config::NvmlConfig;
 use joule_profiler_source_perf_event::PerfEvent;
@@ -59,6 +63,16 @@ async fn main() -> Result<()> {
             Ok(nvml) => {
                 trace!("Using NVML for Nvidia GPU profiling");
                 profiler.add_source(nvml);
+            }
+            Err(err) => warn!("{err}"),
+        }
+    }
+
+    if cli.sources.contains(&Source::AmdSmi) {
+        match AmdSmiSource::new(AmdSmiConfig { poll_interval: Some(Duration::from_millis(1)) }) {
+            Ok(amdsmi) => {
+                trace!("Using AMD SMI for AMD GPU profiling");
+                profiler.add_source(amdsmi);
             }
             Err(err) => warn!("{err}"),
         }
