### 2026-06-02 1e950b4
feat: amdsmi source implementation
