+++ cli/src/lib.rs
@@ -120,6 +120,8 @@ impl TryFrom<CliArgs> for Config {
 pub enum Source {
     Rapl,
     Nvml,
+    #[value(alias = "amdsmi")]
+    AmdSmi,
     #[value(alias = "perf_event")]
     Perf,
     Procfs,
@@ -131,6 +133,7 @@ impl std::fmt::Display for Source {
         let s = match self {
             Source::Rapl => "rapl",
             Source::Nvml => "nvml",
+            Self::AmdSmi => "amdsmi",
             Source::Perf => "perf | perf_event",
             Source::Procfs => "procfs",
             Source::Cgroup => "cgroup",
