### 2026-06-02 1098677
chore(amdsmi): ignore not supported devices at the initialization
