### 2026-06-03 72dc2b0
refactor(amdsmi): source refactoring, lock mutex once for counter update
