### 2026-06-03 b08ba65
refactor(amdsmi): source refactoring, lock mutex once for counter update
