+++ sources/amdsmi/src/lib.rs
@@ -3,7 +3,7 @@ use std::{collections::HashMap, sync::Arc, time::Duration};
 use bitflags::bitflags;
 use futures::StreamExt;
 use joule_profiler_core::{
-    sensor::Sensors,
+    sensor::{Sensor, Sensors},
     source::MetricReader,
     types::{Metric, Metrics},
     unit::{MetricUnit, Unit, UnitPrefix},
@@ -29,6 +29,16 @@ pub type UUID = String;
 
 type Result<T> = std::result::Result<T, AmdSmiError>;
 pub(crate) type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
+
+const MICRO_JOULE_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::Micro,
+    unit: Unit::Joule,
+};
+
+const BYTE_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Byte,
+};
 
 bitflags! {
     #[derive(Debug, Clone, Copy)]
@@ -114,19 +124,32 @@ impl<H: Hardware> AmdSmiSource<H> {
         power_counters: &Arc<Mutex<HashMap<usize, PowerCounter>>>,
         vram_counters: &Arc<Mutex<HashMap<usize, VramCounter>>>,
     ) -> Result<()> {
+        let mut vram_updates = Vec::new();
+        let mut power_updates = Vec::new();
+
         for (index, processor) in processors.iter() {
             if processor.support.contains(ProcessorSupport::Vram) {
-                let mut lock = vram_counters.lock().await;
-                let vram_usage = hardware.get_vram_usage(processor)?;
-                lock.entry(*index).and_modify(|c| c.update(vram_usage));
+                vram_updates.push((*index, hardware.get_vram_usage(processor)?));
             }
-
-            if processor.support.contains(ProcessorSupport::Vram) {
-                let mut lock = power_counters.lock().await;
-                let power = hardware.get_power(processor)?;
-                lock.entry(*index).and_modify(|c| c.update(power));
+            if processor.support.contains(ProcessorSupport::Power) {
+                power_updates.push((*index, hardware.get_power(processor)?));
             }
         }
+
+        {
+            let mut lock = vram_counters.lock().await;
+            for (index, vram_usage) in vram_updates {
+                lock.entry(index).and_modify(|c| c.update(vram_usage));
+            }
+        }
+
+        {
+            let mut lock = power_counters.lock().await;
+            for (index, power) in power_updates {
+                lock.entry(index).and_modify(|c| c.update(power));
+            }
+        }
+
         Ok(())
     }
 }
@@ -144,19 +167,15 @@ impl<H: Hardware> MetricReader for AmdSmiSource<H> {
                     .entry(*index)
                     .or_default()
                     .update(energy);
-            } else if processor.support.contains(ProcessorSupport::Power) {
-                let mut lock = self.power_counters.lock().await;
-                let power = self.hardware.get_power(processor)?;
-                lock.entry(*index).or_default().update(power);
-            }
-
-            if processor.support.contains(ProcessorSupport::Vram) {
-                let mut lock = self.vram_counters.lock().await;
-                let vram_usage = self.hardware.get_vram_usage(processor)?;
-                lock.entry(*index).or_default().update(vram_usage);
             }
         }
-
+        Self::read_polled_counters(
+            &self.hardware,
+            &self.processors,
+            &self.power_counters,
+            &self.vram_counters,
+        )
+        .await?;
         Ok(())
     }
 
@@ -222,7 +241,40 @@ impl<H: Hardware> MetricReader for AmdSmiSource<H> {
     }
 
[…]
