### 2026-06-03 8b553b2
feat(amdsmi): add gpu specification for device filtering
