### 2026-06-03 5675a9c
feat(amdsmi): add gpu specification for device filtering
