### 2026-06-04 b28aec9
chore: update amdsmi config with default in cli
