### 2026-06-04 fc2a200
feat(amdsmi): gpu devices filtering with spec implementation
