### 2026-06-04 6b56b25
feat(amdsmi): gpu devices filtering with spec implementation
