+++ sources/amdsmi/src/config.rs
@@ -1,12 +1,13 @@
-use std::time::Duration;
+use std::{collections::HashSet, time::Duration};
 
 use crate::UUID;
 
+/// Configuration of the AMD SMI source.
 #[derive(Debug, Default)]
 pub struct AmdSmiConfig {
     /// Optional background polling interval.
     pub poll_interval: Option<Duration>,
 
     /// Optional gpus filter.
-    pub gpus_spec: Option<Vec<UUID>>,
+    pub gpus_spec: Option<HashSet<UUID>>,
 }
