### 2026-06-04 2bc9cf2
doc(amdsmi): AMD SMI source documentation
