+++ sources/amdsmi/src/error.rs
@@ -29,6 +29,7 @@ pub enum AmdSmiError {
         JoinError,
     ),
 
+    /// AMD SMI library error.
     #[error("AMD SMI error: {0}")]
     AmdSmiError(#[from] amdsmi::error::AmdSmiError),
 }
