### 2026-06-04 86564b7
doc(amdsmi): AMD SMI source documentation
