### 2026-06-04 00c73b6
doc(amdsmi): AMD SMI source documentation
