### 2026-06-04 e3ec25a
chore(amdsmi): fix clippy warnings
