### 2026-06-04 ac9aebd
feat(nvml): added power, vram and GPU usage
