+++ sources/nvml/src/hardware.rs
@@ -1,49 +1,116 @@
-use std::collections::HashMap;
+use std::collections::HashSet;
 
-use joule_profiler_core::sensor::{Sensor, Sensors};
+use joule_profiler_core::time::get_timestamp_micros;
+use log::{debug, trace};
 
-use crate::{MILLI_JOULE_UNIT, NVML_SOURCE_NAME, Result, snapshot::NvmlSnapshot};
+use crate::{Device, DeviceSupport, Result, UUID, counters::PowerMeasurement, error::NvmlError};
 
 /// Trait abstracting NVML hardware access for testability.
 #[cfg_attr(test, mockall::automock)]
-pub trait NvmlHardware: Send {
-    fn read_snapshot(&self) -> Result<NvmlSnapshot>;
-    fn get_sensors(&self) -> Result<Sensors>;
+pub trait NvmlHardware: Send + Sync + 'static {
+    fn init_devices(&mut self, spec: Option<&HashSet<UUID>>) -> Result<Vec<Device>>;
+    fn get_energy(&self, device: &Device) -> Result<u64>;
+    fn get_power(&self, device: &Device) -> Result<PowerMeasurement>;
+    fn get_vram_usage(&self, device: &Device) -> Result<u64>;
+    fn get_utilization(&self, device: &Device) -> Result<u32>;
 }
 
 /// Hardware adapter for NVML library.
 pub struct NvmlWrapperHardware {
     /// The NVML wrapper instance for interacting with the NVIDIA driver.
     pub nvml: nvml_wrapper::Nvml,
+}
 
-    /// The total number of GPU devices detected.
-    pub devices_max_index: u32,
+impl NvmlWrapperHardware {
+    pub fn new() -> Result<Self> {
+        debug!("Attempting to initialize NVML reader");
+        let nvml = nvml_wrapper::Nvml::init().map_err(|err| match err {
+            nvml_wrapper::error::NvmlError::DriverNotLoaded => NvmlError::NoDriverLoaded,
+            nvml_wrapper::error::NvmlError::NoPermission => NvmlError::NoPermission,
+            _ => err.into(),
+        })?;
+
+        Ok(Self { nvml })
+    }
 }
 
 impl NvmlHardware for NvmlWrapperHardware {
-    /// Reads the current energy consumption snapshot for all GPU devices.
-    ///
-    /// This queries each GPU device and retrieves its total energy consumption counter
-    /// value in millijoules.
-    fn read_snapshot(&self) -> Result<NvmlSnapshot> {
-        let mut gpus_energy = HashMap::with_capacity(self.devices_max_index as usize);
-        for i in 0..self.devices_max_index {
-            let device = self.nvml.device_by_index(i)?;
-            let energy = device.total_energy_consumption()?;
-            gpus_energy.insert(i, energy);
-        }
-        Ok(NvmlSnapshot { gpus_energy })
+    fn init_devices(&mut self, spec: Option<&HashSet<UUID>>) -> Result<Vec<Device>> {
+        let device_count = self.nvml.device_count()?;
+
+        let devices: Vec<_> = (0..device_count)
+            .flat_map(|i| {
+                let device = self.nvml.device_by_index(i)?;
+                let uuid = device.uuid()?;
+                trace!("Discovered GPU device {uuid}.");
+
+                if let Some(spec) = &spec
+                    && !spec.contains(&uuid)
+                {
+                    trace!("Ignoring device {uuid}.");
+                    return Ok::<Option<Device>, NvmlError>(None);
+                }
+
+                let mut support = DeviceSupport::empty();
+
+                if device.total_energy_consumption().is_ok() {
+                    support |= DeviceSupport::Energy;
+                } else if device.power_usage().is_ok() {
[…]
