### 2026-06-04 863193e
feat(nvml): added power, vram and GPU usage
