+++ sources/nvml/src/lib.rs
@@ -4,34 +4,81 @@
 //! It implements the `MetricReader` trait to collect energy metrics from GPU devices and
 //! track energy usage over time.
 
-use std::collections::HashMap;
+use std::{collections::HashMap, sync::Arc, time::Duration};
 
+use bitflags::bitflags;
+use futures::StreamExt;
 use joule_profiler_core::{
-    sensor::Sensors,
+    sensor::{Sensor, Sensors},
     source::MetricReader,
     types::{Metric, Metrics},
     unit::{MetricUnit, Unit, UnitPrefix},
 };
 use log::{debug, trace};
+use tokio::{sync::Mutex, task::JoinHandle};
+use tokio_timerfd::Interval;
+use tokio_util::sync::CancellationToken;
 
 use crate::{
+    config::NvmlConfig,
+    counters::{Counter, EnergyCounter, PowerCounter, UtilizationCounter, VramCounter},
     error::NvmlError,
     hardware::{NvmlHardware, NvmlWrapperHardware},
-    snapshot::{NvmlSnapshot, Phase},
 };
 
+pub mod config;
+pub mod counters;
 mod error;
 mod hardware;
-mod snapshot;
+
+#[allow(clippy::upper_case_acronyms)]
+type UUID = String;
 
 const NVML_SOURCE_NAME: &str = "NVML";
+
 const MILLI_JOULE_UNIT: MetricUnit = MetricUnit {
     prefix: UnitPrefix::Milli,
     unit: Unit::Joule,
 };
 
+const BYTE_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Byte,
+};
+
+const PERCENT_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Percent,
+};
+
+bitflags! {
+    /// The supports of a device.
+    #[derive(Debug, Clone, Copy)]
+    struct DeviceSupport: u8 {
+        const Energy = 1;
+        const Power = 1 << 1;
+        const Vram = 1 << 2;
+        const Utilization = 1 << 3;
+    }
+}
+
 /// Custom result type for NVML.
 type Result<T> = std::result::Result<T, NvmlError>;
+
+/// Polling task handle and its cancellation token.
+type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
+
+#[derive(Debug, Clone)]
+pub struct Device {
+    /// The index of the device.
+    index: u32,
+
+    /// The UUID of the device.
+    uuid: UUID,
+
+    /// The supports of the device (e.g., Energy, Power, VRAM).
+    support: DeviceSupport,
+}
 
 /// NVML-based energy profiler for NVIDIA GPUs.
 ///
@@ -40,14 +87,28 @@ type Result<T> = std::result::Result<T, NvmlError>;
 /// The NVML hardware can be changed for testing purposes, but the default adapter is the NVML one.
 #[allow(private_interfaces, private_bounds)]
 pub struct Nvml<H: NvmlHardware = NvmlWrapperHardware> {
+    /// Source configuration.
+    config: NvmlConfig,
+
     /// The hardware instance for interacting with the NVIDIA driver.
-    hardware: H,
+    hardware: Arc<H>,
 
-    /// Accumulated energy consumption since last retrieval.
-    begin_snapshot: Option<NvmlSnapshot>,
+    /// Map of GPU devices.
+    devices: Arc<Vec<Device>>,
 
-    /// The most recent snapshot taken, used to compute deltas.
-    last_snapshot: Option<NvmlSnapshot>,
+    /// The handle to the polling task and its cancellation token.
+    handle: Option<WorkerHandle>,
+
[…]
