### 2026-06-05 203d9c2
feat(amdsmi): added GPU utilization metrics
