+++ sources/amdsmi/src/lib.rs
@@ -15,7 +15,7 @@ use tokio_util::sync::CancellationToken;
 
 use crate::{
     config::AmdSmiConfig,
-    counters::{Counter, EnergyCounter, PowerCounter, VramCounter},
+    counters::{Counter, EnergyCounter, PowerCounter, UtilizationCounter, VramCounter},
     error::AmdSmiError::{self},
     hardware::{AmdSmi, Hardware},
 };
@@ -42,6 +42,11 @@ const BYTE_UNIT: MetricUnit = MetricUnit {
     unit: Unit::Byte,
 };
 
+const PERCENT_UNIT: MetricUnit = MetricUnit {
+    prefix: UnitPrefix::None,
+    unit: Unit::Percent,
+};
+
 bitflags! {
     /// The supports of a device.
     #[derive(Debug, Clone, Copy)]
@@ -49,6 +54,7 @@ bitflags! {
         const Energy = 1;
         const Power = 1 << 1;
         const Vram = 1 << 2;
+        const Utilization = 1 << 3;
     }
 }
 
@@ -82,6 +88,9 @@ pub struct AmdSmiSource<H: Hardware> {
 
     /// The current power counters.
     power_counters: Arc<Mutex<HashMap<usize, PowerCounter>>>,
+
+    /// The current GPU utilization counters.
+    utilization_counters: Arc<Mutex<HashMap<usize, UtilizationCounter>>>,
 }
 
 impl AmdSmiSource<AmdSmi> {
@@ -102,6 +111,7 @@ impl AmdSmiSource<AmdSmi> {
             energy_counters: HashMap::new(),
             vram_counters: Arc::default(),
             power_counters: Arc::default(),
+            utilization_counters: Arc::default(),
         })
     }
 }
@@ -113,6 +123,7 @@ impl<H: Hardware> AmdSmiSource<H> {
         processors: Arc<HashMap<usize, Processor>>,
         power_counters: Arc<Mutex<HashMap<usize, PowerCounter>>>,
         vram_counters: Arc<Mutex<HashMap<usize, VramCounter>>>,
+        usage_counters: Arc<Mutex<HashMap<usize, UtilizationCounter>>>,
         poll_interval: Duration,
     ) -> Result<WorkerHandle> {
         let mut ticker = Interval::new_interval(poll_interval)?;
@@ -127,7 +138,7 @@ impl<H: Hardware> AmdSmiSource<H> {
                 tokio::select! {
                     _ = ticker.next() => {
                         trace!("Polled AMD SMI source.");
-                        Self::read_polled_counters(&hardware, &processors, &power_counters, &vram_counters).await?;
+                        Self::read_polled_counters(&hardware, &processors, &power_counters, &vram_counters, &usage_counters).await?;
                     }
 
                     () = cancellation_token.cancelled() => {
@@ -149,9 +160,11 @@ impl<H: Hardware> AmdSmiSource<H> {
         processors: &Arc<HashMap<usize, Processor>>,
         power_counters: &Arc<Mutex<HashMap<usize, PowerCounter>>>,
         vram_counters: &Arc<Mutex<HashMap<usize, VramCounter>>>,
+        utilization_counters: &Arc<Mutex<HashMap<usize, UtilizationCounter>>>,
     ) -> Result<()> {
         let mut vram_updates = Vec::new();
         let mut power_updates = Vec::new();
+        let mut utilization_updates = Vec::new();
 
         for (index, processor) in processors.iter() {
             if processor.support.contains(ProcessorSupport::Vram) {
@@ -160,19 +173,29 @@ impl<H: Hardware> AmdSmiSource<H> {
             if processor.support.contains(ProcessorSupport::Power) {
                 power_updates.push((*index, hardware.get_power(processor)?));
             }
[…]
