### 2026-06-05 70fe133
feat(amdsmi): added GPU utilization metrics
