### 2026-06-05 751d766
feat(procfs): child processes detection put in a separate loop (less polling than measure), implemented a Backend trait for testing
