### 2026-06-05 82303cf
feat(procfs): child processes detection put in a separate loop (less polling than measure), implemented a Backend trait for testing
