+++ sources/procfs/src/lib.rs
@@ -11,21 +11,22 @@ use joule_profiler_core::{
     unit::{MetricUnit, Unit, UnitPrefix},
 };
 use log::{debug, trace};
-use procfs::{Current, FromRead, Meminfo};
-use std::{sync::Arc, time::Duration};
+use std::{collections::HashSet, sync::Arc, time::Duration};
 use tokio::sync::Mutex;
 use tokio::task::JoinHandle;
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
 use crate::{
+    backend::{Backend, ProcfsBackend},
     config::ProcfsConfig,
     counters::{Counters, compute_mem_used},
     error::ProcfsError,
-    snapshot::{ProcSnapshot, measure_global, measure_proc},
+    snapshot::ProcSnapshot,
     utils::make_conversion,
 };
 
+mod backend;
 pub mod config;
 pub mod counters;
 pub mod error;
@@ -59,10 +60,13 @@ type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
 ///
 /// let source = Procfs::new(ProcfsConfig::default()).unwrap();
 /// ```
-#[derive(Debug, Default)]
-pub struct Procfs {
+#[derive(Debug)]
+pub struct Procfs<B: Backend = ProcfsBackend> {
     /// procfs source configuration.
     config: ProcfsConfig,
+
+    /// The backend used to interract with procfs. Used for testing.
+    backend: Arc<B>,
 
     /// pid of the profiled process, initialized at -1.
     pid: i32,
@@ -71,10 +75,16 @@ pub struct Procfs {
     counters: Arc<Mutex<Counters>>,
 
     /// The handle to the polling task and it's cancellation token.
-    handle: Option<WorkerHandle>,
+    polling_task_handle: Option<WorkerHandle>,
+
+    /// The handle to the polling task and it's cancellation token.
+    process_discovery_task_handle: Option<WorkerHandle>,
 
     /// Total physical memory in bytes, read once at construction from `/proc/meminfo`.
     mem_total: u64,
+
+    /// The current processes to be measured, updated at each poll from the process discovery task.
+    detected_processes: Arc<Mutex<HashSet<i32>>>,
 }
 
 impl Procfs {
@@ -82,18 +92,26 @@ impl Procfs {
     ///
     /// Reads `MemTotal` from `/proc/meminfo` at construction time.
     pub fn new(config: ProcfsConfig) -> Result<Self> {
-        let mem_total = Meminfo::from_file(Meminfo::PATH)?.mem_total;
+        let backend = ProcfsBackend;
+        let mem_total = backend.mem_total()?;
         Ok(Self {
             config,
             pid: -1,
             mem_total,
-            ..Default::default()
+            backend: Arc::new(backend),
+            counters: Arc::default(),
+            detected_processes: Arc::default(),
+            polling_task_handle: None,
+            process_discovery_task_handle: None,
         })
     }
+}
 
-    fn spawn_worker(
-        pid: i32,
+impl<B: Backend> Procfs<B> {
+    fn spawn_polling_worker(
+        backend: Arc<B>,
         counters: Arc<Mutex<Counters>>,
+        detected_processes: Arc<Mutex<HashSet<i32>>>,
         poll_interval: Duration,
     ) -> Result<WorkerHandle> {
         let mut ticker = Interval::new_interval(poll_interval)?;
@@ -107,20 +125,10 @@ impl Procfs {
                 tokio::select! {
                     _ = ticker.next() => {
                         trace!("Polled procfs source.");
-
-                        let proc = match measure_proc(pid) {
-                            Err(ProcfsError::Procfs(procfs::ProcError::NotFound(_))) => {
-                                Ok(ProcSnapshot::default())
-                            }
[…]
