+++ sources/procfs/src/config.rs
@@ -8,6 +8,8 @@ pub struct ProcfsConfig {
     /// Polling interval. If `None`, metrics are only captured on phase boundaries.
     pub poll_interval: Option<Duration>,
 
+    pub process_detection_poll_interval: Option<Duration>,
+
     /// Memory unit for process metrics (default: `Mega`).
     pub proc_memory_unit: MemoryUnit,
 
@@ -19,6 +21,7 @@ impl Default for ProcfsConfig {
     fn default() -> Self {
         Self {
             poll_interval: None,
+            process_detection_poll_interval: None,
             proc_memory_unit: MemoryUnit::Mega,
             global_memory_unit: MemoryUnit::Giga,
         }
