### 2026-06-05 0064eff
test(procfs): procfs source tests implementation
