+++ sources/procfs/src/lib.rs
@@ -92,13 +92,11 @@ impl Procfs {
     ///
     /// Reads `MemTotal` from `/proc/meminfo` at construction time.
     pub fn new(config: ProcfsConfig) -> Result<Self> {
-        let backend = ProcfsBackend;
-        let mem_total = backend.mem_total()?;
         Ok(Self {
             config,
             pid: -1,
-            mem_total,
-            backend: Arc::new(backend),
+            mem_total: 0,
+            backend: Arc::new(ProcfsBackend),
             counters: Arc::default(),
             detected_processes: Arc::default(),
             polling_task_handle: None,
@@ -213,6 +211,7 @@ impl<B: Backend> MetricReader for Procfs<B> {
     /// Initializes the source to `pid` and starts the background poller if a `poll_interval` is configured.
     async fn init(&mut self, pid: i32) -> Result<()> {
         self.pid = pid;
+        self.mem_total = self.backend.mem_total()?;
         *self.detected_processes.lock().await = self.backend.collect_children(pid);
 
         if let Some(detection_poll_interval) = self.config.process_detection_poll_interval {
@@ -428,5 +427,276 @@ impl<B: Backend> MetricReader for Procfs<B> {
 
     fn get_name() -> &'static str {
         "procfs"
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{collections::HashSet, path::PathBuf, sync::Arc, time::Duration};
+
+    use joule_profiler_core::source::MetricReader;
+    use tokio::time::sleep;
+
+    use crate::{
+        Procfs, backend::MockBackend, config::ProcfsConfig, error::ProcfsError,
+        snapshot::GlobalSnapshot,
+    };
+
+    fn create_source(backend: MockBackend) -> Procfs<MockBackend> {
+        let source = Procfs {
+            backend: Arc::new(backend),
+            pid: 1,
+            config: ProcfsConfig::default(),
+            mem_total: 0,
+            counters: Arc::default(),
+            detected_processes: Arc::default(),
+            polling_task_handle: None,
+            process_discovery_task_handle: None,
+        };
+        source
+    }
+
+    #[tokio::test]
+    async fn test_init_function_initializes_field_correctly() {
+        let mem_total = 4096;
+        let pid = 1;
+        let children: HashSet<_> = vec![pid, 2].into_iter().collect();
+
+        let mut backend = MockBackend::new();
+
+        backend
+            .expect_collect_children()
+            .returning(move |_| vec![pid, 2].into_iter().collect());
+
+        backend
+            .expect_mem_total()
+            .once()
+            .returning(move || Ok(mem_total));
+
+        let mut source = create_source(backend);
+
+        source.init(pid).await.unwrap();
+
+        assert_eq!(source.mem_total, mem_total);
+        assert!(source.detected_processes.lock().await.eq(&children));
+        assert_eq!(source.pid, pid);
+    }
+
+    #[tokio::test]
+    async fn test_process_detection_polling_updates_processes_list() {
+        let children: HashSet<_> = vec![1, 2, 3, 4].into_iter().collect();
+
+        let mut backend = MockBackend::new();
+        backend
+            .expect_collect_children()
+            .once()
+            .returning(|_| HashSet::new());
+
+        backend
+            .expect_collect_children()
+            .returning(|_| vec![1, 2, 3, 4].into_iter().collect());
+
+        backend.expect_mem_total().once().returning(|| Ok(0));
+        let mut source = create_source(backend);
+        source.config.process_detection_poll_interval = Some(Duration::from_millis(1));
+        assert!(source.detected_processes.lock().await.eq(&HashSet::new()));
+
+        source.init(1).await.unwrap();
+
+        sleep(Duration::from_millis(5)).await;
+
[…]
