### 2026-06-05 502f946
test(procfs): procfs source tests implementation
