### 2026-06-05 1d62c0a
doc(procfs): added README.md for procfs source
