### 2026-06-05 0018e97
doc(procfs): added README.md for procfs source
