### 2026-06-07 11beb6b
doc: Partly addressing review from issue #38
Revising content to address issue #38.
