+++ joss/paper/paper.md
@@ -31,11 +31,11 @@ repository: https://github.com/joule-profiler/joule-profiler
 
 # Summary
 
-Joule Profiler is a lightweight Linux command-line tool for measuring a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and to attribute energy use to each. The tool detects phase triggers from program output and automatically queries sources like Intel RAPL (CPUs) or NVML (GPUs) to report system-wide energy consumption.
+Joule Profiler is a lightweight Linux command-line tool for measuring a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and to attribute energy use to each. The tool detects phase triggers in program output and automatically queries sources such as Intel RAPL (CPUs) or NVML (GPUs) to report system-wide energy consumption.
 
 # Statement of need
 
-Energy use in computing is a growing concern across research and industry. Software running in clouds, data centres, and edge devices contributes significantly to global energy consumption. Improving efficiency requires tools that measure energy during execution. Hardware counters provided by modern CPUs and GPUs (e.g., Intel RAPL) make software-based energy measurement possible without external devices.
+Energy use in computing is a growing concern across research and industry. Software running in clouds, data centres, and edge devices contributes significantly to global energy consumption. Improving efficiency requires tools that measure energy during execution. Hardware counters on modern CPUs and GPUs (e.g., Intel RAPL) enable software-based energy measurement without external devices.
 
 Researchers and developers need simple tools to measure the energy use of code segments without complex infrastructure. Joule Profiler addresses this with phase-based profiling that integrates easily into workflows.
 
@@ -45,7 +45,9 @@ Existing tools like PowerAPI [@powerapi], Alumet [@alumet], Scaphandre [@scaphan
 
 These solutions are suited for system-level monitoring, not fine-grained analysis of program phases. Joule Profiler, in contrast, is designed for lightweight, single-invocation use, enabling energy attribution to specific phases within programs.
 
-# Phase-based profiling
+# Software design
+
+## Phase-based profiling
 
 \begin{figure}
 	\centering
@@ -60,11 +62,11 @@ Joule Profiler scans standard output for user-defined patterns to detect phase b
 
 When a phase marker is detected, Joule Profiler records energy counter values at that boundary. After execution, it computes per-phase energy by subtracting these values.
 
-# Software design
+## Software architecture
 
-Joule Profiler’s modular design separates measurement logic from hardware specifics. It accesses energy and performance metrics using `perf_event` [@linux_perf_event] (or powercap as a fallback) for RAPL, and NVML [@nvidia_nvml] for NVIDIA GPUs. The tool can correlate energy with performance counters, supporting extension and maintenance.
+Joule Profiler’s modular design separates measurement logic from hardware specifics. It accesses energy and performance metrics using `perf_event` [@linux_perf_event] (or powercap as a fallback) for RAPL and NVML [@nvidia_nvml] for NVIDIA GPUs. The tool can correlate energy with performance counters, supporting extension and maintenance.
 
-The tool uses a layered structure: the core detects phases and aggregates metrics; sources run asynchronously for parallel data collection; the CLI manages user interaction; and hardware backends are abstracted for easy integration of new sources.
[…]
