### 2026-06-08 14836aa
fix: added CLI flag --init-timeout to be able to configure it on slow devices
