### 2026-06-08 771ce9e
fix: added CLI flag --init-timeout to be able to configure it on slow devices
