### 2026-06-08 13dd130
fix: added CLI flag --init-timeout to be able to configure it on slow devices
