+++ core/src/config.rs
@@ -21,9 +21,12 @@
 //! };
 //! ```
 
+use std::time::Duration;
+
 use derive_builder::Builder;
 
 const PHASE_TOKEN_DEFAULT_REGEX_PATTERN: &str = "__[A-Z0-9_]+__";
+const DEFAULT_INIT_TIMEOUT: Duration = Duration::from_secs(1);
 
 /// Top-level configuration for Joule Profiler.
 #[derive(Debug)]
@@ -61,4 +64,8 @@ pub struct ProfileConfig {
 
     /// Executes the profiled command with root privileges if true and Joule Profiler is launched as root.
     pub use_root: bool,
+
+    /// Duration before aborting sources initialization.
+    #[builder(default = DEFAULT_INIT_TIMEOUT)]
+    pub init_timeout: Duration,
 }
