### 2026-06-08 bbf1860
chore: fix doctests with new CLI flag
