### 2026-06-08 37b5f98
chore: fix doctests with new CLI flag
