### 2026-06-08 525d6ed
fix: skip empty sources results and throw an error only when merging different phases counts
