### 2026-06-08 98eb64e
chore(procfs): added more logging
