### 2026-06-08 233db5b
chore(procfs): added more logging
