### 2026-06-08 570b94e
feat: added default value in CLI for token-pattern and init-timeout for CLI documentation
