### 2026-06-08 b6a3d84
feat: added default value in CLI for token-pattern and init-timeout for CLI documentation
