### 2026-06-08 b5b477e
refactor: improved ProfileConfig builder usage
