### 2026-06-08 fd46f23
refactor: improved ProfileConfig builder usage
