### 2026-06-08 1cbe5e4
refactor: improved ProfileConfig builder usage
