### 2026-06-08 98f0119
chore: removed former default ProfileConfig values in core
