### 2026-06-09 5bc9032
feat(config): Config deserialization using toml crate, override with CLI using CliOverride trait
