+++ sources/perf_event/Cargo.toml
@@ -13,6 +13,7 @@ readme = "README.md"
 joule-profiler-core.workspace = true
 log.workspace = true
 thiserror.workspace = true
+serde.workspace = true
 
 perf-event2 = "0.7.4"
 
