### 2026-06-09 67052d1
feat(config): Config deserialization using toml crate, override with CLI using CliOverride trait
