### 2026-06-09 a8a9e93
feat(config): Config deserialization using toml crate, override with CLI using CliOverride trait
