+++ core/tests/integration_test.rs
@@ -25,7 +25,8 @@ mock! {
     impl MetricReader for MetricReader {
         type Type = ();
         type Error = MockError;
-
+        type Config = ();
+        fn from_config(config: ()) -> Result<Self, MockError>;
         async fn init(&mut self, pid: i32) -> Result<(), MockError>;
         async fn join(&mut self) -> Result<(), MockError>;
         async fn measure(&mut self) -> Result<(), MockError>;
@@ -33,6 +34,7 @@ mock! {
         fn get_sensors(&self) -> Result<Sensors, MockError>;
         fn to_metrics(&self, v: ()) -> Result<Metrics, MockError>;
         fn get_name() -> &'static str;
+        fn get_id() -> &'static str;
     }
 }
 
