### 2026-06-12 1365f27
feat(config): replace CLI override with function to abstract override logic
