### 2026-06-12 fc070b2
feat(config): replace CLI override with function to abstract override logic
