+++ sources/nvml/Cargo.toml
@@ -15,6 +15,7 @@ log.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 futures.workspace = true
+serde.workspace = true
 
 nvml-wrapper = "0.12.0"
 tokio-timerfd = "0.2.0"
