### 2026-06-12 d775f0e
feat(config): replace CLI override with function to abstract override logic
