+++ sources/nvml/src/lib.rs
@@ -104,32 +104,6 @@ pub struct Nvml<H: NvmlHardware = NvmlWrapperHardware> {
     power_counters: Arc<Mutex<HashMap<u32, PowerCounter>>>,
 }
 
-impl Nvml {
-    /// Creates a new NVML source instance.
-    /// This initializes the NVML hardware with the specified devices specification.
-    ///
-    /// This function will return an error if:
-    /// - The NVML library cannot be initialized (driver not installed, incompatible version, etc.)
-    /// - Device information cannot be queried.
-    /// - No GPU devices are detected.
-    /// - The permissions are insufficient to be able to query the NVML driver.
-    pub fn new(config: NvmlConfig) -> Result<Self> {
-        let mut hardware = NvmlWrapperHardware::new()?;
-        let devices = hardware.init_devices(config.gpus_spec.as_ref())?;
-
-        Ok(Self {
-            config,
-            hardware: Arc::new(hardware),
-            devices: Arc::new(devices),
-            handle: None,
-            energy_counters: HashMap::default(),
-            power_counters: Arc::default(),
-            utilization_counters: Arc::default(),
-            vram_counters: Arc::default(),
-        })
-    }
-}
-
 impl<H: NvmlHardware> Nvml<H> {
     /// Creates the worker task for power and vram polling at the specified polling interval.
     pub fn create_worker(
@@ -224,10 +198,24 @@ impl<H: NvmlHardware> MetricReader for Nvml<H> {
 
     type Error = NvmlError;
 
-    type Config = ();
+    type Config = NvmlConfig;
 
-    fn from_config(_config: Self::Config) -> Result<Self> {
-        Self::new()
+    /// Creates a new NVML source instance.
+    /// This initializes the NVML hardware with the specified devices specification.
+    fn from_config(config: NvmlConfig) -> Result<Self> {
+        let mut hardware = H::new()?;
+        let devices = hardware.init_devices(config.gpus_spec.as_ref())?;
+
+        Ok(Self {
+            config,
+            hardware: Arc::new(hardware),
+            devices: Arc::new(devices),
+            handle: None,
+            energy_counters: HashMap::default(),
+            power_counters: Arc::default(),
+            utilization_counters: Arc::default(),
+            vram_counters: Arc::default(),
+        })
     }
 
     async fn measure(&mut self) -> Result<()> {
@@ -443,11 +431,7 @@ mod tests {
     use tokio::time::sleep;
 
     use crate::{
-        Device, DeviceSupport, Nvml,
-        config::NvmlConfig,
-        counters::{Counter, PowerMeasurement},
-        error::NvmlError,
-        hardware::MockNvmlHardware,
+        Device, DeviceSupport, Nvml, config::NvmlConfig, counters::{Counter, PowerMeasurement}, error::NvmlError, hardware::{MockNvmlHardware, NvmlHardware},
     };
 
     fn make_device(index: u32, support: DeviceSupport) -> Device {
@@ -485,7 +469,7 @@ mod tests {
     async fn measure_reads_energy_for_energy_capable_device() {
         let device = make_device(0, DeviceSupport::Energy);
 
-        let mut hw = MockNvmlHardware::new();
+        let mut hw = MockNvmlHardware::new().unwrap();
         hw.expect_get_energy()
             .with(predicate::function(move |d: &Device| {
                 d.index == device.index
@@ -503,7 +487,7 @@ mod tests {
     async fn measure_reads_power_for_power_only_device() {
         let device = make_device(0, DeviceSupport::Power);
 
-        let mut hw = MockNvmlHardware::new();
[…]
