### 2026-06-12 ec15d2e
feat(config): RAPL config fix, removed CLI override
