+++ sources/rapl/Cargo.toml
@@ -15,6 +15,7 @@ log.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 futures.workspace = true
+serde.workspace = true
 
 tokio-timerfd = "0.2.0"
 perf-event2 = "0.7.4"
