### 2026-06-12 1fc2e89
feat(config): RAPL config fix, removed CLI override
