### 2026-06-12 ffcc844
feat(config): RAPL config fix, removed CLI override
