### 2026-06-12 b239f5c
fix: perf_event_paranoid negative level parsing
