### 2026-06-15 74e8614
refactor: replace TryInto by TryFrom for RaplDomainType and removed unnecessary String allocation
