### 2026-06-15 7916a6a
chore: cargo fmt
