### 2026-06-15 49fc6af
fix: check command not empty before spawning process to avoid panic
