### 2026-06-15 4adf12c
fix: check command not empty before spawning process to avoid panic
