### 2026-06-15 3d3d82b
fix: escape quoted fields and semicolons
