+++ cli/Cargo.toml
@@ -28,6 +28,7 @@ env_logger = "0.11.8"
 clap = { version = "4.5.53", features = ["derive"] }
 serde_json = { version = "1.0.149", features = ["preserve_order"] }
 humantime = "2.3.0"
+csv = "1.4.0"
 
 [dev-dependencies]
 tempfile.workspace = true
