### 2026-06-15 37af37c
fix: escape quoted fields and semicolons
