+++ cli/src/output/formats/csv.rs
@@ -1,6 +1,6 @@
 use std::fs::File;
-use std::io::Write;
 
+use csv::{QuoteStyle, Writer};
 use joule_profiler_core::fs::{
     create_file_with_user_permissions, default_results_filename, get_absolute_path,
 };
@@ -13,8 +13,8 @@ type Result<T> = std::result::Result<T, DisplayerError>;
 
 /// CSV output writer to a file.
 pub struct CsvOutput {
-    /// File handle for writing CSV data.
-    file: File,
+    /// CSV writer wrapping the output file.
+    writer: Writer<File>,
 
     /// Path to the output CSV file.
     filename: String,
@@ -28,30 +28,46 @@ impl CsvOutput {
         let absolute_path = get_absolute_path(&filename)?;
         let file = create_file_with_user_permissions(&absolute_path)?;
 
+        let writer = csv::WriterBuilder::new()
+            .delimiter(b';')
+            .quote_style(QuoteStyle::Necessary)
+            .from_writer(file);
+
         Ok(Self {
-            file,
+            writer,
             filename: absolute_path,
         })
     }
 
     /// Write CSV header row.
     fn write_header(&mut self, with_iteration_id: bool) -> Result<()> {
+        let mut record: Vec<&str> = Vec::new();
+
         if with_iteration_id {
-            write!(self.file, "iteration_id;")?;
+            record.push("iteration_id");
         }
 
-        write!(self.file, "phase_id;phase_name;phase_duration_ms;")?;
-        write!(
-            self.file,
-            "metric_name;metric_value;metric_unit;metric_source;"
-        )?;
-        write!(
-            self.file,
-            "start_token;end_token;start_token_line;end_token_line;timestamp;"
-        )?;
-        write!(self.file, "command;exit_code;token_pattern")?;
-        writeln!(self.file)?;
+        record.extend_from_slice(&[
+            "phase_id",
+            "phase_name",
+            "phase_duration_ms",
+            "metric_name",
+            "metric_value",
+            "metric_unit",
+            "metric_source",
+            "start_token",
+            "end_token",
+            "start_token_line",
+            "end_token_line",
+            "timestamp",
+            "command",
+            "exit_code",
+            "token_pattern",
+        ]);
 
+        self.writer
+            .write_record(&record)
+            .map_err(|err| DisplayerError::OutputFormatError(err.into()))?;
         Ok(())
     }
 
@@ -63,51 +79,53 @@ impl CsvOutput {
         cmd: &str,
         token_pattern: &str,
     ) -> Result<()> {
-        for metric in &phase.metrics {
-            let start_token_line = phase
-                .start_token_line
-                .map(|l| l.to_string())
-                .unwrap_or_default();
-            let end_token_line = phase
-                .end_token_line
-                .map(|l| l.to_string())
-                .unwrap_or_default();
+        let start_token_line = phase
+            .start_token_line
+            .map(|l| l.to_string())
+            .unwrap_or_default();
+        let end_token_line = phase
+            .end_token_line
+            .map(|l| l.to_string())
+            .unwrap_or_default();
 
-            write!(
-                self.file,
-                "{};\"{}\";{};",
-                phase.index,
-                phase.get_name(),
-                phase.duration_ms
-            )?;
-            write!(
-                self.file,
-                "{};{};{};{};",
-                metric.name, metric.value, metric.unit, metric.source
-            )?;
-            write!(
-                self.file,
-                "{};{};{};{};{};",
-                phase.start_token,
-                phase.end_token,
-                start_token_line,
-                end_token_line,
-                phase.timestamp
-            )?;
-            write!(
[…]
