### 2026-06-16 905bb7a
feat: added library usage example
