### 2026-06-16 d4bf583
feat: update example and put it in it's own crate, also upgrade tokio version to 1.52.3
