+++ examples/library_usage/README.md
@@ -1,0 +1,8 @@
+# Joule Profiler library usage example
+
+This crate provides an example of usage of Joule Profiler library.
+
+Run the example:
+```bash
+cargo run --bin library_usage_example
+```
