### 2026-06-16 a7347c3
feat: update example and put it in it's own crate, also upgrade tokio version to 1.52.3
