### 2026-06-16 ca46656
chore: format Cargo.toml
