### 2026-06-17 821cd5b
docs: documented orchestrator measure method to indicate that it does not ensure measurement completion
