### 2026-06-18 246da5c
feat: added init validation in orchestrator and refactored orchestrator
