### 2026-06-18 0aac78a
feat: added init validation in orchestrator and refactored orchestrator
