### 2026-06-18 5d9a163
feat: added init validation in orchestrator and refactored orchestrator
