### 2026-06-18 aea2520
test(nvml): test new version of nvml source
