### 2026-06-18 5282d5b
test(nvml): test new version of nvml source
