+++ sources/nvml/src/lib.rs
@@ -112,9 +112,8 @@ pub struct Nvml<H: NvmlHardware = NvmlWrapperHardware> {
 }
 
 impl Nvml {
-    /// Creates a new NVML profiler instance.
-    ///
-    /// This initializes the NVML library and queries all available GPU devices.
+    /// Creates a new NVML source instance.
+    /// This initializes the NVML hardware with the specified devices specification.
     ///
     /// This function will return an error if:
     /// - The NVML library cannot be initialized (driver not installed, incompatible version, etc.)
@@ -233,6 +232,7 @@ impl<H: NvmlHardware> MetricReader for Nvml<H> {
     type Error = NvmlError;
 
     async fn measure(&mut self) -> Result<()> {
+        debug!("NVML measure triggered.");
         for device in self.devices.iter() {
             if device.support.contains(DeviceSupport::Energy) {
                 let energy = self.hardware.get_energy(device)?;
@@ -255,6 +255,7 @@ impl<H: NvmlHardware> MetricReader for Nvml<H> {
 
     /// Retrieve the current counters and reset them for the next phase.
     async fn retrieve(&mut self) -> Result<Self::Type> {
+        debug!("Retrieving NVML counters.");
         let mut energy_counters = self.energy_counters.clone();
         for counter in self.energy_counters.values_mut() {
             counter.reset();
@@ -388,7 +389,6 @@ impl<H: NvmlHardware> MetricReader for Nvml<H> {
                 poll_interval,
             )?);
         }
-
         debug!("NVML source initialized.");
         Ok(())
     }
@@ -408,180 +408,260 @@ impl<H: NvmlHardware> MetricReader for Nvml<H> {
     }
 }
 
-// #[cfg(test)]
-// mod tests {
-//     use joule_profiler_core::{sensor::Sensor, types::MetricValue};
+#[cfg(test)]
+mod tests {
+    use std::{collections::HashMap, sync::Arc, time::Duration};
 
-//     use super::*;
-//     use crate::{hardware::MockNvmlHardware, snapshot::NvmlSnapshot};
+    use joule_profiler_core::source::MetricReader;
+    use mockall::predicate;
+    use tokio::time::sleep;
 
-//     fn snapshot(entries: Vec<(u32, u64)>) -> NvmlSnapshot {
-//         NvmlSnapshot {
-//             gpus_energy: entries.into_iter().collect(),
-//         }
-//     }
+    use crate::{
+        Device, DeviceSupport, Nvml,
+        config::NvmlConfig,
+        counters::{Counter, PowerMeasurement},
+        error::NvmlError,
+        hardware::MockNvmlHardware,
+    };
 
-//     fn sensors(count: u32) -> Sensors {
-//         (0..count)
-//             .map(|i| Sensor::new(format!("GPU-{i}"), MILLI_JOULE_UNIT, NVML_SOURCE_NAME))
-//             .collect()
-//     }
+    fn make_device(index: u32, support: DeviceSupport) -> Device {
+        Device {
+            index,
+            uuid: format!("GPU-mock-{index}"),
+            support,
+        }
+    }
 
-//     fn nvml_with_hardware(hardware: MockNvmlHardware) -> Nvml<MockNvmlHardware> {
-//         Nvml {
-//             hardware,
-//             begin_snapshot: None,
-//             last_snapshot: None,
-//         }
-//     }
+    fn build_nvml(
+        hardware: MockNvmlHardware,
+        devices: Vec<Device>,
+        poll_interval: Option<Duration>,
+    ) -> Nvml<MockNvmlHardware> {
+        Nvml {
+            config: NvmlConfig {
+                poll_interval,
+                gpus_spec: None,
+            },
[…]
