### 2026-06-18 3787bb9
chore(nvml): improved logging
