### 2026-06-18 1f5b9af
chore(amdsmi): rename source and error better handling
