### 2026-06-18 cbb03c8
chore(amdsmi): rename source and error better handling
