### 2026-06-18 b597bd9
chore(amdsmi): rename source and error better handling
