### 2026-06-18 23a11f2
docs(amdsmi): document AMD SMI hardware
