### 2026-06-18 3c06beb
docs(amdsmi): document AMD SMI hardware
