### 2026-06-18 1ce2ff5
docs(amdsmi): document AMD SMI hardware
