### 2026-06-18 9a49028
docs(nvml): document NVML hardware
