### 2026-06-18 d3f8112
docs(nvml): update module doc
