### 2026-06-18 102bb47
test(amdsmi): test AMD SMI source
