+++ sources/amdsmi/src/lib.rs
@@ -1,3 +1,7 @@
+//! AMD SMI (AMD System Management Interface) energy profiling integration for Joule Profiler.
+//!
+//! This module provides energy consumption, VRAM usage and GPU utilization metrics for AMD GPUs using the AMD SMI library.
+
 use std::{collections::HashMap, sync::Arc, time::Duration};
 
 use bitflags::bitflags;
@@ -428,5 +432,391 @@ impl<H: AmdSmiHardware> MetricReader for AmdSmi<H> {
 
     fn get_name() -> &'static str {
         AMDSMI_SOURCE_NAME
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::{collections::HashMap, sync::Arc, time::Duration};
+
+    use amdsmi::types::{EnergyCount, GpuUsageInfo};
+    use joule_profiler_core::source::MetricReader;
+    use mockall::predicate::function;
+    use tokio::time::sleep;
+
+    use crate::{
+        AmdSmi, Processor, ProcessorSupport,
+        config::AmdSmiConfig,
+        counters::{Counter, PowerMeasurement},
+        error::AmdSmiError,
+        hardware::MockAmdSmiHardware,
+    };
+
+    fn make_processor(uuid: &str, support: ProcessorSupport) -> Processor {
+        Processor {
+            uuid: uuid.to_owned(),
+            support,
+        }
+    }
+
+    fn processors_map(list: Vec<Processor>) -> Arc<HashMap<usize, Processor>> {
+        Arc::new(list.into_iter().enumerate().collect())
+    }
+
+    fn build_amdsmi(
+        hardware: MockAmdSmiHardware,
+        processors: Arc<HashMap<usize, Processor>>,
+        poll_interval: Option<Duration>,
+    ) -> AmdSmi<MockAmdSmiHardware> {
+        AmdSmi {
+            config: AmdSmiConfig {
+                poll_interval,
+                gpus_spec: None,
+            },
+            hardware: Arc::new(hardware),
+            handle: None,
+            processors,
+            energy_counters: HashMap::default(),
+            vram_counters: Arc::default(),
+            power_counters: Arc::default(),
+            utilization_counters: Arc::default(),
+        }
+    }
+
+    fn energy_count(value: u64) -> EnergyCount {
+        EnergyCount {
+            energy_accumulator: value,
+            counter_resolution: 1.0,
+            timestamp: 0,
+        }
+    }
+
+    fn gpu_usage(gpu_usage: u32) -> GpuUsageInfo {
+        GpuUsageInfo {
+            gpu_usage,
+            mem_usage: 0,
+        }
+    }
+
+    fn power_measurement(power: u32) -> PowerMeasurement {
+        PowerMeasurement {
+            timestamp: 0,
+            power,
+        }
+    }
+
+    fn by_uuid(uuid: &'static str) -> impl mockall::Predicate<Processor> {
+        function(move |p: &Processor| p.uuid == uuid)
+    }
+
+    #[test]
+    fn get_sensors_energy_support_emits_energy_sensor() {
+        let processors = processors_map(vec![make_processor("UUID_0", ProcessorSupport::Energy)]);
+        let amdsmi = build_amdsmi(MockAmdSmiHardware::new(), processors, None);
+
+        let sensors = amdsmi.get_sensors().unwrap();
+        assert!(sensors.iter().any(|s| s.name == "GPU-UUID_0-energy"));
+    }
+
+    #[test]
+    fn get_sensors_power_support_emits_energy_sensor() {
+        let processors = processors_map(vec![make_processor("UUID_0", ProcessorSupport::Power)]);
+        let amdsmi = build_amdsmi(MockAmdSmiHardware::new(), processors, None);
+
[…]
