### 2026-06-18 0267ef3
chore(amdsmi): replaced local amdsmi path to github repository
