### 2026-06-18 8ce14e2
docs(procfs): documented backend functions
