### 2026-06-18 19b8b68
docs(amdsmi): added source README.md
