### 2026-06-18 94d9b36
docs(nvml): update nvml source README.md
