### 2026-06-18 d7ba9ba
docs(config): documented all config functions
