+++ cli/src/config/table.rs
@@ -1,4 +1,10 @@
-use std::collections::HashSet;
+//! Configuration management for the profiler CLI.
+//!
+//! This module provides [`ConfigTable`], which bridges raw CLI arguments and
+//! global config file settings into a unified [`Config`] object consumed by
+//! the profiler core.
+
+use std::collections::{HashMap, HashSet};
 
 use anyhow::Result;
 use joule_profiler_core::{
@@ -9,19 +15,32 @@ use log::warn;
 
 use crate::{
     CliArgs, ProfilerCommand, Source,
-    config::{GlobalConfig, source::MetricSourceConfig},
-    config_to_displayer,
-    output::displayer::Displayer,
+    config::{GlobalConfig, ProfilerConfig, source::MetricSourceConfig},
+    output::formats::OutputFormat,
 };
 
+/// Holds the resolved configuration state, merging values from the global
+/// config file and CLI arguments before they are built into a final.
 pub struct ConfigTable {
-    pub config: GlobalConfig,
+    /// Global Joule Profiler configuration.
+    pub profiler_config: ProfilerConfig,
+
+    /// Raw TOML values for each named metric source, keyed by source ID.
+    pub sources_config: HashMap<String, toml::Value>,
+
+    /// The set of source IDs that are enabled.
     enabled_sources: HashSet<String>,
 }
 
 impl ConfigTable {
-    pub fn new(config: GlobalConfig, sources: &[Source]) -> Self {
-        let enabled_sources: HashSet<_> = config
+    /// Creates a new [`ConfigTable`] from a [`GlobalConfig`] and the
+    /// sources enabled from the CLI.
+    ///
+    /// The `enabled_sources` set is the union of:
+    /// - sources declared in the global config file, and
+    /// - sources passed directly via the `sources` CLI argument.
+    pub fn new(global_config: GlobalConfig, sources: &[Source]) -> Self {
+        let enabled_sources: HashSet<_> = global_config
             .sources
             .keys()
             .cloned()
@@ -29,48 +48,37 @@ impl ConfigTable {
             .collect();
 
         Self {
-            config,
+            profiler_config: global_config.profiler,
+            sources_config: global_config.sources,
             enabled_sources,
         }
     }
 
-    pub fn build_source_override<R, O>(
-        &mut self,
-        config_override: &mut O,
-        config_override_fn: impl FnOnce(&mut O, &mut R::Config),
-    ) -> Result<Option<R>>
-    where
-        R: MetricReader,
-    {
-        if !self.enabled_sources.contains(R::get_id()) {
-            return Ok(None);
+    /// Applies CLI overrides to Joule Profiler global configuration.
+    pub fn apply_cli(&mut self, cli: &mut CliArgs) {
+        if cli.csv {
+            self.profiler_config.output_format = OutputFormat::Csv
+        } else if cli.json {
+            self.profiler_config.output_format = OutputFormat::Json;
         }
 
-        let config_wrapper = match self.config.sources.remove(R::get_id()) {
-            Some(v) => v.try_into(),
-            None => Ok(MetricSourceConfig::default()),
-        }?;
+        if let Some(output_file) = cli.output_file.take() {
+            self.profiler_config.output_file = Some(output_file);
+        }
 
-        let mut config = config_wrapper.inner;
+        if let Some(rapl_backend) = cli.rapl_backend.take() {
+            self.profiler_config.rapl_backend = rapl_backend;
+        }
 
-        config_override_fn(config_override, &mut config);
-
-        match R::from_config(config) {
-            Ok(reader) => Ok(Some(reader)),
-            Err(e) => {
-                if config_wrapper.ignore_on_failure {
-                    warn!(
[…]
