### 2026-06-18 9dcc1be
docs(config): documented all config functions
