### 2026-06-18 1e632d0
docs(config): documented all config functions
