### 2026-06-18 48b3c1d
docs(config): fix former documentation
