### 2026-06-18 902fff1
chore(config): fix RAPL sockets spec using config, clippy warnings, cargo fmt
