### 2026-06-18 cd02314
chore(config): fix RAPL sockets spec using config, clippy warnings, cargo fmt
