### 2026-06-19 544ffa1
test(nvml): fix tests after updating nvml hardware
