### 2026-07-01 fe66864
refactor(v3): beginning of new architecture
