+++ core/src/util/sys.rs
@@ -1,65 +1,65 @@
-use std::ffi::{CStr, CString};
+// use std::ffi::{CStr, CString};
 
-use crate::JouleProfilerError;
+// use crate::JouleProfilerError;
 
-/// Sends a signal to a process.
-///
-/// SAFETY
-///
-/// - 'pid' must refer to a valid process identifier.
-/// - Calling `libc::kill` is unsafe because it performs a raw syscall using the Linux FFI.
-///
-/// Race Condition
-///
-/// The target process may terminate before the
-/// signal is made. In that case, the system call will fail and
-/// an error will be returned and the signal name will
-/// be retrieved for better error handling.
-///
-/// Errors
-///
-/// Returns [`JouleProfilerError::ProcessControlFailed`] if the
-/// signal could not be delivered (e.g., invalid PID or already exited).
-pub fn signal(pid: i32, sig: i32) -> Result<(), JouleProfilerError> {
-    if unsafe { libc::kill(pid, sig) } != 0 {
-        let sig_name = unsafe {
-            let ptr = libc::strsignal(sig);
-            if ptr.is_null() {
-                "unknown signal".to_string()
-            } else {
-                CStr::from_ptr(ptr).to_string_lossy().to_string()
-            }
-        };
-        Err(JouleProfilerError::ProcessControlFailed(format!(
-            "Cannot send signal {sig_name} to process {pid}"
-        )))
-    } else {
-        Ok(())
-    }
-}
+// /// Sends a signal to a process.
+// ///
+// /// SAFETY
+// ///
+// /// - 'pid' must refer to a valid process identifier.
+// /// - Calling `libc::kill` is unsafe because it performs a raw syscall using the Linux FFI.
+// ///
+// /// Race Condition
+// ///
+// /// The target process may terminate before the
+// /// signal is made. In that case, the system call will fail and
+// /// an error will be returned and the signal name will
+// /// be retrieved for better error handling.
+// ///
+// /// Errors
+// ///
+// /// Returns [`JouleProfilerError::ProcessControlFailed`] if the
+// /// signal could not be delivered (e.g., invalid PID or already exited).
+// pub fn signal(pid: i32, sig: i32) -> Result<(), JouleProfilerError> {
+//     if unsafe { libc::kill(pid, sig) } != 0 {
+//         let sig_name = unsafe {
+//             let ptr = libc::strsignal(sig);
+//             if ptr.is_null() {
+//                 "unknown signal".to_string()
+//             } else {
+//                 CStr::from_ptr(ptr).to_string_lossy().to_string()
+//             }
+//         };
+//         Err(JouleProfilerError::ProcessControlFailed(format!(
+//             "Cannot send signal {sig_name} to process {pid}"
+//         )))
+//     } else {
+//         Ok(())
+//     }
+// }
 
-/// Gets the effective user id of the current process.
-///
-/// SAFETY
-///
-/// - Calling `libc::geteuid` is unsafe because it performs a raw syscall using the Linux FFI.
-pub fn geteuid() -> u32 {
-    unsafe { libc::geteuid() }
-}
+// /// Gets the effective user id of the current process.
+// ///
+// /// SAFETY
+// ///
+// /// - Calling `libc::geteuid` is unsafe because it performs a raw syscall using the Linux FFI.
+// pub fn geteuid() -> u32 {
+//     unsafe { libc::geteuid() }
+// }
 
-/// Gets the current process' user id from its username.
-///
-/// SAFETY
-///
-/// - Calling `libc::getpwnam` is unsafe because it uses the Linux FFI.
[…]
