### 2026-07-02 6e958ec
docs(nvml): added utilization counters doc
