### 2026-07-02 f690a0b
docs: update orchestrator documentation
