### 2026-07-02 54126eb
refactor: orchestrator refactoring for source initialization
