+++ core/src/profiler/mod.rs
@@ -17,7 +17,7 @@ use std::{
 pub mod error;
 
 use crate::config::ProfileConfig;
-use crate::orchestrator::SourceOrchestrator;
+use crate::orchestrator::Orchestrator;
 use crate::phase::{PhaseInfo, PhaseToken};
 use crate::profiler::types::{MeasurePhasesReturnType, Phase, ProfilerResults, Result};
 use crate::sensor::{Sensor, Sensors};
@@ -61,9 +61,6 @@ pub mod types;
 /// ```
 #[derive(Default)]
 pub struct JouleProfiler {
-    /// The sources orchestrator, managing sources and sending them the events sent by the profiler.
-    orchestrator: SourceOrchestrator,
-
     /// The different metric sources.
     sources: Vec<Box<dyn MetricSource>>,
 }
@@ -108,20 +105,31 @@ impl JouleProfiler {
 
     /// Profiles a program spawned with the configured command and return the aggregated results.
     ///
-    /// It starts the orchestrator with the metric sources and profile the program.
+    /// It spawns the program, initializes the metric sources with its pid, starts the
+    /// orchestrator, and profiles the program.
     pub async fn profile(&mut self, config: &ProfileConfig) -> Result<ProfilerResults> {
         info!("Running phase-based profiling");
         debug!("Phase regex: {}", config.token_pattern);
 
+        let regex = Regex::new(&config.token_pattern)?;
+
         let sources = std::mem::take(&mut self.sources);
-        trace!("Starting orchestrator with {} source(s)", sources.len());
-        self.orchestrator.run(sources, config.init_timeout)?;
+        let mut orchestrator = Orchestrator::new(sources);
+
+        debug!("Spawning command: {:?}", config.cmd);
+        let mut child = spawn_profiled_command(config)?;
+        let pid = child.id().cast_signed();
+
+        pause_prosess(pid)?;
+        orchestrator.init(pid, config.init_timeout).await?;
+        orchestrator.run()?;
 
         info!("Starting measurements");
-        let (command_duration_ms, timestamp, exit_code, detected_phases) =
-            self.measure_phases(config).await?;
+        let (command_duration_ms, timestamp, exit_code, detected_phases) = self
+            .measure_phases(&mut orchestrator, config, &regex, &mut child, pid)
+            .await?;
 
-        let (sources_results, sources) = self.orchestrator.finalize().await?;
+        let (sources_results, sources) = orchestrator.finalize().await?;
         self.sources = sources;
 
         let mut phases: Vec<_> = detected_phases
@@ -170,33 +178,27 @@ impl JouleProfiler {
         })
     }
 
-    /// Spawn the configured command and profile it, separating its execution into phases through tokens matching
-    /// a configured regular expression.
+    /// Profile the already spawned and paused command, separating its execution into phases
+    /// through tokens matching a configured regular expression.
     ///
     /// The profiling is composed of several steps:
     ///
-    /// - Firstly, the program is spawned and its pid is retrieved.
-    /// - The process is immediately stopped using a SIGSTOP signal to configure the sources without introducing a significant overhead.
-    /// - The first measure is made and the process is then resume.
+    /// - The metric sources have already been initialized with the profiled program's pid and the process paused
+    ///   (see [`JouleProfiler::profile`]), to configure the sources without introducing a significant overhead.
+    /// - The first measure is made and the process is then resumed.
     /// - The profiler listens for phase token in the standard output of the profiled process and make a measure for every token detected.
[…]
