+++ core/src/source/mod.rs
@@ -5,9 +5,10 @@
 //! Implementations are boxed for flexibility, while internally resolving
 //! concrete types to minimize the profiler overhead.
 
-use std::time::Duration;
+use std::future::Future;
+use std::pin::Pin;
 
-use tokio::sync::{mpsc, oneshot};
+use tokio::sync::mpsc;
 
 pub(crate) mod accumulator;
 pub mod error;
@@ -16,7 +17,6 @@ pub(crate) mod runtime;
 pub(crate) mod types;
 
 use crate::sensor::Sensors;
-use crate::source::runtime::MetricSourceRuntime;
 use crate::source::types::{SourceEvent, SourceWorkerHandle};
 pub use error::MetricSourceError;
 pub use reader::MetricReader;
@@ -28,58 +28,17 @@ pub use types::{MetricReaderErrorBound, MetricReaderTypeBound};
 /// This trait is used to erase the type of the metric source, to be able to have a
 /// convenient API for users while maintaining performance with monomorphization during hot paths.
 pub(crate) trait MetricSource: Send {
-    /// Spawn the source worker and return its handle, control channel and initialization channel.
-    fn run(
-        self: Box<Self>,
-        control_receiver: mpsc::Receiver<SourceEvent>,
-        init_receiver: oneshot::Receiver<i32>,
-        init_validation_sender: oneshot::Sender<Result<(), MetricSourceError>>,
-        init_timeout: Duration,
-    ) -> SourceWorkerHandle;
+    /// Initialize the source with the profiled program's pid.
+    ///
+    /// Must be awaited (and completed) before [`MetricSource::run`] is called.
+    fn init(
+        &mut self,
+        pid: i32,
+    ) -> Pin<Box<dyn Future<Output = Result<(), MetricSourceError>> + Send + '_>>;
+
+    /// Spawn the source worker and return its handle.
+    fn run(self: Box<Self>, control_receiver: mpsc::Receiver<SourceEvent>) -> SourceWorkerHandle;
 
     /// List sensors exposed by this source.
     fn list_sensors(&self) -> Result<Sensors, MetricSourceError>;
-}
-
-impl<R> MetricSource for MetricSourceRuntime<R>
-where
-    R: MetricReader,
-{
-    /// Runs the worker task and returns its handle and the sender, used to send events to manage the metric source.
-    ///
-    /// The metric source is consumed and transformed into a [`MetricSourceRuntime`] with the metric source as a reader.
-    /// This transformation allows to monomorphize the metric source and discover its type after its launch.
-    fn run(
-        self: Box<Self>,
-        control_receiver: mpsc::Receiver<SourceEvent>,
-        init_receiver: oneshot::Receiver<i32>,
-        init_validation_sender: oneshot::Sender<Result<(), MetricSourceError>>,
-        init_timeout: Duration,
-    ) -> SourceWorkerHandle {
-        tokio::spawn(async move {
-            self.run_worker(
-                control_receiver,
-                init_receiver,
-                init_validation_sender,
-                init_timeout,
-            )
-            .await
-        })
-    }
-
-    /// List the sensors of the metric source.
-    fn list_sensors(&self) -> Result<Sensors, MetricSourceError> {
-        self.get_source_sensors()
-    }
-}
-
-/// Converts a [`MetricReader`] into a boxed [`MetricSource`].
-impl<R> From<R> for Box<dyn MetricSource>
-where
-    R: MetricReader,
-{
-    fn from(reader: R) -> Self {
-        let source = MetricSourceRuntime::new(reader);
-        Box::new(source)
-    }
 }
