### 2026-07-02 a0ee7c2
refactor: orchestrator refactoring for source initialization
