+++ core/src/orchestrator/mod.rs
@@ -6,15 +6,13 @@ use std::time::Duration;
 
 use crate::aggregate::sensor_result::SensorResult;
 use crate::orchestrator::error::OrchestratorError;
-use crate::source::error::IntoMetricSourceError;
 use crate::source::types::SourceEvent;
 use crate::source::{MetricSource, MetricSourceError};
+use crate::util::time::get_timestamp_micros;
 use futures::future::try_join_all;
+use log::{debug, trace};
 use tokio::time::timeout;
-use tokio::{
-    sync::{mpsc, oneshot},
-    task::JoinHandle,
-};
+use tokio::{sync::mpsc, task::JoinHandle};
 
 pub mod error;
 
@@ -26,31 +24,9 @@ pub const CONTROL_CHANNEL_SIZE: usize = 16;
 /// The handle describing the return type of a source worker.
 type TaskHandle = JoinHandle<Result<(SensorResult, Box<dyn MetricSource>), MetricSourceError>>;
 
-struct InitChannel {
-    sender: oneshot::Sender<i32>,
-    validation: oneshot::Receiver<Result<(), MetricSourceError>>,
-}
-
-impl InitChannel {
-    async fn initialize(self, pid: i32) -> Result<(), OrchestratorError> {
-        self.sender
-            .send(pid)
-            .map_err(|_| OrchestratorError::InitializationError("Failed to send PID"))?;
-
-        self.validation
-            .await
-            .map_err(|e| OrchestratorError::MetricSourceError(e.into_metric_source_error()))??;
-
-        Ok(())
-    }
-}
-
 struct SourceHandle {
     /// The event channel sender used to manage the metric sources.
     control_sender: mpsc::Sender<SourceEvent>,
-
-    /// The channels used for sources initialization.
-    init_channel: Option<InitChannel>,
 
     /// The handle of the worker task, used for joining sources gracefully.
     handle: TaskHandle,
@@ -58,54 +34,90 @@ struct SourceHandle {
 
 /// Orchestrates the metric sources and send them the profiler's messages through asynchronous channels.
 /// It is a proxy between the profiler and the sources and is responsible of their lifecycle.
-#[derive(Default)]
-pub struct SourceOrchestrator {
+pub struct Orchestrator {
+    /// Sources initialized through [`SourceOrchestrator::init`], not yet started by [`SourceOrchestrator::run`].
+    sources: Vec<Box<dyn MetricSource>>,
+
     handles: Vec<SourceHandle>,
-    init_timeout: Duration,
 }
 
-impl SourceOrchestrator {
-    /// Starts all the metric sources.
+impl Orchestrator {
+    pub fn new(sources: Vec<Box<dyn MetricSource>>) -> Self {
+        Self {
+            sources,
+            handles: Vec::new(),
+        }
+    }
+
+    /// Initializes each metric source with the profiled program's pid, bounded by `init_timeout`.
     ///
-    /// The function uses a one shot sender to forward the profiled program's pid to the sources, used by some for per-process profiling (e.g. `perf_event`).
-    /// Stores the sources handles and the channels senders to be able to gracefully join the sources and send events.
-    #[inline]
-    pub fn run(
+    /// Used by some sources for per-process profiling (e.g. `perf_event`).
+    /// Stores the initialized sources, ready to be started with [`SourceOrchestrator::run`].
+    ///
+    /// Each source is initialized in its own task: some [`MetricReader::init`](`crate::source::MetricReader::init`)
+    /// implementations perform blocking work without yielding, which would otherwise prevent
+    /// `init_timeout` from being enforced if awaited directly on the calling task.
+    pub async fn init(
         &mut self,
-        sources: Vec<Box<dyn MetricSource>>,
[…]
