+++ core/src/source/runtime.rs
@@ -1,17 +1,16 @@
-use std::time::Duration;
+use std::pin::Pin;
 
 use log::debug;
-use tokio::{
-    sync::{mpsc, oneshot},
-    time::timeout,
-};
+use tokio::sync::mpsc;
 
 use crate::{
     aggregate::{phase::SensorPhase, sensor_result::SensorResult},
     sensor::Sensors,
     source::{
-        MetricReader, MetricSource, MetricSourceError, accumulator::MetricAccumulator,
-        error::IntoMetricSourceError, types::SourceEvent,
+        MetricReader, MetricSource, MetricSourceError,
+        accumulator::MetricAccumulator,
+        error::IntoMetricSourceError,
+        types::{SourceEvent, SourceWorkerHandle},
     },
 };
 
@@ -35,20 +34,11 @@ impl<R: MetricReader> MetricSourceRuntime<R> {
     /// Runs the worker responsible for source and accumulator management.
     ///
     /// It listens for events through a channel and execute them.
-    pub async fn run_worker(
+    /// The source must already be initialized (see [`MetricSource::init`]).
+    pub(crate) async fn run_worker(
         mut self,
         mut receiver: mpsc::Receiver<SourceEvent>,
-        init_receiver: oneshot::Receiver<i32>,
-        init_validation_sender: oneshot::Sender<Result<(), MetricSourceError>>,
-        init_timeout: Duration,
     ) -> Result<(SensorResult, Box<dyn MetricSource>), MetricSourceError> {
-        let pid = timeout(init_timeout, init_receiver)
-            .await
-            .map_err(IntoMetricSourceError::into_metric_source_error)?
-            .map_err(|_| MetricSourceError::InitTimeout(init_timeout))?;
-
-        let _ = init_validation_sender.send(self.init_source(pid).await);
-
         loop {
             if let Some(event) = receiver.recv().await {
                 match event {
@@ -73,15 +63,6 @@ impl<R: MetricReader> MetricSourceRuntime<R> {
     async fn measure_source(&mut self) -> Result<(), MetricSourceError> {
         self.source
             .measure()
-            .await
-            .map_err(IntoMetricSourceError::into_metric_source_error)
-    }
-
-    /// Init the source with the profiled program pid.
-    #[inline]
-    async fn init_source(&mut self, pid: i32) -> Result<(), MetricSourceError> {
-        self.source
-            .init(pid)
             .await
             .map_err(IntoMetricSourceError::into_metric_source_error)
     }
@@ -116,13 +97,49 @@ impl<R: MetricReader> MetricSourceRuntime<R> {
             .collect::<Result<Vec<_>, MetricSourceError>>()?;
         Ok(SensorResult { phases: result })
     }
+}
 
-    /// Retrieve source sensors.
-    #[inline]
-    pub fn get_source_sensors(&self) -> Result<Sensors, MetricSourceError> {
+impl<R> MetricSource for MetricSourceRuntime<R>
+where
+    R: MetricReader,
+{
+    /// Initializes the source with the profiled program's pid.
+    fn init(
+        &mut self,
+        pid: i32,
+    ) -> Pin<Box<dyn Future<Output = Result<(), MetricSourceError>> + Send + '_>> {
+        Box::pin(async move {
+            self.source
+                .init(pid)
+                .await
+                .map_err(IntoMetricSourceError::into_metric_source_error)
+        })
+    }
+
+    /// Spawns the worker task and returns its handle, used to send events to manage the metric source.
+    ///
+    /// The metric source is consumed and transformed into a [`MetricSourceRuntime`] with the metric source as a reader.
[…]
