### 2026-07-02 42c71b8
refactor: orchestrator refactoring for source initialization
