### 2026-07-02 02a626b
refactor: orchestrator refactoring
