### 2026-07-02 2da55e8
feat(cgroup): added attach_pid and create_cgroup to cgroup config for already created cgroup or already attached pid
