### 2026-07-02 effb57c
chore: cargo fmt
