### 2026-07-02 45e25e7
refactor(cgroup): improved error handling, readme documentation, improved configuration and removed controller creation (delegated to user for performance)
