+++ sources/cgroup/src/cgroup.rs
@@ -1,19 +1,17 @@
 //! cgroup v2 management utilities.
 //!
 //! This module provides:
-//! - controller activation (`cpu`, `memory`, `io`);
 //! - child cgroup creation and cleanup;
 //! - process attachment to cgroups;
 //! - CPU, memory, and I/O statistics collection.
 
 use std::{
-    collections::HashSet,
-    fmt::Display,
     fs,
+    io::ErrorKind,
     path::{Path, PathBuf},
 };
 
-use log::{debug, trace, warn};
+use log::{debug, warn};
 
 use crate::{
     Result,
@@ -22,37 +20,13 @@ use crate::{
     util::{read_flat_keyed_file, read_io_stat, read_u64_opt},
 };
 
-/// Available cgroup v2 controllers.
-#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
-pub enum Controller {
-    /// I/O controller.
-    Io,
-
-    /// Memory controller.
-    Memory,
-
-    /// CPU controller.
-    Cpu,
-}
-
-impl Display for Controller {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(match self {
-            Controller::Io => "io",
-            Controller::Memory => "memory",
-            Controller::Cpu => "cpu",
-        })
-    }
-}
-
 /// Interface for reading cgroup statistics.
 pub trait CgroupBackend: Send + Sync + 'static {
     /// Initializes the backend.
     fn create(&self, path: &Path) -> Result<()>;
 
+    /// Attaches the provided pid to the cgroup.
     fn attach_pid(&self, path: &Path, pid: i32) -> Result<()>;
-
-    fn enable_controllers(&self, root: &Path, controllers: &HashSet<Controller>) -> Result<()>;
 
     /// Cleanup the backend.
     fn cleanup(&self, path: &Path, root: &Path) -> Result<()>;
@@ -84,44 +58,32 @@ impl SysFsBackend {
 }
 
 impl CgroupBackend for SysFsBackend {
-    /// Creates the cgroup directory on disk.
+    /// Creates the cgroup directory.
     fn create(&self, path: &Path) -> Result<()> {
         debug!("Initializing cgroup at \"{}\"", path.display());
-        fs::create_dir_all(path)?;
+        fs::create_dir_all(path).map_err(|err| match err.kind() {
+            ErrorKind::PermissionDenied => {
+                CgroupError::PermissionDenied("creating a cgroup requires root privileges.")
+            }
+            _ => err.into(),
+        })?;
         Ok(())
     }
 
     /// Attaches a process PID to the cgroup.
     fn attach_pid(&self, path: &Path, pid: i32) -> Result<()> {
         let procs_path = path.join("cgroup.procs");
-        fs::write(&procs_path, pid.to_string()).map_err(|err| CgroupError::FailedToAttachPid {
-            pid,
-            path: procs_path,
-            source: err,
+        fs::write(&procs_path, pid.to_string()).map_err(|err| match err.kind() {
+            ErrorKind::PermissionDenied => CgroupError::PermissionDenied(
+                "attaching a process to this cgroup requires root privileges.",
+            ),
+            _ => CgroupError::FailedToAttachPid {
+                pid,
+                path: procs_path,
+                source: err,
+            },
         })?;
         debug!("Attached PID {pid} to cgroup {}", path.display());
-        Ok(())
-    }
-
-    /// Enables the provided controllers in `cgroup.subtree_control`.
-    ///
-    /// If a controller is already activated, then nothing happens for it.
[…]
