### 2026-07-02 ef6c714
refactor(cgroup): improved error handling, readme documentation, improved configuration and removed controller creation (delegated to user for performance)
