### 2026-07-02 81441df
fix(procfs): ignore measure when race condition between process exit and measure, also clippy warnings fix
