### 2026-07-02 9ad8260
fix(procfs): ignore measure when race condition between process exit and measure, also clippy warnings fix
