### 2026-07-02 5a35a96
docs: update doc for metric reader type bound
