### 2026-07-02 41efa20
chore(cgroup): put default polling interval to 1ms
