### 2026-07-02 b5a27eb
chore(cgroup): put default polling interval to 1ms
