### 2026-07-02 4303438
chore(cgroup): put default polling interval to 1ms
