### 2026-07-03 9978ea8
chore(nvml): changed polling interval to 20 hz because the majority of counters does not have a higher frequency
