### 2026-07-03 bf1d261
chore(nvml): changed polling interval to 20 hz because the majority of counters does not have a higher frequency
