### 2026-07-03 a7ce9ba
chore(nvml): removed UUID and use indexes for gpu spec
