### 2026-07-03 bc810bc
chore(nvml): removed UUID and use indexes for gpu spec
