### 2026-07-03 02b9ec6
perf(amdsmi): parallelized AMD SMI measurements into blocking thread pool
