### 2026-07-03 b19ad25
chore: cargo fmt
