### 2026-07-03 061d7c3
chore(amdsmi): changed default polling interval to 20 milliseconds
