### 2026-07-03 b92587a
chore(amdsmi): changed default polling interval to 20 milliseconds
