+++ cli/src/bin/main.rs
@@ -1,5 +1,3 @@
-use std::time::Duration;
-
 use anyhow::Result;
 use joule_profiler_cli::{
     CliArgs, ProfilerCommand, RaplBackend, Source, init_logging, output_format_to_displayer,
@@ -69,10 +67,7 @@ async fn main() -> Result<()> {
     }
 
     if cli.sources.contains(&Source::AmdSmi) {
-        match AmdSmi::new(AmdSmiConfig {
-            poll_interval: Some(Duration::from_millis(1)),
-            ..Default::default()
-        }) {
+        match AmdSmi::new(AmdSmiConfig::default()) {
             Ok(amdsmi) => {
                 trace!("Using AMD SMI for AMD GPU profiling");
                 profiler.add_source(amdsmi);
