### 2026-07-03 8055710
chore(amdsmi): typo in README.md
