### 2026-07-03 e2451a7
feat(config-file): update all sources configuration and improved configuration file example
