+++ sources/cgroup/src/config.rs
@@ -1,7 +1,16 @@
 use std::{path::PathBuf, time::Duration};
 
+use serde::Deserialize;
+
+const DEFAULT_POLL_INTERVAL: Duration = Duration::from_millis(10);
+
+fn default_poll_interval() -> Duration {
+    DEFAULT_POLL_INTERVAL
+}
+
 /// Configuration for the cgroup metric source.
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Deserialize)]
+#[serde(default)]
 pub struct CgroupConfig {
     /// Path to cgroup v2 hierarchy (usually `/sys/fs/cgroup`).
     pub cgroup_root: Option<PathBuf>,
@@ -10,6 +19,7 @@ pub struct CgroupConfig {
     pub cgroup_name: String,
 
     /// Background polling interval.
+    #[serde(default = "default_poll_interval", with = "humantime_serde")]
     pub poll_interval: Duration,
 
     /// Whether the source must attach the process pid to the cgroup or not.
@@ -24,7 +34,7 @@ impl Default for CgroupConfig {
         Self {
             cgroup_root: None,
             cgroup_name: format!("joule-profiler-{}", std::process::id()),
-            poll_interval: Duration::from_millis(1),
+            poll_interval: DEFAULT_POLL_INTERVAL,
             attach_pid: true,
             create_cgroup: true,
         }
