+++ sources/procfs/Cargo.toml
@@ -15,6 +15,8 @@ log.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 futures.workspace = true
+serde.workspace = true
+humantime-serde.workspace = true
 
 procfs = "0.18.0"
 tokio-timerfd = "0.2.0"

[…]
