### 2026-07-03 dd6ef90
feat(config-file): update all sources configuration and improved configuration file example
