### 2026-07-03 399a49d
feat(config-file): update all sources configuration and improved configuration file example
