+++ sources/amdsmi/src/lib.rs
@@ -76,7 +76,7 @@ pub struct Processor {
     support: ProcessorSupport,
 }
 
-pub struct AmdSmi<H: AmdSmiHardware> {
+pub struct AmdSmi<H: AmdSmiHardware = AmdSmiWrapperHardware> {
     /// Source configuration.
     config: AmdSmiConfig,
 
@@ -100,30 +100,6 @@ pub struct AmdSmi<H: AmdSmiHardware> {
 
     /// The current GPU utilization counters.
     utilization_counters: Arc<Mutex<HashMap<usize, UtilizationCounter>>>,
-}
-
-impl AmdSmi<AmdSmiWrapperHardware> {
-    /// Initializes the AMD SMI source and retrieve the GPU devices.
-    pub fn new(config: AmdSmiConfig) -> Result<Self> {
-        let mut amdsmi = AmdSmiWrapperHardware::new()?;
-
-        let processors = amdsmi
-            .init_processors(config.gpus_spec.as_ref())?
-            .into_iter()
-            .enumerate()
-            .collect();
-
-        Ok(Self {
-            config,
-            hardware: Arc::new(amdsmi),
-            handle: None,
-            processors: Arc::new(processors),
-            energy_counters: HashMap::new(),
-            vram_counters: Arc::default(),
-            power_counters: Arc::default(),
-            utilization_counters: Arc::default(),
-        })
-    }
 }
 
 impl<H: AmdSmiHardware> AmdSmi<H> {
@@ -243,8 +219,29 @@ impl<H: AmdSmiHardware> AmdSmi<H> {
 
 impl<H: AmdSmiHardware> MetricReader for AmdSmi<H> {
     type Type = HashMap<usize, Counter>;
-
     type Error = AmdSmiError;
+    type Config = AmdSmiConfig;
+
+    /// Initializes the AMD SMI source and retrieve the GPU devices.
+    fn from_config(config: AmdSmiConfig) -> Result<Self> {
+        let mut amdsmi = H::new()?;
+        let processors = amdsmi
+            .init_processors(config.gpus_spec.as_ref())?
+            .into_iter()
+            .enumerate()
+            .collect();
+
+        Ok(Self {
+            config,
+            hardware: Arc::new(amdsmi),
+            handle: None,
+            processors: Arc::new(processors),
+            energy_counters: HashMap::new(),
+            vram_counters: Arc::default(),
+            power_counters: Arc::default(),
+            utilization_counters: Arc::default(),
+        })
+    }
 
     /// Makes a measurement for every devices.
     ///
@@ -336,16 +333,14 @@ impl<H: AmdSmiHardware> MetricReader for AmdSmi<H> {
 
     /// Creates the polling task if a polling interval has been configured.
     async fn init(&mut self, _pid: i32) -> Result<()> {
-        if let Some(poll_interval) = self.config.poll_interval {
-            self.handle = Some(Self::create_worker(
-                self.hardware.clone(),
-                self.processors.clone(),
-                self.power_counters.clone(),
-                self.vram_counters.clone(),
-                self.utilization_counters.clone(),
-                poll_interval,
-            )?);
-        }
+        self.handle = Some(Self::create_worker(
+            self.hardware.clone(),
+            self.processors.clone(),
+            self.power_counters.clone(),
+            self.vram_counters.clone(),
+            self.utilization_counters.clone(),
+            self.config.poll_interval,
+        )?);
 
         debug!("AMD SMI source initialized.");
         Ok(())
[…]
