### 2026-07-03 98838e1
chore: changed output format from --json --csv to --output-format and fixed file creation when no parent path is provided.
