+++ cli/src/output/formats/mod.rs
@@ -9,6 +9,7 @@
 //! - [`OutputFormat`] — Enum representing the available output formats.
 //! - `csv`, `json`, `terminal` — Submodules implementing the actual display logic for default output formats.
 
+use clap::ValueEnum;
 use std::fmt::{Display, Formatter, Result};
 
 use serde::Deserialize;
@@ -26,10 +27,10 @@ pub mod terminal;
 ///
 /// # Variants
 ///
-/// - `Terminal` — Display metrics directly in the terminal (default).
-/// - `Json` — Export metrics as JSON for easy parsing or integration.
-/// - `Csv` — Export metrics in CSV format for spreadsheets or analysis.
-#[derive(Debug, Clone, Copy, Default, Deserialize)]
+/// - `Terminal` - Display metrics directly in the terminal (default).
+/// - `Json` - Export metrics as JSON for easy parsing or integration.
+/// - `Csv` - Export metrics in CSV format for spreadsheets or analysis.
+#[derive(Debug, Clone, Copy, Default, ValueEnum, Deserialize)]
 pub enum OutputFormat {
     #[default]
     #[serde(alias = "terminal", alias = "term", alias = "TERM")]
