### 2026-07-03 c1cb02b
chore: changed output format from --json --csv to --output-format and fixed file creation when no parent path is provided.
