### 2026-07-03 2b1acae
chore: changed output format from --json --csv to --output-format and fixed file creation when no parent path is provided.
