### 2026-07-06 0680f2a
fix(config): all sources configurations deserialization fixed
