+++ cli/src/lib.rs
@@ -43,7 +43,7 @@ pub struct CliArgs {
     #[arg(long = "rapl-path")]
     pub rapl_path: Option<String>,
 
-    /// Sockets to measure (e.g., 0 or 0,1)
+    /// Sockets to measure. (e.g., 0 or 0,1)
     #[arg(short = 's', long = "sockets")]
     pub sockets: Option<String>,
 
@@ -51,11 +51,11 @@ pub struct CliArgs {
     #[arg(long = "output-format")]
     pub output_format: Option<OutputFormat>,
 
-    /// Output file for CSV/JSON (else `data<TIMESTAMP>`.csv/json)
+    /// Output file for CSV/JSON. (else `data<TIMESTAMP>`.csv/json)
     #[arg(short = 'o', long = "output-file")]
     pub output_file: Option<String>,
 
-    /// Choose RAPL backend between powercap or perf
+    /// Choose RAPL backend between powercap or perf.
     #[arg(long = "rapl-backend", value_enum)]
     pub rapl_backend: Option<RaplBackend>,
 
@@ -66,7 +66,7 @@ pub struct CliArgs {
     #[arg(long = "config")]
     pub config_file: Option<PathBuf>,
 
-    /// The command to execute
+    /// The command to execute.
     #[command(subcommand)]
     pub command: ProfilerCommand,
 }
@@ -107,7 +107,7 @@ impl std::fmt::Display for Source {
             Source::Rapl => "rapl",
             Source::Nvml => "nvml",
             Source::AmdSmi => "amdsmi",
-            Source::Perf => "perf | perf_event",
+            Source::Perf => "perf",
             Source::Procfs => "procfs",
             Source::Cgroup => "cgroup",
         };
@@ -118,26 +118,20 @@ impl std::fmt::Display for Source {
 #[derive(Debug, Default, Clone, ValueEnum, Deserialize)]
 pub enum RaplBackend {
     #[default]
-    #[serde(rename = "perf", alias = "perf_event")]
+    #[serde(rename = "perf")]
     Perf,
     #[serde(rename = "powercap")]
     Powercap,
 }
 
-pub fn config_table_to_displayer(
-    config_table: &ConfigTable,
-    cli: &CliArgs,
-) -> Result<Box<dyn Displayer>> {
-    let output_format = cli
-        .output_format
-        .unwrap_or(config_table.profiler_config.output_format);
-    let output_file = cli
-        .output_file
-        .as_ref()
-        .or(config_table.profiler_config.output_file.as_ref())
-        .cloned();
+/// Builds the displayer for the configured output format.
+///
+/// Reads only from `config_table`: `apply_cli` must have already resolved
+/// any CLI override into `profiler_config` beforehand.
+pub fn config_table_to_displayer(config_table: &ConfigTable) -> Result<Box<dyn Displayer>> {
+    let output_file = config_table.profiler_config.output_file.clone();
 
-    let displayer = match output_format {
+    let displayer = match config_table.profiler_config.output_format {
         OutputFormat::Terminal => TerminalOutput.into(),
         OutputFormat::Json => JsonOutput::new(output_file)?.into(),
         OutputFormat::Csv => CsvOutput::try_new(output_file)?.into(),
@@ -155,4 +149,100 @@ pub fn parse_sockets_spec(sockets_spec: &str) -> HashSet<u32> {
         .split(',')
         .filter_map(|x| x.trim().parse::<u32>().ok())
         .collect()
+}
+
+#[cfg(test)]
+mod tests {
+    use clap::ValueEnum;
+    use joule_profiler_core::source::MetricReader;
+
+    use super::*;
+
[…]
