### 2026-07-06 ee02c1d
fix(config): all sources configurations deserialization fixed
