### 2026-07-06 aff9df4
fix(config): all sources configurations deserialization fixed
