+++ cli/src/config/table.rs
@@ -20,6 +20,7 @@ use crate::{
 
 /// Holds the resolved configuration state, merging values from the global
 /// config file and CLI arguments before they are built into a final.
+#[derive(Debug)]
 pub struct ConfigTable {
     /// Global Joule Profiler configuration.
     pub profiler_config: ProfilerConfig,
@@ -53,7 +54,13 @@ impl ConfigTable {
         }
     }
 
-    /// Applies CLI overrides to Joule Profiler global configuration.
+    /// Applies every CLI override of the global (profiler-level) configuration
+    /// in one pass, so [`ConfigTable::to_config`] can read `profiler_config`
+    /// directly afterwards instead of re-merging `cli` itself.
+    ///
+    /// Per-source overrides (e.g. RAPL's sockets/path/polling) are a separate
+    /// concern and stay as override closures at each source's registration
+    /// call site (see `register_source_override`).
     pub fn apply_cli(&mut self, cli: &mut CliArgs) {
         if let Some(output_format) = cli.output_format.take() {
             self.profiler_config.output_format = output_format;
@@ -67,7 +74,19 @@ impl ConfigTable {
             self.profiler_config.rapl_backend = rapl_backend;
         }
 
-        if let ProfilerCommand::Profile(profile_args) = &cli.command {
+        if let ProfilerCommand::Profile(profile_args) = &mut cli.command {
+            if let Some(stdout_file) = profile_args.stdout_file.take() {
+                self.profiler_config.stdout_file = Some(stdout_file);
+            }
+
+            if let Some(token_pattern) = profile_args.token_pattern.take() {
+                self.profiler_config.token_pattern = token_pattern;
+            }
+
+            if let Some(init_timeout) = profile_args.init_timeout.take() {
+                self.profiler_config.init_timeout = init_timeout;
+            }
+
             self.profiler_config.use_root |= profile_args.use_root;
         }
     }
@@ -152,35 +171,18 @@ impl ConfigTable {
     /// Consumes the [`ConfigTable`] and a final [`CliArgs`] to produce the
     /// core [`Config`].
     ///
-    /// CLI values override config file values for all fields
-    /// resolved here (stdout file, root mode, timeouts, token patterns).
-    ///
     /// Returns an error if the configuration is invalid.
     pub fn to_config(self, cli: CliArgs) -> Result<Config> {
         let command = match cli.command {
             ProfilerCommand::Profile(profile_args) => {
-                let stdout_file = profile_args
-                    .stdout_file
-                    .or(self.profiler_config.stdout_file);
-
-                let use_root = profile_args.use_root || self.profiler_config.use_root;
-
-                let init_timeout = profile_args
-                    .init_timeout
-                    .unwrap_or(self.profiler_config.init_timeout);
-
-                let token_pattern = profile_args
-                    .token_pattern
-                    .unwrap_or(self.profiler_config.token_pattern);
-
                 let mut builder = ProfileConfigBuilder::default();
 
                 let config = builder
                     .cmd(profile_args.cmd)
-                    .stdout_file(stdout_file)
-                    .token_pattern(token_pattern)
-                    .use_root(use_root)
-                    .init_timeout(init_timeout)
+                    .stdout_file(self.profiler_config.stdout_file)
+                    .token_pattern(self.profiler_config.token_pattern)
+                    .use_root(self.profiler_config.use_root)
+                    .init_timeout(self.profiler_config.init_timeout)
[…]
