### 2026-07-06 0025838
feat(perf): added cgroup scoped counters for perf_event source, optimized reading with spawn blocking and added pre_init function to do some sources work before spawning process
