+++ sources/perf_event/src/lib.rs
@@ -6,19 +6,22 @@
 //! Note: Counters are created individually (not grouped) because
 //! `inherit(true)` is incompatible with `perf_event` groups on Linux.
 
+use std::fs::File;
+
 use joule_profiler_core::{
     sensor::{Sensor, Sensors},
     source::MetricReader,
+    time::get_timestamp_micros,
     types::{Metric, Metrics},
     unit::{MetricUnit, Unit, UnitPrefix},
 };
-use log::{debug, info, trace};
+use log::{debug, info, trace, warn};
 
 use crate::{
     config::PerfConfig,
     error::PerfEventError,
     event::{EVENTS, Event},
-    hardware::{PerfEventCounters, PerfEventHardware},
+    hardware::{PerfEventCounters, PerfEventHardware, Target},
     snapshot::{Phase, Snapshot},
 };
 
@@ -35,16 +38,22 @@ const PERF_EVENT_METRIC_UNIT: MetricUnit = MetricUnit {
     unit: Unit::Count,
 };
 
+/// Root of the cgroup v2 hierarchy that `PerfConfig::cgroup_name` is
+/// resolved against.
+const CGROUP_ROOT: &str = "/sys/fs/cgroup";
+
 /// Hardware performance counter source using `perf_event`.
 ///
 /// Tracks CPU performance metrics (cycles, instructions, cache/branch misses)
-/// for a specific process.
+/// for a specific process, or for every process inside a cgroup if
+/// `PerfConfig::cgroup_name` is set.
 ///
 /// The hardware generic type is used for testing purposes, it allows to change the implementation
 /// used to interact with `perf_event`. The default adapter use the `perf_event2` library.
 pub struct PerfEvent<H: PerfEventHardware = PerfEventCounters> {
     hardware: H,
     events: Vec<Event>,
+    cgroup_name: Option<std::path::PathBuf>,
     begin_snapshot: Option<Snapshot>,
     last_snapshot: Option<Snapshot>,
 }
@@ -52,7 +61,6 @@ pub struct PerfEvent<H: PerfEventHardware = PerfEventCounters> {
 impl<H: PerfEventHardware + 'static> MetricReader for PerfEvent<H> {
     type Type = Phase;
     type Error = PerfEventError;
-
     type Config = PerfConfig;
 
     fn from_config(config: PerfConfig) -> Result<Self> {
@@ -60,22 +68,51 @@ impl<H: PerfEventHardware + 'static> MetricReader for PerfEvent<H> {
             events: config
                 .events
                 .map_or(EVENTS.to_vec(), |e| e.into_iter().collect()),
+            cgroup_name: config.cgroup_name,
             hardware: H::default(),
             begin_snapshot: None,
             last_snapshot: None,
         })
     }
 
-    /// Initialize counters for a specific process and start monitoring.
+    async fn pre_init(&mut self) -> Result<()> {
+        if let Some(name) = &self.cgroup_name {
+            let path = std::path::Path::new(CGROUP_ROOT).join(name);
+            info!(
+                "Initializing perf_event source for cgroup {}",
+                path.display()
+            );
+            let target = Target::Cgroup(File::open(path)?);
+            let begin = get_timestamp_micros();
+            self.hardware.init_counters(&self.events, target).await?;
+            let end = get_timestamp_micros();
+            warn!("init cgroup {}", end - begin);
+        }
+        Ok(())
+    }
+
+    /// Initialize counters and start monitoring: either the given process's
+    /// pid, or the configured cgroup if `PerfConfig::cgroup_name` was set.
     async fn init(&mut self, pid: i32) -> Result<()> {
-        info!("Initializing perf_event source for PID {pid}");
-        self.hardware.init_counters(&self.events, pid)
+        if self.cgroup_name.is_none() {
[…]
