+++ sources/perf_event/src/config.rs
@@ -1,4 +1,4 @@
-use std::collections::HashSet;
+use std::{collections::HashSet, path::PathBuf};
 
 use serde::Deserialize;
 
@@ -6,5 +6,17 @@ use crate::event::Event;
 
 #[derive(Debug, Default, Deserialize)]
 pub struct PerfConfig {
+    /// The events to monitor.
     pub events: Option<HashSet<Event>>,
+
+    /// Name of a cgroup v2 directory under `/sys/fs/cgroup` to monitor
+    /// instead of the profiled process's pid (e.g. `my-cgroup` or
+    /// `parent/child` for a nested cgroup). When set, counters are scoped to
+    /// every process inside that cgroup rather than a single pid and its
+    /// children.
+    /// It can be better to open counters with a cgroup rather than a pid,
+    /// because it allows to open the counters before spawning the process.
+    ///
+    /// The cgroup must be created before opening the counters.
+    pub cgroup_name: Option<PathBuf>,
 }
