+++ sources/perf_event/src/hardware.rs
@@ -1,59 +1,259 @@
 use crate::{Result, error::PerfEventError, event::Event, snapshot::Snapshot};
+use futures::future::try_join_all;
 use log::{debug, info, trace};
-use perf_event::{Builder, Counter, events::Hardware};
-use std::collections::HashMap;
+use perf_event::{
+    Builder, Counter, Group, ReadFormat,
+    events::{Hardware, Software},
+};
+use std::{
+    collections::HashMap,
+    fs::{self, File},
+    sync::Arc,
+};
+use tokio::task::spawn_blocking;
+
+/// What `perf_event` counters should be scoped to.
+pub enum Target {
+    /// Track a single process and any children it spawns, across every CPU
+    /// it runs on (via `inherit`).
+    Pid(i32),
+    /// Track every process inside a cgroup. Cgroup-scoped events don't
+    /// support `inherit`, so this is monitored with one group per online CPU.
+    Cgroup(File),
+}
 
 #[cfg_attr(test, mockall::automock)]
 pub trait PerfEventHardware: Send + Default {
-    fn init_counters(&mut self, events: &[Event], pid: i32) -> Result<()>;
-    fn read_snapshot(&mut self) -> Result<Snapshot>;
+    fn init_counters(
+        &mut self,
+        events: &[Event],
+        target: Target,
+    ) -> impl Future<Output = Result<()>> + Send;
+    fn read_snapshot(&mut self) -> impl Future<Output = Result<Snapshot>> + Send;
 }
+
+/// Per-CPU `perf_event` groups scoped to a cgroup. Kept alongside the cgroup's
+/// file descriptor, which must stay open for as long as the groups are.
+struct CgroupGroups {
+    _cgroup_fd: Arc<File>,
+    groups: HashMap<usize, Group>,
+    counters: HashMap<usize, HashMap<Event, Counter>>,
+}
+
+enum State {
+    /// Individual per-event counters, not yet initialized or PID-scoped.
+    Pid(HashMap<Event, Counter>),
+    Cgroup(CgroupGroups),
+}
+
+impl Default for State {
+    fn default() -> Self {
+        State::Pid(HashMap::new())
+    }
+}
+
+/// Implicit CPU key used for PID-scoped counters, which have
+/// no real per-CPU breakdown (`inherit` follows the process across CPUs).
+const PID_CPU: usize = 0;
 
 #[derive(Default)]
 pub struct PerfEventCounters {
-    counters: HashMap<Event, Counter>,
+    state: State,
 }
 
 impl PerfEventHardware for PerfEventCounters {
-    /// Create individual counters for all configured events.
-    ///
-    /// Each counter is built separately with `inherit(true)` and `observe_pid`,
-    /// since grouped counters do not support inheritance.
-    fn init_counters(&mut self, events: &[Event], pid: i32) -> Result<()> {
-        debug!("Adding {} individual performance counters", events.len());
-        for event in events {
-            trace!("Building counter: {event:?}");
-            let counter = Builder::new(Hardware::from(*event))
-                .inherit(true)
-                .observe_pid(pid)
-                .include_hv()
-                .include_kernel()
-                .build()?;
-            self.counters.insert(*event, counter);
+    async fn init_counters(&mut self, events: &[Event], target: Target) -> Result<()> {
+        match target {
+            Target::Pid(pid) => self.init_pid_counters(events, pid).await,
+            Target::Cgroup(cgroup_fd) => self.init_cgroup_counters(events, cgroup_fd).await,
         }
-        info!("Initialized {} hardware performance counters", events.len());
-        for (event, counter) in &mut self.counters {
-            trace!("Enabling counter: {event:?}");
-            counter.enable()?;
-        }
[…]
