+++ sources/perf_event/Cargo.toml
@@ -14,12 +14,13 @@ joule-profiler-core.workspace = true
 log.workspace = true
 thiserror.workspace = true
 serde.workspace = true
+tokio.workspace = true
+futures.workspace = true
 
 perf-event2 = "0.7.4"
 
 [dev-dependencies]
 mockall.workspace = true
-tokio.workspace = true
 
 [lints]
 workspace = true
