### 2026-07-06 0df4f37
refactor(rapl): moved perf counters enable to pre_init function
