### 2026-07-06 df802a5
refactor(rapl): open rapl counters in pre_init instead of in construction
