+++ sources/rapl/src/perf/socket.rs
@@ -17,16 +17,19 @@ const CPU_TOPOLOGY_SOCKET_ID: &str = "/topology/physical_package_id";
 /// Path to the online CPUs file in sysfs.
 const ONLINE_CPU_SYSFS_PATH: &str = "/sys/devices/system/cpu/online";
 
-/// Represents a CPU socket and the list of CPUs it contains.
-#[derive(Debug)]
+/// Represents a discovered CPU socket that has not been opened yet: its id
+/// and the CPUs it contains (see [`crate::perf::event::open_counters`] to
+/// open it).
 pub struct SocketInfo {
-    /// The ID of the socket.
-    pub socket_id: u32,
+    /// The id of the socket.
+    pub id: u32,
 
     /// List of CPU IDs associated with this socket.
-    pub cpus_id: Vec<u32>,
+    pub cpus: Vec<u32>,
 }
 
+/// Represents an opened CPU socket: its id, its `perf_event` group, and the
+/// RAPL domains discovered on it.
 pub struct Socket {
     /// The id of the socket.
     pub id: u32,
@@ -34,13 +37,14 @@ pub struct Socket {
     /// The group of counters to be able to manage them all at once.
     pub group: Group,
 
-    /// The RAPL domain supported by the hardware.
+    /// The RAPL domains supported by the hardware.
     pub domains: Vec<PerfRaplDomain>,
 }
 
 /// Discover the CPU socket topology of the system with an optional filter to discover only specific socket IDs.
 ///
-/// It returns a list containing each discovered socket and its CPUs.
+/// It returns a list containing each discovered socket and its CPUs, not yet
+/// opened (see [`crate::perf::event::open_counters`]).
 /// An error occurs if reading sysfs files fails or parsing fails.
 pub fn discover_socket_topology(
     sockets_to_discover: Option<&HashSet<u32>>,
@@ -89,14 +93,14 @@ pub(crate) fn discover_socket_topology_from_path(
 
     let mut socket_topology: Vec<SocketInfo> = sockets
         .into_iter()
-        .map(|(socket_id, mut cpus_id)| {
-            trace!("Found {cpus_id:?} cpus for socket {socket_id}");
-            cpus_id.sort_unstable();
-            SocketInfo { socket_id, cpus_id }
+        .map(|(id, mut cpus)| {
+            trace!("Found {cpus:?} cpus for socket {id}");
+            cpus.sort_unstable();
+            SocketInfo { id, cpus }
         })
         .collect();
 
-    socket_topology.sort_unstable_by_key(|s| s.socket_id);
+    socket_topology.sort_unstable_by_key(|s| s.id);
 
     debug!("Discovered {} socket(s)", socket_topology.len());
     Ok(socket_topology)
@@ -212,8 +216,8 @@ mod tests {
                 .unwrap();
 
         assert_eq!(topology.len(), 1);
-        assert_eq!(topology[0].socket_id, 0);
-        assert_eq!(topology[0].cpus_id, vec![0, 1, 2, 3]);
+        assert_eq!(topology[0].id, 0);
+        assert_eq!(topology[0].cpus, vec![0, 1, 2, 3]);
     }
 
     #[test]
@@ -225,10 +229,10 @@ mod tests {
                 .unwrap();
 
         assert_eq!(topology.len(), 2);
-        assert_eq!(topology[0].socket_id, 0);
-        assert_eq!(topology[0].cpus_id, vec![0, 1]);
-        assert_eq!(topology[1].socket_id, 1);
-        assert_eq!(topology[1].cpus_id, vec![2, 3]);
[…]
