### 2026-07-06 a7a4581
refactor(rapl): open rapl counters in pre_init instead of in construction
