### 2026-07-07 d6ad345
perf(perf): removed spawn blocking from read_snapshot because reading is faster than spawning blocking tasks
