### 2026-07-07 1e7a5ed
chore: fix clippy --all-targets warnings
