### 2026-07-07 27f4231
chore: fix clippy --all-targets warnings
