### 2026-07-07 b5a831d
chore: fix clippy --all-targets warnings
