### 2026-07-07 0584d81
perf(procfs): use per-process spawn_blocking to read each smap_rollups instead of blocking runtime
