+++ sources/procfs/src/lib.rs
@@ -3,7 +3,7 @@
 //! Provides memory and I/O metrics for a process and it's children and system-wide,
 //! by reading Linux's `/proc` filesystem via the `procfs` crate.
 
-use futures::StreamExt;
+use futures::{StreamExt, future::try_join_all, try_join};
 use joule_profiler_core::{
     sensor::{Sensor, Sensors},
     source::MetricReader,
@@ -11,9 +11,13 @@ use joule_profiler_core::{
     unit::{MetricUnit, Unit, UnitPrefix},
 };
 use log::{debug, trace};
-use std::{collections::HashSet, sync::Arc, time::Duration};
-use tokio::sync::Mutex;
+use std::{
+    collections::HashSet,
+    sync::{Arc, Mutex},
+    time::Duration,
+};
 use tokio::task::JoinHandle;
+use tokio::task::spawn_blocking;
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
@@ -46,14 +50,7 @@ type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
 /// Procfs-based metric source.
 ///
 /// Reads memory and I/O metrics from `/proc` for a target process and its
-/// entire child hierarchy, as well as system-wide memory statistics.
-///
-/// Metrics are accumulated as min/max over each measurement phase.
-/// A phase starts on [`MetricReader::measure`] and ends on [`MetricReader::retrieve`],
-/// which returns the counters and resets them for the next phase.
-///
-/// If a `poll_interval` is configured, a background task polls `/proc` continuously
-/// at that interval in addition to explicit [`MetricReader::measure`] calls.
+/// entire child hierarchy, as well as system-wide memory metrics.
 ///
 /// # Example
 ///
@@ -151,7 +148,10 @@ impl<B: Backend> Procfs<B> {
                 tokio::select! {
                     _ = ticker.next() => {
                         trace!("Polled procfs discovery task.");
-                        *detected_processes.lock().await = backend.collect_children(pid);
+                        let backend_clone = backend.clone();
+                        let children = spawn_blocking(move || backend_clone.collect_children(pid)).await?;
+                        let mut lock = detected_processes.lock().map_err(|_| ProcfsError::MutexPoisoned)?;
+                        *lock = children;
                     }
                     () = cancellation_token.cancelled() => {
                         debug!("procfs process discovery worker stopped.");
@@ -165,44 +165,65 @@ impl<B: Backend> Procfs<B> {
     }
 
     async fn measure_and_update_pids(
-        backend: &B,
+        backend: &Arc<B>,
         counters: &Arc<Mutex<Counters>>,
         detected_processes: &Arc<Mutex<HashSet<i32>>>,
     ) -> Result<()> {
+        let pids = detected_processes
+            .lock()
+            .map_err(|_| ProcfsError::MutexPoisoned)?
+            .clone();
+
+        let proc_reads: Vec<_> = pids
+            .into_iter()
+            .map(|pid| {
+                let backend = Arc::clone(backend);
+                spawn_blocking(move || {
+                    let result = backend.read_proc(pid);
+                    (pid, result)
+                })
+            })
+            .collect();
+
+        let backend_global = Arc::clone(backend);
+        let global_read = spawn_blocking(move || backend_global.measure_global());
+
+        let (proc_results, global_result) = try_join!(try_join_all(proc_reads), global_read)?;
+        let global = global_result?;
+
         let mut present_pids = HashSet::new();
-        let pids = detected_processes.lock().await.clone();
         let mut snapshot = ProcSnapshot::default();
         let mut pids_updated = false;
 
-        for pid in pids {
-            match backend.read_proc(pid, &mut snapshot) {
-                Ok(()) => {
[…]
