### 2026-07-07 0a50ba6
perf(procfs): use per-process spawn_blocking to read each smap_rollups instead of blocking runtime
