### 2026-07-07 76bfd56
perf(nvml): use spawn blocking for each gpu to avoid blocking the async runtime
