### 2026-07-07 73e98b8
perf(nvml): use spawn blocking for each gpu to avoid blocking the async runtime
