### 2026-07-07 07c2e54
perf(nvml): use spawn blocking for each gpu to avoid blocking the async runtime
