+++ sources/nvml/src/lib.rs
@@ -2,10 +2,14 @@
 //!
 //! This module provides energy consumption, VRAM usage and GPU utilization metrics forNVIDIA GPUs using the NVML library.
 
-use std::{collections::HashMap, sync::Arc, time::Duration};
+use std::{
+    collections::HashMap,
+    sync::{Arc, Mutex},
+    time::Duration,
+};
 
 use bitflags::bitflags;
-use futures::StreamExt;
+use futures::{StreamExt, TryFutureExt, future::try_join_all, try_join};
 use joule_profiler_core::{
     sensor::{Sensor, Sensors},
     source::MetricReader,
@@ -13,7 +17,7 @@ use joule_profiler_core::{
     unit::{MetricUnit, Unit, UnitPrefix},
 };
 use log::{debug, trace};
-use tokio::{sync::Mutex, task::JoinHandle};
+use tokio::task::{JoinHandle, spawn_blocking};
 use tokio_timerfd::Interval;
 use tokio_util::sync::CancellationToken;
 
@@ -63,7 +67,7 @@ type Result<T> = std::result::Result<T, NvmlError>;
 /// Polling task handle and its cancellation token.
 type WorkerHandle = (CancellationToken, JoinHandle<Result<()>>);
 
-#[derive(Debug, Clone)]
+#[derive(Debug, Clone, Copy)]
 pub struct Device {
     /// The index of the device.
     index: u32,
@@ -142,7 +146,11 @@ impl<H: NvmlHardware> Nvml<H> {
         Ok((cancellation_token_clone, handle))
     }
 
-    /// Reads the power and vram counters for each processors and updates the current counters.
+    /// Reads the power, vram and utilization counters for each processors and updates the current counters.
+    ///
+    /// Each device is read on its own blocking thread, in parallel, since NVML calls
+    /// perform blocking driver ioctls. All the queries applicable to a device (vram,
+    /// utilization, power) are grouped in that single blocking task.
     async fn read_polled_counters(
         hardware: &Arc<H>,
         processors: &Arc<Vec<Device>>,
@@ -150,42 +158,43 @@ impl<H: NvmlHardware> Nvml<H> {
         vram_counters: &Arc<Mutex<HashMap<u32, VramCounter>>>,
         utilization_counters: &Arc<Mutex<HashMap<u32, UtilizationCounter>>>,
     ) -> Result<()> {
-        let mut vram_updates = Vec::new();
-        let mut utilization_updates = Vec::new();
-        let mut power_updates = Vec::new();
+        let tasks = processors.iter().copied().map(|device| {
+            let hardware = hardware.clone();
+            spawn_blocking(move || {
+                let vram = device
+                    .support
+                    .contains(DeviceSupport::Vram)
+                    .then(|| hardware.get_vram_usage(device));
+                let utilization = device
+                    .support
+                    .contains(DeviceSupport::Utilization)
+                    .then(|| hardware.get_utilization(device));
+                let power = device
+                    .support
+                    .contains(DeviceSupport::Power)
+                    .then(|| hardware.get_power(device));
+                (device.index, vram, utilization, power)
+            })
+        });
 
-        for device in processors.iter() {
-            if device.support.contains(DeviceSupport::Vram) {
-                vram_updates.push((device.index, hardware.get_vram_usage(device)?));
+        for (index, vram, utilization, power) in try_join_all(tasks).await? {
+            if let Some(vram) = vram {
+                let mut lock = vram_counters.lock().map_err(|_| NvmlError::MutexPoisoned)?;
+                lock.entry(index).or_default().update(vram?);
             }
 
-            if device.support.contains(DeviceSupport::Utilization) {
-                utilization_updates.push((device.index, hardware.get_utilization(device)?));
+            if let Some(utilization) = utilization {
+                let mut lock = utilization_counters
[…]
