### 2026-07-07 059b0f1
chore: fixed warnings and tests
