### 2026-07-07 5e3dd12
chore: fixed warnings and tests
