### 2026-07-07 dc2054b
perf: changed all tokio::sync::Mutex to std::sync::Mutex because there's no contention on Mutexes
