### 2026-07-07 1391afa
perf: changed all tokio::sync::Mutex to std::sync::Mutex because there's no contention on Mutexes
