### 2026-07-07 fdb653f
perf: changed all tokio::sync::Mutex to std::sync::Mutex because there's no contention on Mutexes
