+++ examples/library_usage/README.md
@@ -4,5 +4,5 @@ This crate provides an example of usage of Joule Profiler library.
 
 Run the example:
 ```bash
-cargo run --bin library_usage_example
+cargo run -p library_usage_example
 ```
