### 2026-07-07 edd4f5b
chore: some minor fixes in RAPL and cgroup
