### 2026-07-07 c74a8e5
docs: update RAPL new functions doc and cli doc
