### 2026-07-07 514cb55
docs: update RAPL new functions doc and cli doc
