### 2026-07-07 0e8fa88
docs: update RAPL new functions doc and cli doc
