### 2026-07-07 94adcc6
docs: update config table documentation
