### 2026-07-07 35f7241
fix: config file with empty sources serialized as empty
