+++ cli/src/bin/main.rs
@@ -52,12 +52,16 @@ async fn main() -> Result<()> {
             &mut profiler,
             &mut config_table,
             |config| {
-                config.rapl_path = cli.rapl_path.take();
+                if let Some(rapl_path) = cli.rapl_path.take() {
+                    config.rapl_path = Some(rapl_path);
+                }
                 if let Some(sockets_spec) = &cli.sockets {
                     config.sockets_spec = Some(parse_sockets_spec(sockets_spec));
                 }
-                if let ProfilerCommand::Profile(profile_args) = &cli.command {
-                    config.poll_interval = profile_args.rapl_polling;
+                if let ProfilerCommand::Profile(profile_args) = &cli.command
+                    && let Some(rapl_polling) = profile_args.rapl_polling
+                {
+                    config.poll_interval = Some(rapl_polling);
                 }
             },
         ),
