### 2026-07-07 5ffa3a8
fix: config file with empty sources serialized as empty
