### 2026-07-07 9f8fd84
docs(rapl): removed module former doc
