### 2026-07-07 3069156
docs: update readmes, cli and cgroup doc
