### 2026-07-07 9a99828
docs: update readmes, cli and cgroup doc
