### 2026-07-07 94961b9
docs: update readmes, cli and cgroup doc
