### 2026-07-07 3976ecd
fix: sources configurations serialization fixed
