+++ sources/cgroup/src/config.rs
@@ -8,14 +8,22 @@ fn default_poll_interval() -> Duration {
     DEFAULT_POLL_INTERVAL
 }
 
+fn default_true() -> bool {
+    true
+}
+
+fn default_cgroup_name() -> String {
+    format!("joule-profiler-{}", std::process::id())
+}
+
 /// Configuration for the cgroup metric source.
 #[derive(Debug, Clone, Deserialize)]
-#[serde(default)]
 pub struct CgroupConfig {
     /// Path to cgroup v2 hierarchy (usually `/sys/fs/cgroup`).
     pub cgroup_root: Option<PathBuf>,
 
     /// Name of the created cgroup for the monitored process.
+    #[serde(default = "default_cgroup_name")]
     pub cgroup_name: String,
 
     /// Background polling interval.
@@ -23,9 +31,11 @@ pub struct CgroupConfig {
     pub poll_interval: Duration,
 
     /// Whether the source must attach the process pid to the cgroup or not.
+    #[serde(default = "default_true")]
     pub attach_pid: bool,
 
     /// Whether the cgroup is already created or if the source must create it itself.
+    #[serde(default = "default_true")]
     pub create_cgroup: bool,
 }
 
@@ -33,7 +43,7 @@ impl Default for CgroupConfig {
     fn default() -> Self {
         Self {
             cgroup_root: None,
-            cgroup_name: format!("joule-profiler-{}", std::process::id()),
+            cgroup_name: default_cgroup_name(),
             poll_interval: DEFAULT_POLL_INTERVAL,
             attach_pid: true,
             create_cgroup: true,
