+++ cli/src/config/mod.rs
@@ -24,7 +24,7 @@ pub struct ProfilerConfig {
     pub stdout_file: Option<String>,
 
     /// Regex used to detect phase tokens in program output.
-    #[serde(default)]
+    #[serde(default = "default_token_pattern")]
     pub token_pattern: String,
 
     /// Executes the profiled command with root privileges if true and Joule Profiler is launched as root.
@@ -39,7 +39,7 @@ pub struct ProfilerConfig {
     pub output_format: OutputFormat,
 
     /// Duration before aborting sources initialization. (default: 1s)
-    #[serde(default, with = "humantime_serde")]
+    #[serde(default = "default_timeout", with = "humantime_serde")]
     pub init_timeout: Duration,
 
     #[serde(default)]
