+++ sources/procfs/src/config.rs
@@ -17,7 +17,6 @@ fn default_process_detection_poll_interval() -> Duration {
 
 /// Configuration for the `Procfs` metric source.
 #[derive(Debug, Clone, Deserialize)]
-#[serde(default)]
 pub struct ProcfsConfig {
     /// Polling interval.
     #[serde(default = "default_poll_interval", with = "humantime_serde")]
@@ -31,9 +30,11 @@ pub struct ProcfsConfig {
     pub process_detection_poll_interval: Duration,
 
     /// Memory unit for process metrics.
+    #[serde(default)]
     pub proc_memory_unit: MemoryUnit,
 
     /// Memory unit for global metrics.
+    #[serde(default)]
     pub global_memory_unit: MemoryUnit,
 }
 
