### 2026-07-07 864a311
fix: sources configurations serialization fixed
