### 2026-07-07 ac145c4
feat(perf): added cgroup_root to configure the cgroup base path
