### 2026-07-07 3287538
feat(perf): added cgroup_root to configure the cgroup base path
