### 2026-07-07 8c9e1a6
feat(perf): added cgroup_root to configure the cgroup base path
