### 2026-07-08 baad7c7
feat(perf): added cpu specification to not open counters on each core when using cgroup
