+++ sources/perf_event/src/config.rs
@@ -22,4 +22,10 @@ pub struct PerfConfig {
 
     /// The root path of the cgroup if cgroup name is provided.
     pub cgroup_root: Option<PathBuf>,
+
+    /// If the source uses the cgroup, then you can specify on
+    /// which CPU cores you want to open the counters.
+    /// It can be useful if your program has a configured CPU affinity (e.g., taskset)
+    /// Not opening counters on all cores can improve the performance of the source.
+    pub cpu_spec: Option<HashSet<u32>>,
 }
