### 2026-07-08 58091a5
feat(perf): added cpu specification to not open counters on each core when using cgroup
