### 2026-07-08 395075e
fix: update former phases to profile to launch profiling in examples
