### 2026-07-09 c345e5b
perf: spawn phase detection mechanism in its own thread instead of blocking async runtime
