### 2026-07-09 283040d
perf: spawn phase detection mechanism in its own thread instead of blocking async runtime
