+++ core/src/profiler/mod.rs
@@ -8,11 +8,13 @@ use log::{debug, info, trace};
 use regex::Regex;
 use std::io::BufWriter;
 use std::os::unix::process::CommandExt;
-use std::process::{Child, Command};
+use std::process::{Child, ChildStdout, Command};
+use std::thread::JoinHandle;
 use std::{
     io::{BufRead, BufReader, ErrorKind, Write},
     process::{self, Stdio},
 };
+use tokio::sync::mpsc::{Receiver, Sender};
 
 pub mod error;
 
@@ -28,6 +30,8 @@ use crate::util::time::get_timestamp_micros;
 pub use error::JouleProfilerError;
 
 pub mod types;
+
+const PHASE_CHANNEL_SIZE: usize = 2;
 
 /// Orchestrates program profiling and metric collection.
 ///
@@ -118,6 +122,9 @@ impl JouleProfiler {
             return Err(JouleProfilerError::NoSourceConfigured);
         }
         let mut orchestrator = Orchestrator::new(sources);
+
+        let (tx, rx) = tokio::sync::mpsc::channel::<PhaseInfo>(PHASE_CHANNEL_SIZE);
+
         orchestrator.pre_init().await?;
 
         debug!("Spawning command: {:?}", config.cmd);
@@ -125,15 +132,24 @@ impl JouleProfiler {
         let pid = child.id().cast_signed();
 
         pause_prosess(pid)?;
+
+        let child_stdout = child
+            .stdout
+            .take()
+            .ok_or(JouleProfilerError::StdOutCaptureFail)?;
+        let reader_thread =
+            spawn_reader_thread(child_stdout, regex, config.stdout_file.clone(), tx)?;
+
         orchestrator.init(pid, config.init_timeout).await?;
         orchestrator.run();
 
         info!("Starting measurements");
         let (command_duration_ms, timestamp, exit_code, detected_phases) = self
-            .measure_phases(&mut orchestrator, config, &regex, &mut child, pid)
+            .measure_phases(&mut orchestrator, &mut child, rx, pid)
             .await?;
 
         let (sources_results, sources) = orchestrator.finalize().await?;
+        join_reader_thread(reader_thread)?;
         self.sources = sources;
 
         let mut phases: Vec<_> = detected_phases
@@ -160,7 +176,7 @@ impl JouleProfiler {
         if phases.is_empty()
             && let Some(end_phase) = sources_results.phases.into_iter().last()
         {
-            let phase = Phase {
+            phases.push(Phase {
                 index: 0,
                 metrics: end_phase.metrics,
                 start_token: PhaseToken::Start,
@@ -169,8 +185,7 @@ impl JouleProfiler {
                 duration_ms: command_duration_ms,
                 start_token_line: None,
                 end_token_line: None,
-            };
-            phases.push(phase);
+            });
         }
 
         debug!("Collected {} sensor phase(s)", phases.len());
@@ -197,40 +212,24 @@ impl JouleProfiler {
     async fn measure_phases(
         &mut self,
         orchestrator: &mut Orchestrator,
-        config: &ProfileConfig,
-        regex: &Regex,
         child: &mut Child,
+        mut rx: Receiver<PhaseInfo>,
         pid: i32,
     ) -> Result<MeasurePhasesReturnType> {
-        let mut sink = create_output_sink(config.stdout_file.as_ref())?;
-
-        let child_stdout = child
-            .stdout
-            .take()
-            .ok_or(JouleProfilerError::StdOutCaptureFail)?;
-
-        let reader = BufReader::new(child_stdout);
         let mut detected_phases = Vec::with_capacity(2);
 
         let begin_timestamp = get_timestamp_micros();
         trace!("Begin timestamp: {begin_timestamp}");
 
         orchestrator.measure().await?;
-
         resume_process(pid)?;
-
[…]
