### 2026-07-10 899d6d1
Improve clarity and detail in the software design and impact sections
Revised the description of Joule Profiler for clarity and detail. Enhanced explanations of the tool's architecture, measurement validation, and phase detection precision.

closes #38
