### 2026-07-10 481e242
perf: spawn reader task in a separate thread to not block the async runtime
