### 2026-07-10 8890a77
perf: spawn reader task in a separate thread to not block the async runtime
