+++ core/src/profiler/mod.rs
@@ -14,14 +14,13 @@ use std::{
     io::{BufRead, BufReader, ErrorKind, Write},
     process::{self, Stdio},
 };
-use tokio::sync::mpsc::{Receiver, Sender};
 
 pub mod error;
 
 use crate::config::ProfileConfig;
 use crate::orchestrator::Orchestrator;
 use crate::phase::{PhaseInfo, PhaseToken};
-use crate::profiler::types::{MeasurePhasesReturnType, Phase, ProfilerResults, Result};
+use crate::profiler::types::{Phase, ProfilerResults, Result};
 use crate::sensor::{Sensor, Sensors};
 use crate::source::{MetricReader, MetricSource, MetricSourceError};
 use crate::util::fs::create_file_with_user_permissions;
@@ -31,7 +30,11 @@ pub use error::JouleProfilerError;
 
 pub mod types;
 
-const PHASE_CHANNEL_SIZE: usize = 2;
+/// Phases and begin/end timestamps collected by the reader thread.
+type MeasureData = (Vec<PhaseInfo>, u128, u128);
+
+/// Reader thread handle: gives the orchestrator back with the collected data.
+type ReaderThreadHandle = JoinHandle<(Orchestrator, Result<MeasureData>)>;
 
 /// Orchestrates program profiling and metric collection.
 ///
@@ -123,8 +126,6 @@ impl JouleProfiler {
         }
         let mut orchestrator = Orchestrator::new(sources);
 
-        let (tx, rx) = tokio::sync::mpsc::channel::<PhaseInfo>(PHASE_CHANNEL_SIZE);
-
         orchestrator.pre_init().await?;
 
         debug!("Spawning command: {:?}", config.cmd);
@@ -137,19 +138,40 @@ impl JouleProfiler {
             .stdout
             .take()
             .ok_or(JouleProfilerError::StdOutCaptureFail)?;
-        let reader_thread =
-            spawn_reader_thread(child_stdout, regex, config.stdout_file.clone(), tx)?;
 
         orchestrator.init(pid, config.init_timeout).await?;
         orchestrator.run();
 
         info!("Starting measurements");
-        let (command_duration_ms, timestamp, exit_code, detected_phases) = self
-            .measure_phases(&mut orchestrator, &mut child, rx, pid)
-            .await?;
+        let reader_thread = spawn_reader_thread(
+            orchestrator,
+            child_stdout,
+            regex,
+            config.stdout_file.clone(),
+            pid,
+        )?;
+
+        let (mut orchestrator, measure_result) =
+            tokio::task::spawn_blocking(move || join_reader_thread(reader_thread))
+                .await
+                .map_err(|_| JouleProfilerError::ReaderThreadPanicked)??;
+
+        let (detected_phases, begin_timestamp, end_timestamp) = match measure_result {
+            Ok(data) => data,
+            Err(err) => {
+                return Err(match orchestrator.finalize().await {
+                    Err(source_err) => source_err.into(),
+                    Ok(_) => err,
+                })?;
+            }
+        };
+
+        let command_duration_ms = (end_timestamp - begin_timestamp) / 1000;
+        let timestamp = begin_timestamp;
+        let exit_code = wait_for_child_exit(&mut child)?;
+        info!("Command finished: duration={command_duration_ms} ms exit_code={exit_code}");
 
         let (sources_results, sources) = orchestrator.finalize().await?;
-        join_reader_thread(reader_thread)?;
         self.sources = sources;
 
         let mut phases: Vec<_> = detected_phases
@@ -196,88 +218,80 @@ impl JouleProfiler {
             phases,
         })
     }
-
-    /// Profile the already spawned and paused command, separating its execution into phases
[…]
