### 2026-07-10 80cf039
refactor: profiler.rs improvement
