### 2026-07-10 35834d3
perf: removed spawn blocking for waiting and put oneshot sender
