### 2026-07-10 48324a3
perf: removed spawn blocking for waiting and put oneshot sender
