### 2026-07-10 9f0e2a8
chore: moved new types into types.rs
