### 2026-07-14 97b8cf9
Update paper.md to specify cloud compatibility
Clarify that Joule Profiler can run on bare-metal machines and suggests checking measurement source availability in virtual environments.
