+++ joss/paper/paper.md
@@ -68,7 +68,7 @@ Joule Profiler has been designed to be modular. These modules collect energy and
 
 ![](images/archi.png){ width=90% }
 
-Internally, the tool is structured into layers. The core layer handles the main logic: detecting phases, aggregating metrics, and coordinating the measurement sources. Each source runs as an asynchronous task, enabling parallel data collection and maintaining temporal precision. The *Command-Line Interface* (CLI) layer manages user interaction, parses configuration options, and displays results. A source abstraction layer encapsulates each hardware backend, such as RAPL, NVML, or performance counters, in a separate module. This separation eases the integration of new sources in the future without affecting the rest of the system. This design allows Joule Profiler to run on a large diversity of machines based on Intel and AMD processors.
+Internally, the tool is structured into layers. The core layer handles the main logic: detecting phases, aggregating metrics, and coordinating the measurement sources. Each source runs as an asynchronous task, enabling parallel data collection and maintaining temporal precision. The *Command-Line Interface* (CLI) layer manages user interaction, parses configuration options, and displays results. A source abstraction layer encapsulates each hardware backend, such as RAPL, NVML, or performance counters, in a separate module. This separation eases the integration of new sources in the future without affecting the rest of the system. This design allows Joule Profiler to run on a large diversity of bare-metal machines based on Intel and AMD processors. While Joule Profiler can also be used in virtual environments, users are encouraged to check the availability of measurement sources.
 
 # Software assessment
 
