### 2026-07-19 0465e5d
fix(perf-event): include host and guest for virtualization when opening counters
