### 2026-08-09 09997f9
Enhance paper content and improve clarity
Addressing the JOSS review from @ygrange to include comments about the required usage of bare-metal environments.
