+++ joss/paper/paper.md
@@ -6,6 +6,7 @@ tags:
   - energy measurement
   - profiling
   - Intel RAPL
+  - NVIDIA NVML
   - Linux
   - Rust
   - green computing
@@ -31,11 +32,11 @@ repository: https://github.com/joule-profiler/joule-profiler
 
 # Summary
 
-Joule Profiler is a lightweight Linux command-line tool for profiling a program’s energy consumption with minimal instrumentation overhead. It enables users to break execution into user-defined phases (e.g., data loading, computation) and attribute energy use to each phase. The tool detects phase triggers in program output and automatically queries sources such as Intel RAPL (CPUs) or NVML (GPUs) to report system-wide energy consumption.
+Joule Profiler is a lightweight Linux command-line tool for profiling a program’s energy consumption on bare-metal computing environments with minimal instrumentation overhead. It enables users to break a program's execution into user-defined phases (e.g., data loading, computation) and attribute energy use to each phase. The tool detects phase triggers from program output and automatically queries available hardware sources such as Intel RAPL (CPUs) or NVML (GPUs) to report system-wide energy consumption.
 
 # Statement of need
 
-Energy use in computing is a growing concern across research and industry. Software running in clouds, data centres, and edge devices contributes significantly to global energy consumption. Improving efficiency requires tools that measure energy during execution. Hardware counters on modern CPUs and GPUs (e.g., Intel RAPL) enable software-based energy measurement without external devices.
+Energy use in computing is a growing concern across research and industry. Software running in clouds, data centers, and edge devices contributes significantly to global energy consumption. Improving efficiency requires tools that measure energy during execution. Hardware counters on modern CPUs and GPUs (e.g., Intel RAPL) enable software-based energy measurement without external devices.
 
 Researchers and developers need simple tools to measure the energy use of code segments without complex infrastructure. Joule Profiler addresses this with phase-based profiling that integrates easily into workflows.
 
@@ -43,7 +44,7 @@ Researchers and developers need simple tools to measure the energy use of code s
 
 Existing tools like PowerAPI [@powerapi], Alumet [@alumet], Scaphandre [@scaphandre], and EnergiBridge [@sallou_energibridge_2024] monitor energy using these counters, often focusing on distributed and system-level observability. JouleIt [@jouleit], which inspired this work, demonstrated a light wrapper approach but lacked phase decomposition, GPU support, and modularity.
 
-These solutions are suited for system-level monitoring, not fine-grained analysis of program phases. Joule Profiler, in contrast, is designed for lightweight, single-invocation use, enabling energy attribution to specific phases within programs.
+These solutions are suited for system-level monitoring, not fine-grained analysis of program phases. Joule Profiler, in contrast, is designed for lightweight, single-invocation use, enabling energy attribution to specific program phases.
 
 # Software design
 
@@ -68,7 +69,7 @@ Joule Profiler has been designed to be modular. These modules collect energy and
 
 ![](images/archi.png){ width=90% }
 
[…]
