### 2026-08-12 ca2231b
Apply editorial corrections to paper.md
