### 2026-08-12 c74e800
feat(cgroups): added tests to cgroup source and improve error handling
