+++ sources/cgroup/src/cgroup.rs
@@ -1,9 +1,4 @@
-//! cgroup v2 management utilities.
-//!
-//! This module provides:
-//! - child cgroup creation and cleanup;
-//! - process attachment to cgroups;
-//! - CPU, memory, and I/O statistics collection.
+//! Cgroup v2 management utilities.
 
 use std::{
     fs,
@@ -17,15 +12,17 @@ use crate::{
     Result,
     error::CgroupError,
     snapshot::{CpuSnapshot, IoSnapshot, MemorySnapshot},
-    util::{read_flat_keyed_file, read_io_stat, read_u64_opt},
+    util::{is_cgroup_dir, read_flat_keyed_file, read_io_stat, read_u64_opt},
 };
 
 /// Interface for reading cgroup statistics.
-///
-/// `Default` and `Clone` let [`RootCgroup::build`] construct a backend from
-/// scratch and share that same instance with its child cgroup, for any
-/// implementation of this trait (not just [`SysFsBackend`]).
 pub trait CgroupBackend: Send + Sync + Clone + Default + 'static {
+    /// Checks that `path` is a cgroup the backend can read, `root` being the
+    /// hierarchy it belongs to.
+    fn verify(&self, _path: &Path, _root: &Path) -> Result<()> {
+        Ok(())
+    }
+
     /// Initializes the backend.
     fn create(&self, path: &Path) -> Result<()>;
 
@@ -54,7 +51,11 @@ impl SysFsBackend {
     fn pids(path: &Path) -> Result<Vec<i32>> {
         debug!("Retrieving cgroup `{}` PIDs.", path.display());
         let path = path.join("cgroup.procs");
-        let content = fs::read_to_string(&path)?;
+        let content = match fs::read_to_string(&path) {
+            Ok(content) => content,
+            Err(err) if err.kind() == ErrorKind::NotFound => return Ok(Vec::new()),
+            Err(err) => return Err(CgroupError::io(&path, err)),
+        };
         Ok(content
             .lines()
             .filter_map(|l| l.trim().parse::<i32>().ok())
@@ -63,6 +64,20 @@ impl SysFsBackend {
 }
 
 impl CgroupBackend for SysFsBackend {
+    /// Checks that the hierarchy is mounted and that the cgroup exists in it.
+    fn verify(&self, path: &Path, root: &Path) -> Result<()> {
+        if !is_cgroup_dir(root) {
+            return Err(CgroupError::NotAvailable(root.to_path_buf()));
+        }
+        if !path.exists() {
+            return Err(CgroupError::NotFound(path.to_path_buf()));
+        }
+        if !is_cgroup_dir(path) {
+            return Err(CgroupError::NotAvailable(path.to_path_buf()));
+        }
+        Ok(())
+    }
+
     /// Creates the cgroup directory.
     fn create(&self, path: &Path) -> Result<()> {
         debug!("Initializing cgroup at \"{}\"", path.display());
@@ -70,7 +85,7 @@ impl CgroupBackend for SysFsBackend {
             ErrorKind::PermissionDenied => {
                 CgroupError::PermissionDenied("creating a cgroup requires root privileges.")
             }
-            _ => err.into(),
+            _ => CgroupError::io(path, err),
         })?;
         Ok(())
     }
@@ -82,6 +97,12 @@ impl CgroupBackend for SysFsBackend {
             ErrorKind::PermissionDenied => CgroupError::PermissionDenied(
                 "attaching a process to this cgroup requires root privileges.",
             ),
+            // The kernel reports a missing `cgroup.procs` and a process that
+            // died before being attached the same way, so only the file tells
+            // them apart.
[…]
