+++ sources/cgroup/src/util.rs
@@ -1,15 +1,28 @@
-use std::{collections::HashMap, fs, path::Path};
+use std::{collections::HashMap, fs, io::ErrorKind, path::Path};
 
 use crate::{Result, error::CgroupError};
 
+/// Reads a cgroup file, tagging any I/O failure with the file it happened on.
+pub fn read_file(path: &Path) -> Result<String> {
+    fs::read_to_string(path).map_err(|err| CgroupError::io(path, err))
+}
+
+/// Reads a cgroup file, returning `None` when it does not exist.
+///
+/// Many interface files are only exposed by some kernel versions or some
+/// controller configurations, so a missing file is a valid answer here while
+/// any other I/O failure stays an error.
+fn read_file_opt(path: &Path) -> Result<Option<String>> {
+    match fs::read_to_string(path) {
+        Ok(raw) => Ok(Some(raw)),
+        Err(err) if err.kind() == ErrorKind::NotFound => Ok(None),
+        Err(err) => Err(CgroupError::io(path, err)),
+    }
+}
+
 /// Reads a `u64` value from a file, returns an error if unable to parse.
-fn read_u64(path: &Path) -> Result<u64> {
-    let raw = fs::read_to_string(path).map_err(|e| CgroupError::IoPath {
-        path: path.to_path_buf(),
-        source: e,
-    })?;
-    let trimmed = raw.trim();
-    trimmed.parse::<u64>().map_err(|err| CgroupError::Parse {
+fn parse_u64(path: &Path, raw: &str) -> Result<u64> {
+    raw.trim().parse::<u64>().map_err(|err| CgroupError::Parse {
         path: path.to_path_buf(),
         source: err,
     })
@@ -19,11 +32,28 @@ fn read_u64(path: &Path) -> Result<u64> {
 ///
 /// Return None if the file does not exist, else try to parse the file.
 pub fn read_u64_opt(path: &Path) -> Result<Option<u64>> {
-    if fs::exists(path)? {
-        Ok(Some(read_u64(path)?))
-    } else {
-        Ok(None)
-    }
+    read_file_opt(path)?
+        .map(|raw| parse_u64(path, &raw))
+        .transpose()
+}
+
+/// Returns whether `path` is a directory of a cgroup v2 hierarchy.
+///
+/// Every cgroup v2 directory, root or not, exposes a read-only
+/// `cgroup.controllers` file, which a plain directory never has.
+pub fn is_cgroup_dir(path: &Path) -> bool {
+    path.join("cgroup.controllers").is_file()
+}
+
+/// Reads a space separated list of controller names.
+///
+/// Used for `cgroup.controllers` (controllers the hierarchy knows about) and
+/// `cgroup.subtree_control` (controllers a cgroup enables for its children).
+pub fn read_controllers(path: &Path) -> Result<Vec<String>> {
+    Ok(read_file(path)?
+        .split_whitespace()
+        .map(str::to_owned)
+        .collect())
 }
 
 /// Reads a flat key-value file.
@@ -37,10 +67,7 @@ pub fn read_u64_opt(path: &Path) -> Result<Option<u64>> {
 /// Returns a map of parsed metrics.
 /// Lines that fail to parse are ignored.
 pub fn read_flat_keyed_file(path: &Path) -> Result<HashMap<String, u64>> {
-    let raw = fs::read_to_string(path).map_err(|err| CgroupError::IoPath {
-        path: path.to_owned(),
-        source: err,
-    })?;
+    let raw = read_file(path)?;
 
     let map = raw
         .lines()
[…]
