### 2026-08-12 293db16
feat(cgroups): added tests to cgroup source and improve error handling
