+++ sources/cgroup/src/error.rs
@@ -7,12 +7,24 @@ use std::{
 use thiserror::Error;
 use tokio::task::JoinError;
 
+use crate::util::{is_cgroup_dir, read_controllers};
+
 /// Main error type for all cgroup-related operations.
 #[derive(Debug, Error)]
 pub enum CgroupError {
     /// cgroup v2 is not available or not mounted as a unified hierarchy.
-    #[error("Cgroup v2 not available at `{0}`. Check if the unified hierarchy is mounted.")]
-    NotAvailable(String),
+    #[error(
+        "`{0}` is not a cgroup v2 directory. Check that the unified hierarchy is mounted \
+        (`mount -t cgroup2`) and that `cgroup_root` points to it."
+    )]
+    NotAvailable(PathBuf),
+
+    /// The cgroup directory does not exist.
+    #[error(
+        "Cgroup `{0}` does not exist. Create it beforehand or let the source do it \
+        with `create_cgroup = true`."
+    )]
+    NotFound(PathBuf),
 
     /// Attempted to create a cgroup that already exists.
     #[error("Cgroup with name \"{0}\" already created.")]
@@ -37,17 +49,19 @@ pub enum CgroupError {
         path: PathBuf,
     },
 
+    /// A cgroup controller is not available in the hierarchy at all.
+    #[error(
+        "Cgroup controller `{controller}` is not available in `{path}`. \
+        The kernel was built without it or it is not delegated to this hierarchy."
+    )]
+    ControllerNotAvailable {
+        controller: &'static str,
+        path: PathBuf,
+    },
+
     /// A metric expected to always exist in kernel stats was missing.
     #[error("Missing always present metric `{0}`")]
     MissingAlwaysPresentMetric(&'static str),
-
-    /// Generic I/O error.
-    #[error("I/O error")]
-    Io(
-        #[from]
-        #[source]
-        std::io::Error,
-    ),
 
     /// I/O error tied to a specific file path.
     #[error("I/O error on `{path}`: {source}")]
@@ -56,6 +70,13 @@ pub enum CgroupError {
         #[source]
         source: std::io::Error,
     },
+
+    /// Failed to create the timer driving the background memory polling.
+    ///
+    /// Not built from [`From`] on purpose: every I/O error of the source must
+    /// name the file or the operation it comes from.
+    #[error("Failed to create the cgroup polling timer: {0}")]
+    Timer(#[source] std::io::Error),
 
     /// Failed to parse a numeric value from a cgroup file.
     #[error("Failed to parse `{path}`: {source}")]
@@ -81,16 +102,179 @@ pub enum CgroupError {
 }
 
 impl CgroupError {
+    /// Builds an [`CgroupError::IoPath`] carrying the file the operation failed on.
+    pub(crate) fn io(path: &Path, source: std::io::Error) -> Self {
+        CgroupError::IoPath {
+            path: path.to_path_buf(),
+            source,
+        }
+    }
+
+    /// Refines a missing controller interface file into its actual cause.
+    ///
+    /// A `memory.stat` (or `cpu.stat`, `io.stat`) that cannot be found has
+    /// several plausible causes, which are checked from the most general to the
+    /// most specific so the reported one is always the first thing to fix:
+    /// the cgroup does not exist, it is not part of a cgroup v2 hierarchy, the
+    /// controller is unknown to that hierarchy, or the parent cgroup does not
+    /// delegate it through `cgroup.subtree_control`.
+    ///
+    /// Any error that is not a missing file, and any cause that cannot be
+    /// established, is returned untouched.
     pub(crate) fn into_controller_error(self, controller: &'static str, cgroup: &Path) -> Self {
[…]
