### 2026-08-12 8b1cf62
docs: update examples command line
