### 2026-08-12 f722d1d
docs: update examples command line
