+++ examples/programs/mandelbrot_nvidia.cu
@@ -1,7 +1,7 @@
 /*
   Source: https://github.com/canonizer/mandelbrot-dyn
 
-  Compile: nvcc -O3 -Xcompiler -fopenmp mandelbrot_nvidia.cu -o mandelbrot_nvidia
+  Compile: nvcc -O3 -Xcompiler -fopenmp examples/programs/mandelbrot_nvidia.cu -o mandelbrot_nvidia
   Usage: joule-profiler --gpu profile -- ./mandelbrot_nvidia {n}
 */
 
