### 2026-08-12 c04bead
docs: update examples command line
