### 2026-08-12 334f6df
chore: add warning if the cgroup is already created and the configuration specify to create it
