### 2026-08-12 d54c87e
docs: update readme examples
