### 2026-08-12 a9c3ac3
chore: update cgroup error doc
