### 2026-08-12 4715491
docs: update readme with examples directory doc
