### 2026-08-12 427d704
feat(cli): override configuration with -D flag
