+++ cli/src/config/mod.rs
@@ -1,9 +1,11 @@
-use std::{collections::HashMap, time::Duration};
+use std::{collections::HashMap, fs, path::Path, time::Duration};
 
+use anyhow::{Context, Result};
 use serde::Deserialize;
 
-use crate::{RaplBackend, output::formats::OutputFormat};
+use crate::{RaplBackend, config::overrides::ConfigOverride, output::formats::OutputFormat};
 
+pub mod overrides;
 pub mod source;
 pub mod table;
 
@@ -18,7 +20,10 @@ fn default_token_pattern() -> String {
     DEFAULT_TOKEN_PATTERN.to_owned()
 }
 
+/// Unknown keys are rejected: a misspelled one would otherwise be dropped
+/// without a trace, which is easy to miss on a `-D` override.
 #[derive(Debug, Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct ProfilerConfig {
     /// Optional file to redirect the profiled program stdout.
     pub stdout_file: Option<String>,
@@ -61,6 +66,7 @@ impl Default for ProfilerConfig {
 }
 
 #[derive(Debug, Default, Deserialize)]
+#[serde(deny_unknown_fields)]
 pub struct GlobalConfig {
     /// The global configuration of the profiler.
     #[serde(default)]
@@ -69,4 +75,154 @@ pub struct GlobalConfig {
     /// The sources configurations.
     #[serde(default)]
     pub sources: HashMap<String, toml::Value>,
+}
+
+/// Loads the configuration from `config_file`, with the `-D` overrides applied
+/// on top of it.
+///
+/// Overrides are applied to the raw TOML rather than to the deserialized
+/// configuration, so a `-D` reaches any key a configuration file can set,
+/// including per-source ones, and goes through the same checks.
+pub fn load_global_config(
+    config_file: Option<&Path>,
+    overrides: &[ConfigOverride],
+) -> Result<GlobalConfig> {
+    let mut value = match config_file {
+        Some(path) => {
+            let content = fs::read_to_string(path)
+                .with_context(|| format!("configuration file error on `{}`", path.display()))?;
+            toml::from_str(&content).context("error parsing configuration file")?
+        }
+        None => toml::Value::Table(toml::map::Map::new()),
+    };
+
+    for config_override in overrides {
+        config_override.apply(&mut value)?;
+    }
+
+    value
+        .try_into()
+        .context("error applying the configuration. Check the `-D` overrides")
+}
+
+#[cfg(test)]
+mod tests {
+    use std::io::Write;
+
+    use super::*;
+
+    fn overrides(raw: &[&str]) -> Vec<ConfigOverride> {
+        raw.iter().map(|s| s.parse().unwrap()).collect()
+    }
+
+    fn config_file(content: &str) -> tempfile::NamedTempFile {
+        let mut file = tempfile::NamedTempFile::new().unwrap();
+        write!(file, "{content}").unwrap();
+        file
+    }
+
+    #[test]
+    fn load_without_a_file_nor_overrides_is_the_default_config() {
+        let config = load_global_config(None, &[]).unwrap();
+
+        assert_eq!(config.profiler.token_pattern, DEFAULT_TOKEN_PATTERN);
+        assert!(config.sources.is_empty());
+    }
+
+    #[test]
+    fn load_applies_overrides_without_a_config_file() {
+        let config = load_global_config(
+            None,
+            &overrides(&[
+                "profiler.token_pattern=__CLI__",
+                "sources.cgroup.create_cgroup=false",
+            ]),
+        )
+        .unwrap();
+
+        assert_eq!(config.profiler.token_pattern, "__CLI__");
+        assert_eq!(
[…]
