### 2026-08-12 2468e8b
feat(cli): override configuration with -D flag
