+++ cli/src/config/overrides.rs
@@ -1,0 +1,212 @@
+//! CLI overrides of the TOML configuration (`-D KEY=VALUE`).
+
+use std::str::FromStr;
+
+use anyhow::{Result, anyhow, bail};
+use toml::{Value, map::Map};
+
+/// A single `-D KEY=VALUE` override, `KEY` being the dotted path of the key in
+/// the configuration file.
+#[derive(Debug, Clone, PartialEq)]
+pub struct ConfigOverride {
+    path: Vec<String>,
+    value: Value,
+}
+
+impl ConfigOverride {
+    /// Writes the override into `root`, creating the tables it goes through.
+    ///
+    /// Returns an error if the path traverses a key that is not a table.
+    pub fn apply(&self, root: &mut Value) -> Result<()> {
+        let (key, parents) = self
+            .path
+            .split_last()
+            .expect("an override path always has at least one segment");
+
+        let key_path = self.path.join(".");
+
+        let mut table = root.as_table_mut().ok_or_else(|| {
+            anyhow!("cannot override `{key_path}`: the configuration is not a table.")
+        })?;
+
+        for (depth, segment) in parents.iter().enumerate() {
+            let node = table
+                .entry(segment.clone())
+                .or_insert_with(|| Value::Table(Map::new()));
+            let kind = node.type_str();
+
+            table = node.as_table_mut().ok_or_else(|| {
+                anyhow!(
+                    "cannot override `{key_path}`: `{}` is a {kind}, not a table.",
+                    self.path[..=depth].join("."),
+                )
+            })?;
+        }
+
+        table.insert(key.clone(), self.value.clone());
+
+        Ok(())
+    }
+}
+
+impl FromStr for ConfigOverride {
+    type Err = anyhow::Error;
+
+    fn from_str(s: &str) -> Result<Self> {
+        let Some((key, value)) = s.split_once('=') else {
+            bail!("expected KEY=VALUE, got `{s}`.");
+        };
+
+        let path: Vec<String> = key
+            .trim()
+            .split('.')
+            .map(|segment| segment.trim().to_owned())
+            .collect();
+
+        if path.iter().any(String::is_empty) {
+            bail!("`{key}` is not a valid configuration key.");
+        }
+
+        Ok(Self {
+            path,
+            value: parse_value(value),
+        })
+    }
+}
+
+/// Parses a `-D` value as a TOML value, falling back to a bare string.
+///
+/// The fallback is what lets durations, regexes and paths go unquoted while
+/// `false`, `10` or `[0,1]` keep the type the sources expect.
+fn parse_value(raw: &str) -> Value {
+    let raw = raw.trim();
+
+    toml::from_str::<Value>(&format!("value = {raw}"))
+        .ok()
+        .and_then(|mut parsed| parsed.as_table_mut()?.remove("value"))
+        .unwrap_or_else(|| Value::String(raw.to_owned()))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn parse(s: &str) -> ConfigOverride {
+        s.parse::<ConfigOverride>().unwrap()
+    }
+
+    fn applied(overrides: &[&str]) -> Value {
+        let mut root = Value::Table(Map::new());
+        for raw in overrides {
+            parse(raw).apply(&mut root).unwrap();
+        }
+        root
+    }
+
+    #[test]
+    fn parses_a_dotted_key_path() {
+        let parsed = parse("sources.rapl.rapl_path=/sys/class/powercap");
+
+        assert_eq!(parsed.path, ["sources", "rapl", "rapl_path"]);
+        assert_eq!(
+            parsed.value,
[…]
