+++ examples/example_config.toml
@@ -3,7 +3,12 @@
 # Usage:
 #   joule-profiler --config examples/example_config.toml ...
 #
-# `ignore_on_failure = true` on a configured source ignore the error 
+# Any key below can be overridden on the command line with its dotted path,
+# with or without this file:
+#   joule-profiler --config examples/example_config.toml \
+#     -D profiler.output_format=csv -D sources.cgroup.create_cgroup=false ...
+#
+# `ignore_on_failure = true` on a configured source ignore the error
 # if one occur while building the source.
 #
 # Notes: all fields are optional and Joule Profiler does not require

[…]
