### 2026-08-12 c1a3429
docs: update readme for -D CLI flag
