+++ README.md
@@ -52,8 +52,11 @@ Any key of the configuration file can be set from the command line with `-D KEY=
 
 ```bash
 # Switch the RAPL backend and restrict it to the first socket
-joule-profiler -D profiler.rapl_backend=powercap -D 'sources.rapl.sockets_spec=[0]' profile -- <COMMAND>
+joule-profiler -D profiler.rapl_backend=powercap -D "sources.rapl.sockets_spec=[0]" profile -- <COMMAND>
 ```
+
+> [!IMPORTANT]
+> If a config option is not in string format, then you must add brackets.
 
 For the RAPL source with the perf_event backend, you might be asked to run Joule Profiler with the root privileges, you can configure the perf_event_paranoid level to allow using the source without them:
 ```bash
