### 2026-08-12 b74cfd0
chore(cli): remove rapl polling in profiler config because now it can be set with -D
