### 2026-08-12 c02dbe7
chore(cli): remove rapl polling in profiler config because now it can be set with -D
