### 2026-08-12 f10c782
docs: add examples directory link into readme
