### 2026-08-12 5d53b66
chore: add features flags to enable or disable features
