+++ cli/src/lib.rs
@@ -4,6 +4,7 @@ use clap::{ArgAction, Parser, ValueEnum};
 
 use anyhow::{Result, bail};
 pub use commands::ProfilerCommand;
+#[cfg(feature = "rapl")]
 use serde::Deserialize;
 
 use crate::{
@@ -25,7 +26,7 @@ mod output;
 #[command(name = "joule-profiler")]
 #[command(
     version,
-    about = "Measure program metrics from various sources like RAPL"
+    about = "Measure program metrics from various sources like RAPL, perf_event or NVML"
 )]
 pub struct CliArgs {
     /// Verbosity (-v, -vv, -vvv)
@@ -40,8 +41,12 @@ pub struct CliArgs {
     #[arg(short = 'o', long = "output-file")]
     pub output_file: Option<String>,
 
-    /// Sources activation list. All sources must be separated with a comma (e.g., "perf,nvml").
-    #[arg(long, value_delimiter = ',', default_value = "rapl")]
+    /// Sources activation list. All sources must be separated with a comma (e.g., "rapl,nvml").
+    #[cfg_attr(
+        feature = "rapl",
+        arg(long, value_delimiter = ',', default_value = "rapl")
+    )]
+    #[cfg_attr(not(feature = "rapl"), arg(long, value_delimiter = ','))]
     pub sources: Vec<Source>,
 
     #[allow(clippy::doc_markdown)]
@@ -91,38 +96,86 @@ impl CliArgs {
     }
 }
 
+#[cfg(not(any(
+    feature = "rapl",
+    feature = "perf_event",
+    feature = "nvml",
+    feature = "amdsmi",
+    feature = "procfs",
+    feature = "cgroup",
+)))]
+compile_error!("At least one source feature must be enabled");
+
 #[derive(Clone, Debug, PartialEq, Eq, Hash, ValueEnum)]
 pub enum Source {
+    #[cfg(feature = "rapl")]
     Rapl,
-    Nvml,
-    #[value(name = "amdsmi")]
-    AmdSmi,
+
+    #[cfg(feature = "perf_event")]
     #[value(alias = "perf_event")]
     Perf,
+
+    #[cfg(feature = "nvml")]
+    Nvml,
+
+    #[cfg(feature = "amdsmi")]
+    #[value(name = "amdsmi")]
+    AmdSmi,
+
+    #[cfg(feature = "procfs")]
     Procfs,
+
+    #[cfg(feature = "cgroup")]
     Cgroup,
 }
 
 impl std::fmt::Display for Source {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let s = match self {
+            #[cfg(feature = "rapl")]
             Source::Rapl => "rapl",
-            Source::Nvml => "nvml",
-            Source::AmdSmi => "amdsmi",
+
+            #[cfg(feature = "perf_event")]
             Source::Perf => "perf",
+
+            #[cfg(feature = "nvml")]
+            Source::Nvml => "nvml",
+
+            #[cfg(feature = "amdsmi")]
+            Source::AmdSmi => "amdsmi",
+
+            #[cfg(feature = "procfs")]
             Source::Procfs => "procfs",
+
+            #[cfg(feature = "cgroup")]
             Source::Cgroup => "cgroup",
         };
         write!(f, "{s}")
     }
 }
 
+#[cfg(all(
+    feature = "rapl",
+    not(any(feature = "rapl-backend-perf", feature = "rapl-backend-powercap"))
+))]
+compile_error!("Enable at least one backend: `rapl-backend-perf` or `rapl-backend-powercap`.");
+
 /// RAPL backend, selected with `profiler.rapl_backend` in the configuration.
+///
[…]
