+++ sources/rapl/Cargo.toml
@@ -19,10 +19,15 @@ serde.workspace = true
 humantime-serde.workspace = true
 
 tokio-timerfd = "0.2.0"
-perf-event2 = "0.7.4"
+perf-event2 = { version = "0.7.4", optional = true }
 
 [dev-dependencies]
 tempfile.workspace = true
+
+[features]
+default = ["backend-perf", "backend-powercap"]
+backend-perf = ["dep:perf-event2"]
+backend-powercap = []
 
 [lints]
 workspace = true

[…]
