+++ sources/nvml/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "joule-profiler-source-nvml"
-version = "1.0.2"
+version = "2.0.0"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
