### 2026-08-12 52e99a2
build: Fix amdsmi dep
