+++ examples/library_usage/Cargo.toml
@@ -10,7 +10,7 @@ readme.workspace = true
 [dependencies]
 anyhow = "1.0.102"
 joule-profiler-core = { version = "1.0.2", path = "../../core" }
-joule-profiler-source-rapl = { version = "1.0.2", path = "../../sources/rapl" }
+joule-profiler-source-rapl = { version = "2.0.0", path = "../../sources/rapl" }
 tokio.workspace = true
 
 [lints]
