### 2026-08-12 140b270
feat: Fix core version
