+++ core/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "joule-profiler-core"
-version = "2.0.0"
+version = "2.1.0"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
