+++ Cargo.toml
@@ -20,7 +20,7 @@ repository = "https://github.com/joule-profiler/joule-profiler"
 readme = "README.md"
 
 [workspace.dependencies]
-joule-profiler-core = { path = "core", version = "2.0.0" }
+joule-profiler-core = { path = "core", version = "2.1.0" }
 
 log = "0.4.28"
 tokio = { version = "1.52.3", features = ["time", "rt-multi-thread", "macros", "sync", "rt"] }
