### 2026-08-12 7163de5
fix: Remove library usage example
