### 2026-08-12 9bfd528
Change profiling command in usage comment
Updated usage instructions for profiling the program.
