### 2026-08-13 607dff7
fix: use hidden _rapl feature to activate rapl using rapl-perf and rapl-powercap feature flags
