### 2026-08-13 0d9f644
fix: use hidden _rapl feature to activate rapl using rapl-perf and rapl-powercap feature flags
