+++ cli/src/config/mod.rs
@@ -3,7 +3,7 @@ use std::{collections::HashMap, fs, path::Path, time::Duration};
 use anyhow::{Context, Result};
 use serde::Deserialize;
 
-#[cfg(feature = "rapl")]
+#[cfg(feature = "_rapl")]
 use crate::RaplBackend;
 use crate::{config::overrides::ConfigOverride, output::formats::OutputFormat};
 
@@ -49,7 +49,7 @@ pub struct ProfilerConfig {
     #[serde(default = "default_timeout", with = "humantime_serde")]
     pub init_timeout: Duration,
 
-    #[cfg(feature = "rapl")]
+    #[cfg(feature = "_rapl")]
     #[serde(default)]
     pub rapl_backend: RaplBackend,
 }
@@ -60,7 +60,7 @@ impl Default for ProfilerConfig {
             token_pattern: default_token_pattern(),
             init_timeout: default_timeout(),
             output_format: OutputFormat::Terminal,
-            #[cfg(feature = "rapl")]
+            #[cfg(feature = "_rapl")]
             rapl_backend: RaplBackend::default(),
             use_root: false,
             stdout_file: None,
