+++ cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "joule-profiler-cli"
-version = "2.1.1"
+version = "3.0.0"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
@@ -14,7 +14,7 @@ name = "joule-profiler"
 path = "src/bin/main.rs"
 
 [dependencies]
-joule-profiler-source-rapl = { path = "../sources/rapl", version = "1.0.2", default-features = false, optional = true }
+joule-profiler-source-rapl = { path = "../sources/rapl", version = "1.0.3", default-features = false, optional = true }
 joule-profiler-source-perf_event = { path = "../sources/perf_event", version = "1.0.2", optional = true }
 joule-profiler-source-nvml = { path = "../sources/nvml", version = "1.0.2", optional = true }
 joule-profiler-source-amdsmi = { path = "../sources/amdsmi", version = "0.1.0", optional = true }
