+++ cli/Cargo.toml
@@ -14,7 +14,7 @@ name = "joule-profiler"
 path = "src/bin/main.rs"
 
 [dependencies]
-joule-profiler-source-rapl = { path = "../sources/rapl", version = "1.0.4", default-features = false, optional = true }
+joule-profiler-source-rapl = { path = "../sources/rapl", version = "1.0.5", default-features = false, optional = true }
 joule-profiler-source-perf_event = { path = "../sources/perf_event", version = "1.0.2", optional = true }
 joule-profiler-source-nvml = { path = "../sources/nvml", version = "1.0.2", optional = true }
 joule-profiler-source-amdsmi = { path = "../sources/amdsmi", version = "0.1.0", optional = true }
