+++ sources/rapl/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "joule-profiler-source-rapl"
-version = "1.0.4"
+version = "1.0.5"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
