+++ sources/rapl/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "joule-profiler-source-rapl"
-version = "1.0.3"
+version = "1.0.4"
 edition.workspace = true
 authors.workspace = true
 license.workspace = true
