### 2026-08-13 809cdbc
ci: Update pipeline order
