### 2026-08-17 #10575 @mbarzegary
@jwoirhaye and @rouvoy, you have used tab characters for indentation while adding the author ORCIDs as can be seen in the error above and here: https://github.com/joule-profiler/joule-profiler/blob/main/joss/paper/paper.md. This is invalid for YAML. Please every tab with spaces.

Also, make sure that the (zenodo) archive title matches the title of the paper.
