### 2026-08-17 #10575 @mbarzegary
@jwoirhaye Fig. 4 currently represents many circular points on the plots in a vector format, which makes it slow for the PDF viewer to render it (https://github.com/joule-profiler/joule-profiler/blob/main/joss/paper/images/full_comparison_parallel2.pdf). I recommend the acceptance of the paper in its current form but will ask the Managing EiC to check if it is okay for JOSS. You may need to replace it with a more lightweight or a raster graphics version.
