### 2026-08-17 #10575 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
