### 2026-08-17 #10575 @mbarzegary
@openjournals/csism-eics, Fig. 4 currently represents many circular points on the plots in a high quality vector format, which makes it slow for most PDF viewers to render it. This is the figure file: https://github.com/joule-profiler/joule-profiler/blob/main/joss/paper/images/full_comparison_parallel2.pdf. Please check if the current high quality version is acceptable or not. Otherwise the authors need to replace it with a lighter or bitmap version.
