### 2026-08-17 #10575 @jwoirhaye
@mbarzegary Thank you for your feedback. We’ve changed the image format to make it more lightweight. Is this okay for you?
