### 2026-08-17 #10575 @danielskatz
@jwoirhaye - everything looks good, except one minor issue in the bib, addressed in https://github.com/joule-profiler/joule-profiler/pull/87

Please merge this, then we can continue to acceptance and publication
