### 2026-08-17 #10575 @jwoirhaye
@danielskatz Thanks for your feedback. We’ve merged the PR. Is this okay for you?
