### 2026-08-18 #10575 @danielskatz
Congratulations to @jwoirhaye (Jérémy Woirhaye) and co-authors on your JOSS publication!!

Thanks to @shrishar, @StewMH, and @ygrange for reviewing, and to @mbarzegary for editing!
JOSS is completely dependent on volunteers - we couldn't do this without you
